### 2023-12-21 #6168 @editorialbot
[PRE REVIEW]: MapReader: v1.1.0
**Submitting author:** @kmcdono2 (<a href="http://orcid.org/0000-0001-7506-1025">Katherine McDonough</a>)
**Repository:** https://github.com/Living-with-machines/MapReader
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
