+++ tests/test_load/test_images.py
@@ -12,7 +12,6 @@ from PIL import Image
 from pytest import approx
 from shapely.geometry import Polygon
 
-from mapreader import loader
 from mapreader.load.images import MapImages
 
 
@@ -22,27 +21,31 @@ def sample_dir():
 
 
 @pytest.fixture
-def init_ts_maps(sample_dir, tmp_path):
+def image_id():
+    return "cropped_74488689.png"
+
+
+@pytest.fixture
+def init_maps(sample_dir, image_id, tmp_path):
     """Initializes MapImages object (with metadata from csv and patches).
 
     Returns
     -------
     list
-        image_ID (of parent png image), ts_map (MapImages object), parent_list (== image_ID) and patch_list (list of patches).
+        maps (MapImages object), parent_list (== image_id) and patch_list (list of patches).
     """
-    image_ID = "cropped_74488689.png"
-    ts_map = loader(f"{sample_dir}/{image_ID}")
-    ts_map.add_metadata(f"{sample_dir}/ts_downloaded_maps.csv")
-    ts_map.patchify_all(patch_size=3, path_save=tmp_path)  # gives 9 patches
-    parent_list = ts_map.list_parents()
-    patch_list = ts_map.list_patches()
+    maps = MapImages(f"{sample_dir}/{image_id}")
+    maps.add_metadata(f"{sample_dir}/ts_downloaded_maps.csv")
+    maps.patchify_all(patch_size=3, path_save=tmp_path)  # gives 9 patches
+    parent_list = maps.list_parents()
+    patch_list = maps.list_patches()
 
-    return image_ID, ts_map, parent_list, patch_list
+    return maps, parent_list, patch_list
 
 
 @pytest.fixture
-def keys():
-    return ["parent_id", "image_path", "shape", "name", "coord", "other"]
+def ts_metadata_keys():
+    return ["name", "url", "coordinates", "crs", "published_date", "grid_bb", "polygon"]
 
 
 @pytest.fixture
@@ -54,6 +57,11 @@ def metadata_df():
             "other": [1, 2, 3],
         }
     )
+
+
+@pytest.fixture
+def metadata_keys():
+    return ["parent_id", "image_path", "shape", "name", "coord", "other"]
 
 
 @pytest.fixture
@@ -94,120 +102,102 @@ def missing_metadata_dir(tmp_path, metadata_df):
     return test_path
 
 
-# ---- tests ----
-
-# ---- png tests (separate geo info) ---
+# --- test init ---
 
 
-def test_loader_png(sample_dir):
-    image_ID = "cropped_74488689.png"
-    ts_map = loader(f"{sample_dir}/{image_ID}")
-    assert len(ts_map) == 1
-    assert isinstance(ts_map, MapImages)
+def test_init_png(sample_dir, image_id):
+    maps = MapImages(f"{sample_dir}/{image_id}")
+    assert len(maps.list_parents()) == 1
+    assert len(maps.list_patches()) == 0
+    assert isinstance(maps, MapImages)
+    str(maps)
+    len(maps)
 
 
-# add_metadata tests w/ png files
-def test_loader_add_metadata(sample_dir):
+def test_init_tiff(sample_dir):
+    image_id = "cropped_non_geo.tif"
+    tiffs = MapImages(f"{sample_dir}/{image_id}")
+    assert len(tiffs) == 1
+    assert isinstance(tiffs, MapImages)
+
+
+def test_init_geotiff(sample_dir):
+    image_id = "cropped_geo.tif"
+    geotiffs = MapImages(f"{sample_dir}/{image_id}")
+    assert len(geotiffs) == 1
+    assert isinstance(geotiffs, MapImages)
+
+
[…]
