+++ tests/test_load/test_images.py
@@ -31,7 +31,7 @@ def init_maps(sample_dir, image_id, tmp_path):
 
     Returns
     -------
-    list
+    tuple
         maps (MapImages object), parent_list (== image_id) and patch_list (list of patches).
     """
     maps = MapImages(f"{sample_dir}/{image_id}")
@@ -41,6 +41,28 @@ def init_maps(sample_dir, image_id, tmp_path):
     patch_list = maps.list_patches()
 
     return maps, parent_list, patch_list
+
+
+@pytest.fixture
+def init_dataframes(sample_dir, image_id, tmp_path):
+    """Initializes MapImages object (with metadata from csv and patches) and creates parent and patch dataframes.
+
+    Returns
+    -------
+    tuple
+        path to parent and patch dataframes
+    """
+    maps = MapImages(f"{sample_dir}/{image_id}")
+    maps.add_metadata(f"{sample_dir}/ts_downloaded_maps.csv")
+    maps.patchify_all(patch_size=3, path_save=tmp_path)  # gives 9 patches
+    maps.add_center_coord(tree_level="parent")
+    maps.add_patch_polygons()
+    maps.calc_pixel_stats()
+    _, _ = maps.convert_images(save=True)
+    assert os.path.isfile("./parent_df.csv")
+    assert os.path.isfile("./patch_df.csv")
+
+    return "./parent_df.csv", "./patch_df.csv"
 
 
 @pytest.fixture
@@ -128,13 +150,32 @@ def test_init_geotiff(sample_dir):
     assert isinstance(geotiffs, MapImages)
 
 
-def test_init_tiff_32bit(sample_dir):
+def test_init_parent_path(sample_dir, image_id, capfd):
+    maps = MapImages(
+        f"{sample_dir}/{image_id}",
+        tree_level="patch",
+        parent_path=f"{sample_dir}/{image_id}",
+    )
+    assert len(maps.list_parents()) == 1
+    assert len(maps.list_patches()) == 1
+
+    # without passing tree level should get warning
+    maps = MapImages(f"{sample_dir}/{image_id}", parent_path=f"{sample_dir}/{image_id}")
+    out, _ = capfd.readouterr()
+    assert (
+        "[WARNING] Ignoring `parent_path` as `tree_level`  is set to 'parent'." in out
+    )
+    assert len(maps.list_parents()) == 1
+    assert len(maps.list_patches()) == 0
+
+
+def test_init_tiff_32bit_error(sample_dir):
     image_id = "cropped_32bit.tif"
     with pytest.raises(NotImplementedError, match="Image mode"):
         MapImages(f"{sample_dir}/{image_id}")
 
 
-def test_init_non_image(sample_dir):
+def test_init_non_image_error(sample_dir):
     file_name = "ts_downloaded_maps.csv"
     with pytest.raises(PIL.UnidentifiedImageError, match="not an image"):
         MapImages(f"{sample_dir}/{file_name}")
@@ -269,6 +310,46 @@ def test_add_metadata_columns(matching_metadata_dir, metadata_df):
             assert isinstance(my_files.parents[parent_id]["coord"], tuple)
 
 
+def test_add_metadata_parent(sample_dir, image_id, init_dataframes, ts_metadata_keys):
+    parent_df, _ = init_dataframes
+    maps = MapImages(f"{sample_dir}/{image_id}")
+    maps.add_metadata(parent_df, tree_level="parent")
+    assert all(
+        [
+            x in maps.parents[image_id].keys()
+            for x in [*ts_metadata_keys, "center_lat", "center_lon"]
+        ]
+    )
[…]
