### 2024-01-05 b8d3da4
calc shape from height, width and channels explicitly and allow for single band images
