+++ mapreader/annotate/annotator.py
@@ -22,8 +22,6 @@ from ..load.loader import load_patches
 
 warnings.filterwarnings("ignore", category=UserWarning)
 
-MAX_SIZE = 1000
-
 _CENTER_LAYOUT = widgets.Layout(
     display="flex", flex_flow="column", align_items="center"
 )
@@ -64,8 +62,23 @@ class Annotator(pd.DataFrame):
     sortby : str or None, optional
         Name of the column to use to sort the patch DataFrame, by default None.
         Default sort order is ``ascending=True``. Pass ``ascending=False`` keyword argument to sort in descending order.
-    **kwargs
-        Additional keyword arguments
+    ascending : bool, optional
+        Whether to sort the DataFrame in ascending order when using the ``sortby`` argument, by default True.
+    username : str or None, optional
+        Username to use when saving annotations file, by default None.
+        If not provided, a random string is generated.
+    task_name : str or None, optional
+        Name of the annotation task, by default None.
+    min_values : dict, optional
+        A dictionary consisting of column names (keys) and minimum values as floating point values (values), by default None.
+    max_values : dict, optional
+        A dictionary consisting of column names (keys) and maximum values as floating point values (values), by default None.
+    surrounding : int, optional
+        The number of surrounding images to show for context, by default 1.
+    max_size : int, optional
+        The size in pixels for the longest side to which constrain each patch image, by default 1000.
+    resize_to : int or None, optional
+        The size in pixels for the longest side to which resize each patch image, by default None.
 
     Raises
     ------
@@ -79,21 +92,6 @@ class Annotator(pd.DataFrame):
         If labels provided are not in the form of a list
     SyntaxError
         If labels provided are not in the form of a list
-
-    Notes
-    -----
-
-    Additional kwargs:
-
-    - ``username``: Username to use when saving annotations file. Default: Randomly generated string.
-    - ``task_name``: Name of the annotation task. Default: "task".
-    - ``min_values``: A dictionary consisting of column names (keys) and minimum values as floating point values (values). Default: {}.
-    - ``max_values``: A dictionary consisting of column names (keys) and maximum values as floating point values (values). Default: {}.
-    - ``buttons_per_row``: Number of buttons to display per row. Default: None.
-    - ``ascending``: Whether to sort the DataFrame in ascending order. Default: True.
-    - ``surrounding``: The number of surrounding images to show for context. Default: 1.
-    - ``max_size``: The size in pixels for the longest side to which constrain each patch image. Default: 1000.
-    - ``resize_to``: The size in pixels for the longest side to which resize each patch image. Default: None.
     """
 
     def __init__(
@@ -111,7 +109,14 @@ class Annotator(pd.DataFrame):
         auto_save: bool = True,
         delimiter: str = ",",
         sortby: str | None = None,
-        **kwargs,
+        ascending: bool = True,
+        username: str | None = None,
+        task_name: str | None = None,
+        min_values: dict | None = None,
+        max_values: dict | None = None,
+        surrounding: int = 1,
+        max_size: int = 1000,
+        resize_to: int | None = None,
     ):
         if labels is None:
             labels = []
@@ -195,13 +200,12 @@ class Annotator(pd.DataFrame):
         )
 
         # Set up annotations file
-        username = kwargs.get(
-            "username",
-            "".join(
+        if not username:
+            username = "".join(
                 [random.choice(string.ascii_letters + string.digits) for n in range(30)]
-            ),
-        )
-        task_name = kwargs.get("task_name", "task")
+            )
+        if not task_name:
+            task_name = "task"
[…]
