### 2024-01-05 6dac40f
remove error if no url
