### 2024-01-05 932df44
allow for image_id to be column 0
