### 2024-01-05 af7df2d
add saving of one band geotiffs
