### 2024-01-05 e02f857
add tests for grayscale images
