### 2024-01-05 a086c1f
remove fail on no url col
