### 2024-01-08 8997b4d
add missing tests
