### 2024-01-11 #6168 @kthyng
@kmcdono2 Also usually our submissions have more of a descriptive title and the version is added in a separate step at the end of the review. Could you take a look at other JOSS submission names/titles and see if you might want to add a descriptor after "MapReader"?
