### 2024-01-17 02e2436
enable easier saving of predictions to csv
