### 2024-01-18 f6f5e89
add docstrings, allow user to specify conf
