### 2024-01-18 e978b40
replace `square_cuts` with padding at edge patches
