### 2024-01-22 34014b1
return df after eval
