+++ mapreader/classify/datasets.py
@@ -2,13 +2,14 @@
 from __future__ import annotations
 
 import os
+from ast import literal_eval
+from itertools import product
 from typing import Callable
 
 import matplotlib.pyplot as plt
-import numpy as np
 import pandas as pd
 import torch
-from PIL import Image, ImageOps
+from PIL import Image
 from torch.utils.data import DataLoader, Dataset
 from torchvision import transforms
 
@@ -105,6 +106,8 @@ class PatchDataset(Dataset):
             if os.path.isfile(patch_df):
                 print(f'[INFO] Reading "{patch_df}".')
                 patch_df = pd.read_csv(patch_df, sep=delimiter)
+                # ensure tuple/list columns are read as such
+                patch_df = self._eval_df(patch_df)
                 self.patch_df = patch_df
             else:
                 raise ValueError(f'[ERROR] "{patch_df}" cannot be found.')
@@ -113,6 +116,12 @@ class PatchDataset(Dataset):
             raise ValueError(
                 "[ERROR] Please pass ``patch_df`` as a string (path to csv file) or pd.DataFrame."
             )
+
+        # force index to be integer
+        if self.patch_df.index.name == "image_id":
+            if "image_id" in self.patch_df.columns:
+                self.patch_df.drop(columns=["image_id"], inplace=True)
+            self.patch_df.reset_index(drop=False, names="image_id", inplace=True)
 
         self.label_col = label_col
         self.label_index_col = label_index_col
@@ -152,6 +161,15 @@ class PatchDataset(Dataset):
         else:
             self.transform = transform
 
+    @staticmethod
+    def _eval_df(df):
+        for col in df.columns:
+            try:
+                df[col] = df[col].apply(literal_eval)
+            except (ValueError, TypeError, SyntaxError):
+                pass
+        return df
+
     def __len__(self) -> int:
         """
         Return the length of the dataset.
@@ -186,7 +204,7 @@ class PatchDataset(Dataset):
         if torch.is_tensor(idx):
             idx = idx.tolist()
 
-        img_path = self.patch_df.iloc[idx][self.patch_paths_col]
+        img_path = self.patch_df.at[idx, self.patch_paths_col]
 
         if os.path.exists(img_path):
             img = Image.open(img_path).convert(self.image_mode)
@@ -199,12 +217,12 @@ Please check the image exists, your file paths are correct and that ``.patch_pat
         img = self.transform(img)
 
         if self.label_col in self.patch_df.iloc[idx].keys():
-            image_label = self.patch_df.iloc[idx][self.label_col]
+            image_label = self.patch_df.at[idx, self.label_col]
         else:
             image_label = ""
 
         if self.label_index_col in self.patch_df.iloc[idx].keys():
-            image_label_index = self.patch_df.iloc[idx][self.label_index_col]
+            image_label_index = self.patch_df.at[idx, self.label_index_col]
         else:
             image_label_index = -1
 
@@ -236,7 +254,7 @@ Please check the image exists, your file paths are correct and that ``.patch_pat
         if torch.is_tensor(idx):
             idx = idx.tolist()
 
-        img_path = self.patch_df.iloc[idx][self.patch_paths_col]
+        img_path = self.patch_df.at, [idx, self.patch_paths_col]
 
         if os.path.exists(img_path):
             img = Image.open(img_path).convert(self.image_mode)
@@ -378,12 +396,9 @@ class PatchContextDataset(PatchDataset):
         label_col: str | None = None,
         label_index_col: str | None = None,
         image_mode: str | None = "RGB",
-        context_save_path: str | None = "./maps/maps_context",
-        create_context: bool | None = False,
+        context_dir: str | None = "./maps/maps_context",
[…]
