### 2024-01-22 c1b596c
ensure geotiffs are saved correctly
