+++ mapreader/annotate/annotator.py
@@ -288,8 +288,6 @@ class Annotator(pd.DataFrame):
         self._min_values = min_values or {}
         self._max_values = max_values or {}
 
-        self.patch_width, self.patch_height = self.get_patch_size()
-
         # Create annotations_dir
         Path(annotations_dir).mkdir(parents=True, exist_ok=True)
 
@@ -382,26 +380,6 @@ class Annotator(pd.DataFrame):
                 pass
         return df
 
-    def get_patch_size(self):
-        """
-        Calculate and return the width and height of the patches based on the
-        first patch of the DataFrame, assuming the same shape of patches
-        across the frame.
-
-        Returns
-        -------
-        Tuple[int, int]
-            Width and height of the patches.
-        """
-        patch_width = (
-            self.sort_values("min_x").max_x[0] - self.sort_values("min_x").min_x[0]
-        )
-        patch_height = (
-            self.sort_values("min_y").max_y[0] - self.sort_values("min_y").min_y[0]
-        )
-
-        return patch_width, patch_height
-
     def _setup_buttons(self) -> None:
         """
         Set up buttons for each label to be annotated.
@@ -452,7 +430,7 @@ class Annotator(pd.DataFrame):
         self, as_type: str | None = "list"
     ) -> list[int] | (pd.Index | pd.Series):
         """
-        Gets the indices of rows which are legible for annotation.
+        Gets the indices of rows which are eligible for annotation.
 
         Parameters
         ----------
@@ -468,7 +446,7 @@ class Annotator(pd.DataFrame):
             pd.Index object, or a pd.Series of legible rows.
         """
 
-        def check_legibility(row):
+        def check_eligibility(row):
             if row.label is not None:
                 return False
 
@@ -481,18 +459,16 @@ class Annotator(pd.DataFrame):
 
             return True
 
-        test = self.copy()
-        test["eligible"] = test.apply(check_legibility, axis=1)
-        test = test[
-            ["eligible"] + [col for col in test.columns if not col == "eligible"]
-        ]
+        queue_df = self.copy(deep=True)
+        queue_df["eligible"] = queue_df.apply(check_eligibility, axis=1)
+        queue_df = queue_df[queue_df.eligible].sample(frac=1)  # shuffle
 
-        indices = test[test.eligible].index
+        indices = queue_df[queue_df.eligible].index
         if as_type == "list":
             return list(indices)
         if as_type == "index":
             return indices
-        return test[test.eligible]
+        return queue_df[queue_df.eligible]
 
     def get_context(self):
         """
@@ -516,9 +492,16 @@ class Annotator(pd.DataFrame):
                 im = Image.fromarray(im_array.astype(np.uint8))
             return im
 
-        def get_empty_square():
+        def get_empty_square(patch_size: tuple[int, int]):
+            """Generates an empty square image.
+
+            Parameters
+            ----------
+            patch_size : tuple[int, int]
+                Patch size in pixels as tuple of `(width, height)`.
+            """
             im = Image.new(
-                size=(self.patch_width, self.patch_height),
+                size=patch_size,
                 mode="RGB",
                 color="white",
             )
@@ -533,17 +516,26 @@ class Annotator(pd.DataFrame):
 
         ix = self._queue[self.current_index]
 
-        x = self.at[ix, "min_x"]
-        y = self.at[ix, "min_y"]
-        current_parent = self.at[ix, "parent_id"]
+        min_x = self.at[ix, "min_x"]
+        min_y = self.at[ix, "min_y"]
 
+        # cannot assume all patches are same size
+        try:
+            height, width, _ = self.at[ix, "shape"]
+        except KeyError:
+            im_path = self.at[ix, self.patch_paths_col]
[…]
