+++ mapreader/annotate/annotator.py
@@ -283,6 +283,7 @@ class Annotator(pd.DataFrame):
         self.auto_save = auto_save
         self.username = username
         self.task_name = task_name
+        self._annotate_context = False
 
         # set up for the annotator
         self._min_values = min_values or {}
@@ -485,8 +486,12 @@ class Annotator(pd.DataFrame):
             # Resize the image
             im = Image.open(image_path)
 
+            # Never dim when annotating context
+            if self._annotate_context:
+                dim = False
+
             # Dim the image
-            if dim is True or dim == "True":
+            if dim in [True, "True"]:
                 im_array = np.array(im)
                 im_array = 256 - (256 - im_array) * 0.4  # lighten image
                 im = Image.fromarray(im_array.astype(np.uint8))
@@ -597,6 +602,102 @@ class Annotator(pd.DataFrame):
         resize_to: int | None = None,
         max_size: int | None = None,
     ) -> None:
+        """Annotate at the patch-level of the current patch.
+        Renders the annotation interface for the first image.
+
+        Parameters
+        ----------
+        show_context : bool or None, optional
+            Whether or not to display the surrounding context for each image.
+            Default is None.
+        min_values : dict or None, optional
+            Minimum values for each property to filter images for annotation.
+            It should be provided as a dictionary consisting of column names
+            (keys) and minimum values as floating point values (values).
+            Default is None.
+        max_values : dict or None, optional
+            Maximum values for each property to filter images for annotation.
+            It should be provided as a dictionary consisting of column names
+            (keys) and minimum values as floating point values (values).
+            Default is None
+        surrounding : int or None, optional
+            The number of surrounding images to show for context. Default: 1.
+        max_size : int or None, optional
+            The size in pixels for the longest side to which constrain each
+            patch image. Default: 100.
+
+        Notes
+        -----
+        This method is a wrapper for the ``_annotate`` method.
+        """
+        self._annotate_context = False
+
+        self._annotate(
+            show_context=show_context,
+            min_values=min_values,
+            max_values=max_values,
+            surrounding=surrounding,
+            resize_to=resize_to,
+            max_size=max_size,
+        )
+
+    def annotate_context(
+        self,
+        min_values: dict | None = None,
+        max_values: dict | None = None,
+        resize_to: int | None = None,
+        max_size: int | None = None,
+    ) -> None:
+        """Annotate at the context-level of the current patch.
+        Renders the annotation interface for the first image plus surrounding context.
+
+        Parameters
+        ----------
+        min_values : dict or None, optional
+            Minimum values for each property to filter images for annotation.
+            It should be provided as a dictionary consisting of column names
+            (keys) and minimum values as floating point values (values).
+            Default is None.
+        max_values : dict or None, optional
+            Maximum values for each property to filter images for annotation.
+            It should be provided as a dictionary consisting of column names
+            (keys) and minimum values as floating point values (values).
+            Default is None
+        surrounding : int or None, optional
+            The number of surrounding images to show for context. Default: 1.
+        max_size : int or None, optional
+            The size in pixels for the longest side to which constrain each
+            patch image. Default: 100.
+
+        Notes
+        -----
+        This method is a wrapper for the ``_annotate`` method.
+        """
+        self._annotate_context = True
+
+        if "context_label" not in self.columns:
+            self["context_label"] = None
+        if "context_changed" not in self.columns:
+            self["context_changed"] = False
+
[…]
