+++ mapreader/annotate/annotator.py
@@ -183,7 +183,6 @@ class Annotator(pd.DataFrame):
         # Add label column if not present
         if label_col not in patch_df.columns:
             patch_df[label_col] = None
-        patch_df["changed"] = False
 
         # Check for image paths column
         if patch_paths_col not in patch_df.columns:
@@ -214,47 +213,26 @@ class Annotator(pd.DataFrame):
         # Ensure unique values in list
         labels = sorted(set(labels), key=labels.index)
 
-        # Test for existing file
+        # Test for existing patch annotation file
         if os.path.exists(annotations_file):
-            print(f"[INFO] Loading existing annotations for {username}.")
-            existing_annotations = pd.read_csv(
-                annotations_file, index_col=0, sep=delimiter
+            print("[INFO] Loading existing patch annotations.")
+            patch_df = self._load_annotations(
+                patch_df=patch_df,
+                annotations_file=annotations_file,
+                labels=labels,
+                col=label_col,
+                delimiter=delimiter,
             )
 
-            if label_col not in existing_annotations.columns:
-                raise ValueError(
-                    f"[ERROR] Your existing annotations do not have the label column: {label_col}."
-                )
-
-            print(existing_annotations[label_col].dtype)
-
-            if existing_annotations[label_col].dtype == int:
-                # convert label indices (ints) to labels (strings)
-                # this is to convert old annotations format to new annotations format
-                existing_annotations[label_col] = existing_annotations[label_col].apply(
-                    lambda x: labels[x]
-                )
-
-            patch_df = patch_df.join(
-                existing_annotations, how="left", lsuffix="_x", rsuffix="_y"
-            )
-            patch_df[label_col] = patch_df["label_y"].fillna(patch_df[f"{label_col}_x"])
-            patch_df = patch_df.drop(
-                columns=[
-                    f"{label_col}_x",
-                    f"{label_col}_y",
-                ]
-            )
-            patch_df["changed"] = patch_df[label_col].apply(
-                lambda x: True if x else False
-            )
-
-            patch_df[patch_paths_col] = patch_df[f"{patch_paths_col}_x"]
-            patch_df = patch_df.drop(
-                columns=[
-                    f"{patch_paths_col}_x",
-                    f"{patch_paths_col}_y",
-                ]
+        # Test for existing context annotation file
+        if os.path.exists(f"{annotations_file[:-4]}_context.csv"):
+            print("[INFO] Loading existing context annotations.")
+            patch_df = self._load_annotations(
+                patch_df=patch_df,
+                annotations_file=f"{annotations_file[:-4]}_context.csv",
+                labels=labels,
+                col="context_label",
+                delimiter=delimiter,
             )
 
         # initiate as a DataFrame
@@ -323,7 +301,7 @@ class Annotator(pd.DataFrame):
         self._setup_box()
 
         # Setup queue
-        self._queue = self.get_queue()
+        self._queue = []
 
     @staticmethod
     def _load_dataframes(
@@ -380,6 +358,53 @@ class Annotator(pd.DataFrame):
             except (ValueError, TypeError, SyntaxError):
                 pass
         return df
+
+    @staticmethod
+    def _load_annotations(
+        patch_df: pd.DataFrame,
+        annotations_file: str,
+        labels: list,
+        col: str,
+        delimiter: str,
+    ):
+        """Load existing annotations from file.
+
+        Parameters
+        ----------
+        patch_df : pd.DataFrame
+            Current patch dataframe.
+        annotations_file : str
+            Name of the annotations file
+        labels : list
+            List of labels for annotation.
+        col : str
+            Name of the column in which labels are stored in annotations file
+        delimiter : str
+            Delimiter used in CSV files
+
[…]
