### 2024-01-25 a5e36ca
enable classifier to work with patch+ context labels
