+++ mapreader/classify/classifier.py
@@ -37,7 +37,8 @@ class ClassifierContainer:
         dataloaders: dict[str, DataLoader] | None = None,
         device: str | None = "default",
         input_size: int | None = (224, 224),
-        is_inception: bool | None = False,
+        is_inception: bool = False,
+        context: bool = False,
         load_path: str | None = None,
         force_device: bool | None = False,
         **kwargs,
@@ -65,6 +66,9 @@ class ClassifierContainer:
         is_inception : bool, optional
             Whether the model is an Inception-style model.
             Default is ``False``.
+        context : bool, optional
+            Whether the model is uses patch and context inputs.
+            Default is `False`.
         load_path : str, optional
             The path to an ``.obj`` file containing a
         force_device : bool, optional
@@ -88,8 +92,10 @@ class ClassifierContainer:
             The model.
         input_size : None or tuple of int
             The size of the input to the model.
-        is_inception : None or bool
+        is_inception : bool
             A flag indicating if the model is an Inception model.
+        context : bool
+            A flag indicating if the model uses patch and context as inputs.
         optimizer : None or torch.optim.Optimizer
             The optimizer being used for training the model.
         scheduler : None or torch.optim.lr_scheduler._LRScheduler
@@ -150,6 +156,7 @@ class ClassifierContainer:
                 self.model = model.to(self.device)
                 self.input_size = input_size
                 self.is_inception = is_inception
+                self.context = context
             elif isinstance(model, str):
                 self._initialize_model(model, **kwargs)
 
@@ -183,7 +190,6 @@ class ClassifierContainer:
         min_lr: float,
         max_lr: float,
         spacing: str | None = "linspace",
-        parameter_groups: bool = False,
     ) -> list[dict]:
         """
         Calculates layer-wise learning rates for a given set of model
@@ -201,16 +207,19 @@ class ClassifierContainer:
             where `"linspace"` uses evenly spaced learning rates over a
             specified interval and `"geomspace"` uses learning rates spaced
             evenly on a log scale (a geometric progression). By default ``"linspace"``.
-        parameter_groups : bool, optional
-            When using context mode, whether to consider parameters belonging to the patch model and context model as separate groups.
-            If True, layers belonging to each group will be assigned the same learning rate.
-            Defaults to ``False``.
 
         Returns
         -------
         list of dicts
             A list of dictionaries containing the parameters and learning
             rates for each layer.
+
+        Notes
+        -----
+        parameter_groups : bool, optional
+            When using context mode, whether to consider parameters belonging to the patch model and context model as separate groups.
+            If True, layers belonging to each group will be assigned the same learning rate.
+            Defaults to ``False``.
         """
 
         if spacing.lower() not in ["linspace", "geomspace"]:
@@ -218,10 +227,12 @@ class ClassifierContainer:
                 '[ERROR] ``spacing`` must be one of "linspace" or "geomspace"'
             )
 
-        if parameter_groups:
+        if self.context:
             params2optimize = []
 
-            for group in ["patch_model", "context_model"]:
+            for group in set(
+                tuple[0].split(".")[0] for tuple in [*self.model.named_parameters()]
+            ):
                 group_params = [
                     params
                     for (name, params) in self.model.named_parameters()
@@ -309,6 +320,10 @@ class ClassifierContainer:
         if optim_param_dict is None:
             optim_param_dict = {"lr": 0.001}
         if params2optimize == "default":
+            if self.context:
+                raise ValueError(
+                    "[ERROR] When using context model, first call `params2optimize` cannot be set to `default`."
+                )
             params2optimize = filter(lambda p: p.requires_grad, self.model.parameters())
 
         if optim_type.lower() in ["adam"]:
@@ -899,7 +914,9 @@ Use ``initialize_optimizer`` or ``add_optimizer`` to define one."  # noqa
[…]
