### 2024-01-25 e58acd2
return only context image for context datasets
