+++ mapreader/classify/classifier.py
@@ -37,7 +37,7 @@ class ClassifierContainer:
         dataloaders: dict[str, DataLoader] | None = None,
         device: str | None = "default",
         input_size: int | None = (224, 224),
-        is_inception: bool | None = False,
+        is_inception: bool = False,
         load_path: str | None = None,
         force_device: bool | None = False,
         **kwargs,
@@ -88,7 +88,7 @@ class ClassifierContainer:
             The model.
         input_size : None or tuple of int
             The size of the input to the model.
-        is_inception : None or bool
+        is_inception : bool
             A flag indicating if the model is an Inception model.
         optimizer : None or torch.optim.Optimizer
             The optimizer being used for training the model.
@@ -183,7 +183,6 @@ class ClassifierContainer:
         min_lr: float,
         max_lr: float,
         spacing: str | None = "linspace",
-        parameter_groups: bool = False,
     ) -> list[dict]:
         """
         Calculates layer-wise learning rates for a given set of model
@@ -201,10 +200,6 @@ class ClassifierContainer:
             where `"linspace"` uses evenly spaced learning rates over a
             specified interval and `"geomspace"` uses learning rates spaced
             evenly on a log scale (a geometric progression). By default ``"linspace"``.
-        parameter_groups : bool, optional
-            When using context mode, whether to consider parameters belonging to the patch model and context model as separate groups.
-            If True, layers belonging to each group will be assigned the same learning rate.
-            Defaults to ``False``.
 
         Returns
         -------
@@ -218,42 +213,14 @@ class ClassifierContainer:
                 '[ERROR] ``spacing`` must be one of "linspace" or "geomspace"'
             )
 
-        if parameter_groups:
-            params2optimize = []
-
-            for group in ["patch_model", "context_model"]:
-                group_params = [
-                    params
-                    for (name, params) in self.model.named_parameters()
-                    if group in name
-                ]
-
-                if spacing.lower() == "linspace":
-                    lrs = np.linspace(min_lr, max_lr, len(group_params))
-                elif spacing.lower() in ["log", "geomspace"]:
-                    lrs = np.geomspace(min_lr, max_lr, len(group_params))
-
-                params2optimize.extend(
-                    [
-                        {"params": params, "learning rate": lr}
-                        for params, lr in zip(group_params, lrs)
-                    ]
-                )
-
-        else:
-            if spacing.lower() == "linspace":
-                lrs = np.linspace(
-                    min_lr, max_lr, len(list(self.model.named_parameters()))
-                )
-            elif spacing.lower() in ["log", "geomspace"]:
-                lrs = np.geomspace(
-                    min_lr, max_lr, len(list(self.model.named_parameters()))
-                )
-
-            params2optimize = [
-                {"params": params, "learning rate": lr}
-                for (_, params), lr in zip(self.model.named_parameters(), lrs)
-            ]
+        if spacing.lower() == "linspace":
+            lrs = np.linspace(min_lr, max_lr, len(list(self.model.named_parameters())))
+        elif spacing.lower() in ["log", "geomspace"]:
+            lrs = np.geomspace(min_lr, max_lr, len(list(self.model.named_parameters())))
+        params2optimize = [
+            {"params": params, "learning rate": lr}
+            for (_, params), lr in zip(self.model.named_parameters(), lrs)
+        ]
 
         return params2optimize
 
@@ -931,29 +898,22 @@ Use ``add_criterion`` to define one."
                             ):
                                 outputs, aux_outputs = self.model(*inputs)
 
-                                if not all(
-                                    isinstance(out, torch.Tensor)
-                                    for out in [outputs, aux_outputs]
-                                ):
-                                    try:
-                                        outputs = outputs.logits
-                                        aux_outputs = aux_outputs.logits
[…]
