### 2024-01-25 fe05f91
remove context annotations from annotator
