+++ mapreader/annotate/annotator.py
@@ -224,17 +224,6 @@ class Annotator(pd.DataFrame):
                 delimiter=delimiter,
             )
 
-        # Test for existing context annotation file
-        if os.path.exists(f"{annotations_file[:-4]}_context.csv"):
-            print("[INFO] Loading existing context annotations.")
-            patch_df = self._load_annotations(
-                patch_df=patch_df,
-                annotations_file=f"{annotations_file[:-4]}_context.csv",
-                labels=labels,
-                col="context_label",
-                delimiter=delimiter,
-            )
-
         # initiate as a DataFrame
         super().__init__(patch_df)
 
@@ -261,7 +250,6 @@ class Annotator(pd.DataFrame):
         self.auto_save = auto_save
         self.username = username
         self.task_name = task_name
-        self._annotate_context = False
 
         # set up for the annotator
         self._min_values = min_values or {}
@@ -497,37 +485,6 @@ class Annotator(pd.DataFrame):
             return indices
         return queue_df
 
-    def get_context_queue(
-        self, as_type: str | None = "list"
-    ) -> list[int] | (pd.Index | pd.Series):
-        """
-        Gets the indices of rows which are eligible for annotation at the context-level.
-
-        Parameters
-        ----------
-        as_type : str, optional
-            The format in which to return the indices. Options: "list",
-            "index". Default is "list". If any other value is provided, it
-            returns a pandas.Series.
-
-        Returns
-        -------
-        List[int] or pandas.Index or pandas.Series
-            Depending on "as_type", returns either a list of indices, a
-            pd.Index object, or a pd.Series of legible rows.
-        """
-
-        queue_df = self.copy(deep=True)
-        queue_df = queue_df[queue_df["context_label"].isna()]  # only unlabelled
-        queue_df = queue_df[queue_df[self.label_col].notna()].sample(frac=1)  # shuffle
-
-        indices = queue_df.index
-        if as_type == "list":
-            return list(indices)
-        if as_type == "index":
-            return indices
-        return queue_df
-
     def get_context(self):
         """
         Provides the surrounding context for the patch to be annotated.
@@ -542,10 +499,6 @@ class Annotator(pd.DataFrame):
         def get_path(image_path, dim=True):
             # Resize the image
             im = Image.open(image_path)
-
-            # Never dim when annotating context
-            if self._annotate_context:
-                dim = False
 
             # Dim the image
             if dim in [True, "True"]:
@@ -688,8 +641,6 @@ class Annotator(pd.DataFrame):
         This method is a wrapper for the ``_annotate`` method.
         """
 
-        self._annotate_context = False
-
         if min_values is not None:
             self._min_values = min_values
         if max_values is not None:
@@ -701,51 +652,6 @@ class Annotator(pd.DataFrame):
         self._annotate(
             show_context=show_context,
             surrounding=surrounding,
-            resize_to=resize_to,
-            max_size=max_size,
-        )
-
-    def annotate_context(
-        self,
-        resize_to: int | None = None,
-        max_size: int | None = None,
-    ) -> None:
-        """Annotate at the context-level of the current patch.
-        Renders the annotation interface for the first image plus surrounding context.
-
-        Parameters
-        ----------
-        min_values : dict or None, optional
-            Minimum values for each property to filter images for annotation.
-            It should be provided as a dictionary consisting of column names
-            (keys) and minimum values as floating point values (values).
-            Default is None.
-        max_values : dict or None, optional
-            Maximum values for each property to filter images for annotation.
-            It should be provided as a dictionary consisting of column names
-            (keys) and minimum values as floating point values (values).
-            Default is None
-        surrounding : int or None, optional
[…]
