### 2024-01-25 d8a08e2
force image_id index
