+++ mapreader/classify/load_annotations.py
@@ -24,7 +24,6 @@ class AnnotationsLoader:
         """
         self.annotations = pd.DataFrame()
         self.reviewed = pd.DataFrame()
-        self.id_col = None
         self.patch_paths_col = None
         self.label_col = None
         self.datasets = None
@@ -36,14 +35,13 @@ class AnnotationsLoader:
         images_dir: str | None = None,
         remove_broken: bool | None = True,
         ignore_broken: bool | None = False,
-        id_col: str | None = "image_id",
         patch_paths_col: str | None = "image_path",
         label_col: str | None = "label",
         append: bool | None = True,
         scramble_frame: bool | None = False,
         reset_index: bool | None = False,
     ):
-        """Loads annotations from a csv file or dataframe and can be used to set the ``id_col``, ``patch_paths_col`` and ``label_col`` attributes.
+        """Loads annotations from a csv file or dataframe and can be used to set the ``patch_paths_col`` and ``label_col`` attributes.
 
         Parameters
         ----------
@@ -65,8 +63,6 @@ class AnnotationsLoader:
             Whether to ignore broken image paths (only valid if remove_broken=False).
             If True, annotations with broken paths will remain in annotations dataframe and no error will be raised. This may cause issues!
             If False, annotations with broken paths will raise error. By default, False.
-        id_col : Optional[str], optional
-            The name of the column which contains the image IDs, by default "image_id".
         patch_paths_col : Optional[str], optional
             The name of the column containing the image paths, by default "image_path".
         label_col : Optional[str], optional
@@ -85,14 +81,6 @@ class AnnotationsLoader:
         ValueError
             If ``annotations`` is passed as something other than a string or pd.DataFrame.
         """
-
-        if not self.id_col:
-            self.id_col = id_col
-        elif self.id_col != id_col:
-            print(
-                f'[WARNING] ID column was previously "{self.id_col}, but will now be set to {id_col}.'
-            )
-            self.id_col = id_col
 
         if not self.patch_paths_col:
             self.patch_paths_col = patch_paths_col
@@ -121,7 +109,7 @@ class AnnotationsLoader:
 
         if images_dir:
             abs_images_dir = os.path.abspath(images_dir)
-            annotations[self.patch_paths_col] = annotations[self.id_col].apply(
+            annotations[self.patch_paths_col] = annotations.index.apply(
                 lambda x: os.path.join(abs_images_dir, x)
             )
 
@@ -183,8 +171,6 @@ class AnnotationsLoader:
         if os.path.isfile(annotations):
             print(f'[INFO] Reading "{annotations}"')
             annotations = pd.read_csv(annotations, sep=delimiter, index_col=0)
-            if annotations.index.name in ["name", "image_id"]:
-                annotations.reset_index(inplace=True, drop=False)
         else:
             raise ValueError(f'[ERROR] "{annotations}" cannot be found.')
 
@@ -193,7 +179,9 @@ class AnnotationsLoader:
         if reset_index:
             annotations.reset_index(drop=True, inplace=True)
 
-        annotations.drop_duplicates(subset=self.id_col, inplace=True, keep="first")
+        annotations = annotations[
+            ~annotations.index.duplicated(keep="first")
+        ]  # remove duplicates
         return annotations
 
     def _check_patch_paths(
@@ -274,9 +262,8 @@ Number of annotations remaining: {len(self.annotations)}"
         if len(self.annotations) == 0:
             raise ValueError("[ERROR] No annotations loaded.")
 
-        patch_row = self.annotations[self.annotations[self.id_col] == patch_id]
-        patch_path = patch_row[self.patch_paths_col].values[0]
-        patch_label = patch_row[self.label_col].values[0]
[…]
