### 2024-01-25 dc848c3
use total_df to build context images
