+++ mapreader/classify/datasets.py
@@ -390,8 +390,8 @@ class PatchContextDataset(PatchDataset):
     def __init__(
         self,
         patch_df: pd.DataFrame | str,
-        patch_transform: str,
-        context_transform: str,
+        total_df: pd.DataFrame | str,
+        transform: str,
         delimiter: str = ",",
         patch_paths_col: str | None = "image_path",
         label_col: str | None = None,
@@ -409,11 +409,10 @@ class PatchContextDataset(PatchDataset):
         ----------
         patch_df : pandas.DataFrame or str
             DataFrame or path to csv file containing the paths to image patches and their labels.
-        patch_transform : str
-            Torchvision transform to be applied to input images.
-            Either "train" or "val".
-        context_transform : str
-            Torchvision transform to be applied to target images.
+        total_df : pandas.DataFrame or str
+            DataFrame or path to csv file containing the paths to all images and their labels.
+        transform : str
+            Torchvision transform to be applied to context  images.
             Either "train" or "val".
         delimiter : str
             The delimiter to use when reading the csv file. By default ``","``.
@@ -465,6 +464,7 @@ class PatchContextDataset(PatchDataset):
             if os.path.isfile(patch_df):
                 print(f'[INFO] Reading "{patch_df}".')
                 patch_df = pd.read_csv(patch_df, sep=delimiter)
+                patch_df = self._eval_df(patch_df)
                 self.patch_df = patch_df
             else:
                 raise ValueError(f'[ERROR] "{patch_df}" cannot be found.')
@@ -474,11 +474,26 @@ class PatchContextDataset(PatchDataset):
                 "[ERROR] Please pass ``patch_df`` as a string (path to csv file) or pd.DataFrame."
             )
 
-        # force index to be integer
-        if self.patch_df.index.name in ["image_id", "name"]:
-            if "image_id" in self.patch_df.columns:
-                self.patch_df.drop(columns=["image_id"], inplace=True)
-            self.patch_df.reset_index(drop=False, names="image_id", inplace=True)
+        if isinstance(total_df, pd.DataFrame):
+            self.total_df = total_df
+
+        elif isinstance(total_df, str):
+            if os.path.isfile(total_df):
+                total_df = pd.read_csv(total_df, sep=delimiter)
+                total_df = self._eval_df(total_df)
+                self.total_df = total_df
+            else:
+                raise ValueError(f'[ERROR] "{total_df}" cannot be found.')
+
+        else:
+            raise ValueError(
+                "[ERROR] Please pass ``total_df`` as a string (path to csv file) or pd.DataFrame."
+            )
+
+        # force index to be image_id
+        for df in [self.patch_df, self.total_df]:
+            if "image_id" in df.columns and df.index.name != "image_id":
+                df.set_index("image_id", drop=True, inplace=True)
 
         self.label_col = label_col
         self.label_index_col = label_index_col
@@ -504,25 +519,15 @@ class PatchContextDataset(PatchDataset):
                     self.label_col
                 ].apply(self._get_label_index)
 
-        if isinstance(patch_transform, str):
-            if patch_transform in ["train", "val", "test"]:
-                self.patch_transform = self._default_transform(patch_transform)
+        if isinstance(transform, str):
+            if transform in ["train", "val", "test"]:
+                self.transform = self._default_transform(transform)
             else:
                 raise ValueError(
                     '[ERROR] ``transform`` can only be "train", "val" or "test" or, a transform.'
                 )
         else:
-            self.patch_transform = patch_transform
-
-        if isinstance(context_transform, str):
[…]
