### 2024-01-25 9b9003c
force image_id index
