### 2024-01-30 #6168 @emdupre
👋  Hi @kmcdono2, and thanks for your submission to JOSS !

I see that you have suggested Rebecca Sutton Koeser as a reviewer.  If you have suggestions for additional potential reviewers, please let me know by listing their names or GitHub handles (without the @, so they don't receive a notification) here. This would be particularly helpful as I don't have directly overlapping expertise with this project, so it's useful to know who you would consider to be a good reviewer for the work !

You can suggest reviewers from any relevant project, though we often recommend starting with [this database](https://reviewers.joss.theoj.org/) of people who have already agreed to review for JOSS.

I'll put together a list of folks to reach out to as reviewers based on these suggestions and my own recommendations. I'll update this thread with that outreach.
