### 2024-02-01 #6168 @kthyng
thanks @emdupre! And I'll note that mikemahoney218 is actually a new editor to JOSS so not a great choice for reviewer as he's got other things going on with JOSS now.
