### 2024-02-07 #6168 @emdupre
👋  Hi @kmcdono2 ! While we're waiting to hear back from potential reviewers, I wanted to quickly check in re: a few details on the software paper.

1. Have you decided on a new subtitle for this submission ? I noticed a subtitle in the documentation, which I've included in the invitation to reviewers. This isn't in [the paper draft](https://github.com/Living-with-machines/MapReader/blob/main/paper/paper.md), though, so I wanted to explicitly check with you.
2. In the listed author affiliations, please explicitly write out all acronyms; e.g., "UK" should be "United Kingdom." Please also include the city for each affiliation such that all affiliations follow "Department, Institute/Univ., City, Country" as closely as possible.
3. Your references are not currently rendering in the compiled PDF. Looking at [the raw markdown file](https://raw.githubusercontent.com/Living-with-machines/MapReader/main/paper/paper.md), I believe this is because you've enclosed them in backticks. If you remove these, I expect the references should compile correctly ! Please see [the example JOSS submission](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) for more examples on how these should be included.

If you have any questions, of course, please don't hesitate to ask !
