### 2024-02-10 #6168 @geekysquirrel
Hm I'm not a computer vision expert but experienced in geospatial software and would be happy to review.
