### 2024-02-12 0e1eef4
fix index map vs apply
