### 2024-02-12 #6168 @emdupre
Thank you, @geekysquirrel ! We don't expect reviewers to be experts in all areas of the submission, but you have very relevant expertise that I know would be appreciated in this review of _MapReader_.

I'll add you as a reviewer on this issue now. Once we have a sufficient number of reviewers, I'll open a dedicated review issue 🚀
