### 2024-02-13 #6168 @emdupre
Hi @kmcdono2, thanks for your reply ! [rlskoeser](https://github.com/rlskoeser) is actually already reviewing a JOSS submission, so we prioritize reviewers who are not actively managing other reviews.
