### 2024-02-15 Living-with-machines/MapReader#353 @rwood-97
Choose subtitle for JOSS paper
https://github.com/openjournals/joss-reviews/issues/6168

"Open software for the visual analysis of maps"

- [x] Rosie to update paper in MapReader repo
- [x] Make changes from issue linked above
