### 2024-02-15 #6168 @Anaphory
Hi @emdupre, as much as I like JOSS and would like to help out I'm unfortunately not able to review in the forseeable future.
