### 2024-02-16 #6168 @emdupre
Thank you for letting us know, @Anaphory ! If you'd like me to remove you from the [JOSS reviewer database](https://reviewers.joss.theoj.org/), please let me know. Otherwise, feel free to mute this issue so you don't receive any additional pings !
