### 2024-02-18 #6168 @kmcdono2
@emdupre - the issues with the subtitle, affiliations, and references have been fixed and updated.
