+++ tests/test_classify/test_annotations_loader.py
@@ -8,7 +8,7 @@ from torch.utils.data import DataLoader, RandomSampler
 from torchvision import transforms
 
 from mapreader import AnnotationsLoader
-from mapreader.classify.datasets import PatchDataset
+from mapreader.classify.datasets import PatchContextDataset, PatchDataset
 
 
 @pytest.fixture
@@ -31,18 +31,18 @@ def load_annots(sample_dir):
 @pytest.mark.dependency(name="load_annots_csv", scope="session")
 def test_load_csv(load_annots, sample_dir):
     annots = load_annots
-    assert len(annots.annotations) == 29
+    assert len(annots.annotations) == 81
     assert isinstance(annots.annotations, pd.DataFrame)
-    assert annots.labels_map == {0: "stuff", 1: "nothing"}
+    assert annots.labels_map == {0: "no", 1: "railspace"}
     annots.load(
         f"{sample_dir}/test_annots_append.csv",
         append=True,
         remove_broken=False,
         ignore_broken=True,
     )  # test append
-    assert len(annots.annotations) == 31
-    assert annots.unique_labels == ["stuff", "nothing", "new"]
-    assert annots.labels_map == {0: "stuff", 1: "nothing", 2: "new"}
+    assert len(annots.annotations) == 83
+    assert annots.unique_labels == ["no", "railspace", "building"]
+    assert annots.labels_map == {0: "no", 1: "railspace", 2: "building"}
 
 
 @pytest.mark.dependency(name="load_annots_df", scope="session")
@@ -50,16 +50,60 @@ def test_load_df(sample_dir):
     annots = AnnotationsLoader()
     df = pd.read_csv(f"{sample_dir}/test_annots.csv", sep=",", index_col=0)
     annots.load(df, remove_broken=False, ignore_broken=True)
-    assert len(annots.annotations) == 29
+    assert len(annots.annotations) == 81
     assert isinstance(annots.annotations, pd.DataFrame)
-    assert annots.labels_map == {0: "stuff", 1: "nothing"}
+    assert annots.labels_map == {0: "no", 1: "railspace"}
+
+
+def test_init_images_dir(sample_dir):
+    annots = AnnotationsLoader()
+    annots.load(
+        f"{sample_dir}/test_annots.csv",
+        remove_broken=False,
+        ignore_broken=True,
+        images_dir=sample_dir,
+    )
+    assert annots.annotations.iloc[0]["image_path"].startswith(str(sample_dir))
+
+
+def test_scramble_frame(sample_dir):
+    annots = AnnotationsLoader()
+    annots.load(
+        f"{sample_dir}/test_annots.csv",
+        reset_index=False,
+        scramble_frame=True,
+        remove_broken=False,
+        ignore_broken=True,
+    )
+    assert len(annots.annotations) == 81
+    assert annots.annotations.index[0] != "patch-0-0-1-1-#cropped_74488689.png#.png"
+
+    # with reset_index
+    annots = AnnotationsLoader()
+    annots.load(
+        f"{sample_dir}/test_annots.csv",
+        reset_index=True,
+        scramble_frame=True,
+        remove_broken=False,
+        ignore_broken=True,
+    )
+    assert len(annots.annotations) == 81
+    assert annots.annotations.index[0] == 0
 
 
 def test_create_datasets_default_transforms(load_annots):
     annots = load_annots
     annots.create_datasets(0.5, 0.3, 0.2)
-    assert annots.dataset_sizes == {"train": 14, "val": 9, "test": 6}
+    assert annots.dataset_sizes == {"train": 40, "val": 24, "test": 17}
     assert isinstance(annots.datasets["train"], PatchDataset)
[…]
