### 2024-02-21 7315b4f
ensure pixel stats are correct for edge patches
