### 2024-02-21 05ad379
add tests for geotiff saving (edge patches)
