### 2024-02-21 a61d4c6
add tests for datasets
