### 2024-02-21 6bca80a
only save important cols in annotator
