### 2024-02-21 113f3a1
better test for scramble frame
