### 2024-02-21 02418ee
fix plants worked example
