### 2024-02-22 133d0b6
add readme for workshop notebooks
