+++ worked_examples/geospatial/workshop_june_2023/README.md
@@ -1,0 +1,4 @@
+## Workshop - June 2023
+
+These notebooks are from our workshop in June 2023 and so may contain outdated code.
+Please refer to our other worked examples for up to date notebooks.
