+++ docs/source/Worked-examples/Worked-examples.rst
@@ -1,26 +1,60 @@
 Worked Examples
 ================
 
-.. todo:: Explain these different use cases
+We have provided a number of worked examples to demonstrate how to use the MapReader package.
+These examples can be found in the `worked_examples <https://github.com/Living-with-machines/MapReader/tree/main/worked_examples>`_ directory of the repository.
 
-.. list-table::
-   :widths: 25 25 25
-   :header-rows: 0
-   :stub-columns: 0
+Maps and geospatial images
+--------------------------
 
-   * - **classification_one_inch_maps**
-         .. image:: https://raw.githubusercontent.com/Living-with-machines/MapReader/main/figs/tutorial_classification_one_inch_maps_001.png
-            :width: 200px
-     - **classification_plant_phenotype**
-         .. image:: https://raw.githubusercontent.com/Living-with-machines/MapReader/main/figs/tutorial_classification_plant_phenotype.png
-            :width: 200px
-     - **classification_mnist**
-         .. image:: https://raw.githubusercontent.com/Living-with-machines/MapReader/main/figs/tutorial_classification_mnist.png
-            :width: 200px
+MapReader was developed for maps and geospatial images.
 
-.. toctree::
-   :maxdepth: 1
+Classification of one-inch OS maps
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-   Classification of one-inch OS maps <one_inch_pipeline>
-   Classification of plant phenotypes <plant_pipeline>
-   Classification of mnist dataset <mnist_pipeline>
+.. image:: https://raw.githubusercontent.com/Living-with-machines/MapReader/main/figs/tutorial_classification_one_inch_maps_001.png
+   :width: 400px
+   :target: https://github.com/Living-with-machines/MapReader/tree/main/worked_examples/geospatial
+
+We have provided two examples of how to use MapReader to identify railspace patches in one-inch OS maps.
+Both examples demonstrate how to use MapReader with maps hosted on a tileserver.
+
+Our examples show a full end-to-end use of the MapReader pipeline, including downloading, loading and patchifying map images, annotating patches to create training data, training a model and using the model to classify patches.
+
+The first examples demonstrates how to use MapReader to classify patches using a standard patch-level classification model in which patches are used as inputs to the model.
+
+It can be found `here <https://github.com/Living-with-machines/MapReader/blob/main/worked_examples/geospatial/classification_one_inch_maps/Pipeline.ipynb>`__.
+
+The second example demonstrates how to use MapReader to classify patches using a context-level classification model in which patches and their surrounding patches (i.e. context) are used as inputs to the model.
+
+It can be found `here <https://github.com/Living-with-machines/MapReader/blob/main/worked_examples/geospatial/context_classification_one_inch_maps/Pipeline.ipynb>`__.
+
+Non-geospatial images
+---------------------
+
+MapReader can also for non-geospatial images.
+We have provided two examples of this.
+
+Classification of plant phenotypes
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. image:: https://raw.githubusercontent.com/Living-with-machines/MapReader/main/figs/tutorial_classification_plant_phenotype.png
+   :width: 400px
+   :target: https://github.com/Living-with-machines/MapReader/blob/main/worked_examples/non-geospatial/classification_plant_phenotype/Pipeline.ipynb
+
+In our plant phenotypes example, we demonstrate how to use MapReader to classify plant phenotypes in images of plants.
+Importantly, this worked example demonstrates how to use MapReader with non-georeferenced images (e.g. non-georeferenced map images).
+
[…]
