+++ docs/source/Worked-examples/Worked-examples.rst
@@ -22,11 +22,9 @@ Both examples demonstrate how to use MapReader with maps hosted on a tileserver.
 Our examples show a full end-to-end use of the MapReader pipeline, including downloading, loading and patchifying map images, annotating patches to create training data, training a model and using the model to classify patches.
 
 The first examples demonstrates how to use MapReader to classify patches using a standard patch-level classification model in which patches are used as inputs to the model.
-
 It can be found `here <https://github.com/Living-with-machines/MapReader/blob/main/worked_examples/geospatial/classification_one_inch_maps/Pipeline.ipynb>`__.
 
 The second example demonstrates how to use MapReader to classify patches using a context-level classification model in which patches and their surrounding patches (i.e. context) are used as inputs to the model.
-
 It can be found `here <https://github.com/Living-with-machines/MapReader/blob/main/worked_examples/geospatial/context_classification_one_inch_maps/Pipeline.ipynb>`__.
 
 Non-geospatial images
@@ -44,7 +42,6 @@ Classification of plant phenotypes
 
 In our plant phenotypes example, we demonstrate how to use MapReader to classify plant phenotypes in images of plants.
 Importantly, this worked example demonstrates how to use MapReader with non-georeferenced images (e.g. non-georeferenced map images).
-
 It can be found `here <ttps://github.com/Living-with-machines/MapReader/blob/main/worked_examples/non-geospatial/classification_plant_phenotype/Pipeline.ipynb>`__.
 
 Classification of MNIST digits
@@ -56,5 +53,4 @@ Classification of MNIST digits
 
 In our MNIST example, we demonstrate how to use MapReader to classify MNIST digits.
 Importantly, this example demonstrates how to use MapReader to classify whole images instead of patches and therefore how MapReader can generalize to much broader use cases.
-
 It can be found `here <https://github.com/Living-with-machines/MapReader/blob/main/worked_examples/non-geospatial/classification_mnist/Pipeline.ipynb>`__.
