### 2024-02-29 f106c95
Update Events.rst
Data/Culture workshop details added
