+++ docs/source/Events.rst
@@ -8,21 +8,21 @@ Thanks to the new AHRC-supported project Data/Culture based at The Alan Turing I
 
 **Spring Data/Culture Workshop: MapReader**
 April 30-May 1, 2024 at The Alan Turing Institute, London, UK
-- 2-day workshop for historians and other historically-inclined researchers about learning to use MapReader. 
-- Bursaries available.
-- Apply and find more details at https://www.turing.ac.uk/events/spring-dataculture-workshop-search-inside-maps-mapreader 
+* 2-day workshop for historians and other historically-inclined researchers about learning to use MapReader. 
+* Bursaries available.
+* Apply and find more details at https://www.turing.ac.uk/events/spring-dataculture-workshop-search-inside-maps-mapreader 
 
 **Summer Data/Culture Workshop: MapReader + Newspapers**
 June 5-7, 2024 in Lancaster, UK
-- Co-sponsored by the N8
-- 3-day workshop for historians and other historically-inclined researchers about learning to use MapReader (2 days) paired with an introduction to tools and data from Living with Machines related to digitized British newspaper collections.
-- Bursaries available.
-- Application will be available soon.
+* Co-sponsored by the N8
+* 3-day workshop for historians and other historically-inclined researchers about learning to use MapReader (2 days) paired with an introduction to tools and data from Living with Machines related to digitized British newspaper collections.
+* Bursaries available.
+* Application will be available soon.
 
 **MapReader Workshop at Spatial Humanities 2024**
 September 25, 2024 in Bamberg, Germany
-- 1/2 day workshop co-located with the Spatial Humanities conference
+* 1/2 day workshop co-located with the Spatial Humanities conference
 
 **Fall Data/Culture Workshop: MapReader + Newspapers**
 Date to be announced shortly! 
-- This workshop will be most likely hybrid: online and at The Alan Turing Institute, London, UK
+* This workshop will be most likely hybrid: online and at The Alan Turing Institute, London, UK
