### 2024-03-05 Living-with-machines/MapReader#364 @PipGrylls
Installation of MapReader - JOSS Review
Links: #6434

When following the code installation I found a few items I think should be addressed.

- [ ] Python version is 3.8 which reaches end of support 2024-10, I think this should ideally be 3.12 if possible other wise the highest current release it is compatible with, if this is 3.8 then please justify why.
- [ ] Recommended environment is Conda, but first and second install paths are from pip. PyPi should not be mixed with conda when possible. Refactor this so there are 'the Conda path', 'the pip path’, and ‘the pip source path’. You could also have a ‘the conda source path’
- [ ] Cartopy also is a pip installed by default so it should be noted to users that they need to pick the conda or pip cartopy install
- [ ] Cartopy requires 3.9 which is out of step with the 3.8 listed as above
- [ ] Numpy 1.21.5 install fails
- [ ] All pip installs fail on M2

Package install:
Test 1-
Conda for the env, the experimental Conda install, and Conda to install Cartopy. All good.
Test 2-
pip for the env, pypi for install. Failed
M1 (I’m on M2) troubleshooting, pip install numpy 1.21.5. Failed
Test 3-
using env from test 2, brew install openblas, OPENBLAS="$(brew --prefix openblas)" pip3 install mapreader FAILED
