### 2024-03-05 Living-with-machines/MapReader#365 @PipGrylls
ImportError on SheetDownloader - JOSS Review
Links: #6434

Cannot import SheetDownloader from mapreader

**To Reproduce**
Steps to reproduce the behavior:
1. Launch a jupyter notebook with the Python (mr_py) kernel as specified in install docs
2. Run `from mapreader import SheetDownloader`
3. See error
4. Run `import mapreader`
5. In next cell use `mapreader.` then tab to see possible items in the namespace
6. No SheetDownloader

**Expected behavior**
Sheetdownloader should be available `from mapreader` or the documentation updated to reflect its actutal location.

**Screenshots**
Full notebook:
![Screenshot 2024-03-05 at 10 01 33](https://github.com/Living-with-machines/MapReader/assets/22835635/a6937934-88c7-4986-8e8a-c7fc17e4b009)
Scrolling down in tab complete.
![Screenshot 2024-03-05 at 10 02 03](https://github.com/Living-with-machines/MapReader/assets/22835635/99ac1074-5fea-4af0-af50-d097c706e706)

 - OS: MacOS 14.2.1
 - Installed via conda
