### 2024-03-05 #6434 @kmcdono2
> I will pause my review at this point as the Install, Functionality, and Documentation prevent me from using the software or running tests.

@emdupre - is this a prompt for us to work on? I'm unclear what the blockers are specifically.
