### 2024-03-05 #6434 @PipGrylls
Hi @kmcdono2 See these two issues:
https://github.com/Living-with-machines/MapReader/issues/365
https://github.com/Living-with-machines/MapReader/issues/364
