### 2024-03-05 Living-with-machines/MapReader#364 @kmcdono2
Thanks - re: python version: we'll talk this week about moving to 3.12, it's a good shift to make now and we've just been focused on some other tasks.

Cartopy has been a thorn in our side for a while, we'll work on getting this resolved in a better way.

More soon on resolving the other issues.
