### 2024-03-06 Living-with-machines/MapReader#365 @rwood-97
Hi, can you check which MapReader version this is? I think the conda install is outdated so won't have the `SheetDownloader` class?
