### 2024-03-06 Living-with-machines/MapReader#364 @rwood-97
Hi, can you let me know python version and mapreader version for each package install so I can troubleshoot?
