### 2024-03-06 Living-with-machines/MapReader#364 @PipGrylls
Conda:
python                    3.8.18          h2469fbe_0_cpython    conda-forge
mapreader                 0.3.4+200.g542eca2          py38_0    anothersmith

Pip:
Python 3.11.6
Mapreader wont intall using `pip3 install mapreader`
