### 2024-03-06 Living-with-machines/MapReader#365 @PipGrylls
Conda:
python 3.8.18 h2469fbe_0_cpython conda-forge
mapreader 0.3.4+200.g542eca2 py38_0 anothersmith

If this is not up to date then it should be removed from the docs. Also the mixing of conda and pip/pypi becomes more of an issue.
