### 2024-03-06 #6434 @PipGrylls
Finally, regarding `Contribution and authorship: Has the submitting author made major contributions to the software? Does the full list of paper authors seem appropriate and complete?`
I note that in the paper last two authors listed are the people with the highest level of contribution to the GH, whilst lines of code is not always equal to contribution I would like to see a justification of this author list ordering and an acknowledgment of developer roles.
