### 2024-03-06 #6434 @kmcdono2
Thanks for this @PipGrylls - we can include a note about this and also possibly re-order names (I don't think we paid this much attention, and of course we should).

We also use these 2 spaces in the repo to document roles: https://github.com/Living-with-machines/MapReader?tab=readme-ov-file#contributors & https://github.com/Living-with-machines/MapReader/blob/main/contributors.md, if it helps in the meantime.
