### 2024-03-06 Living-with-machines/MapReader#364 @rwood-97
What is the error message you get with pip install? Something related to building wheels for numpy?
