### 2024-03-06 Living-with-machines/MapReader#364 @PipGrylls
Using cached mapreader-0.3.2-py3-none-any.whl.metadata (15 kB)
ERROR: Cannot install mapreader==0.3.2, mapreader==0.3.3 and mapreader==0.3.4 because these package versions have conflicting dependencies.

The conflict is caused by:
    mapreader 0.3.4 depends on torch<2.0.0 and >=1.10.0
    mapreader 0.3.3 depends on torch<2.0.0 and >=1.10.0
    mapreader 0.3.2 depends on torch<2.0.0 and >=1.10.0

To fix this you could try to:
1. loosen the range of package versions you've specified
2. remove package versions to allow pip attempt to solve the dependency conflict
