### 2024-03-06 Living-with-machines/MapReader#364 @rwood-97
Ah okay, can you do `pip install mapreader==1.1.0` and see how this does?
