### 2024-03-06 Living-with-machines/MapReader#364 @PipGrylls
Seems to have worked:
Package                   Version
------------------------- ---------------
absl-py                   2.1.0
affine                    2.4.0
aiohttp                   3.9.3
aiosignal                 1.3.1
anyio                     4.3.0
appnope                   0.1.4
argon2-cffi               23.1.0
argon2-cffi-bindings      21.2.0
arrow                     1.3.0
asttokens                 2.4.1
async-lru                 2.0.4
async-timeout             4.0.3
attrs                     23.2.0
azure-core                1.30.1
azure-storage-blob        12.19.1
Babel                     2.14.0
beautifulsoup4            4.12.3
bleach                    6.1.0
certifi                   2024.2.2
cffi                      1.16.0
charset-normalizer        3.3.2
click                     8.1.7
click-plugins             1.1.1
cligj                     0.7.2
comm                      0.2.1
contourpy                 1.2.0
cryptography              42.0.5
cycler                    0.12.1
Cython                    0.29.37
debugpy                   1.8.1
decorator                 5.1.1
defusedxml                0.7.1
et-xmlfile                1.1.0
exceptiongroup            1.2.0
executing                 2.0.1
fastjsonschema            2.19.1
fiona                     1.9.5
fonttools                 4.49.0
fqdn                      1.5.1
frozenlist                1.4.1
geographiclib             1.52
geopandas                 0.14.3
geopy                     2.1.0
grpcio                    1.62.0
h11                       0.14.0
httpcore                  1.0.4
httpx                     0.27.0
idna                      3.6
imageio                   2.34.0
ipyannotate               0.1.0b0
ipykernel                 6.29.3
ipython                   8.22.2
ipywidgets                8.1.2
isodate                   0.6.1
isoduration               20.11.0
jedi                      0.19.1
Jinja2                    3.1.3
joblib                    1.3.2
json5                     0.9.20
jsonpointer               2.4
jsonschema                4.21.1
jsonschema-specifications 2023.12.1
jupyter                   1.0.0
jupyter_client            8.6.0
jupyter-console           6.6.3
jupyter_core              5.7.1
jupyter-events            0.9.0
jupyter-lsp               2.2.4
jupyter_server            2.13.0
jupyter_server_terminals  0.5.2
jupyterlab                4.1.3
jupyterlab_pygments       0.3.0
jupyterlab_server         2.25.3
jupyterlab_widgets        3.0.10
kiwisolver                1.4.5
lazy_loader               0.3
mapreader                 1.1.0
Markdown                  3.5.2
MarkupSafe                2.1.5
matplotlib                3.8.3
matplotlib-inline         0.1.6
mistune                   3.0.2
multidict                 6.0.5
nbclient                  0.9.0
nbconvert                 7.16.2
nbformat                  5.9.2
nest-asyncio              1.6.0
networkx                  3.2.1
notebook                  7.1.1
notebook_shim             0.2.4
numpy                     1.26.4
openpyxl                  3.1.2
overrides                 7.7.0
packaging                 23.2
pandas                    1.5.3
pandocfilters             1.5.1
parhugin                  0.0.3
parso                     0.8.3
pexpect                   4.9.0
pillow                    10.2.0
pip                       23.2.1
platformdirs              4.2.0
prometheus_client         0.20.0
prompt-toolkit            3.0.43
protobuf                  4.25.3
psutil                    5.9.8
ptyprocess                0.7.0
pure-eval                 0.2.2
pycparser                 2.21
Pygments                  2.17.2
pyparsing                 3.1.2
pyproj                    3.6.1
python-dateutil           2.9.0.post0
python-json-logger        2.0.7
pytz                      2024.1
PyYAML                    6.0.1
pyzmq                     25.1.2
qtconsole                 5.5.1
QtPy                      2.4.1
rasterio                  1.3.9
referencing               0.33.0
requests                  2.31.0
rfc3339-validator         0.1.4
rfc3986-validator         0.1.1
rpds-py                   0.18.0
scikit-image              0.22.0
scikit-learn              1.4.1.post1
scipy                     1.12.0
Send2Trash                1.8.2
setuptools                68.1.2
shapely                   2.0.3
simplekml                 1.3.6
six                       1.16.0
sniffio                   1.3.1
snuggs                    1.4.7
soupsieve                 2.5
stack-data                0.6.3
tensorboard               2.16.2
tensorboard-data-server   0.7.2
terminado                 0.18.0
threadpoolctl             3.3.0
tifffile                  2024.2.12
tinycss2                  1.2.1
tomli                     2.0.1
torch                     1.13.1
torchinfo                 1.8.0
torchvision               0.12.0
tornado                   6.4
tqdm                      4.66.2
traitlets                 5.14.1
types-python-dateutil     2.8.19.20240106
typing_extensions         4.10.0
uri-template              1.3.0
urllib3                   2.2.1
versioneer                0.29
wcwidth                   0.2.13
webcolors                 1.13
webencodings              0.5.1
websocket-client          1.7.0
Werkzeug                  3.0.1
widgetsnbextension        4.0.10
yarl                      1.9.4
