### 2024-03-08 Living-with-machines/MapReader#364 @rwood-97
Hi @PipGrylls , can I check which of these still need addressing or are causing you an issue:

>  - Python version is 3.8 which reaches end of support 2024-10, I think this should ideally be 3.12 if possible other wise the highest current release it is compatible with, if this is 3.8 then please justify why.
>  - Recommended environment is Conda, but first and second install paths are from pip. PyPi should not be mixed with conda when possible. Refactor this so there are 'the Conda path', 'the pip path’, and ‘the pip source path’. You could also have a ‘the conda source path’
>  - Cartopy also is a pip installed by default so it should be noted to users that they need to pick the conda or pip cartopy install
>  - Cartopy requires 3.9 which is out of step with the 3.8 listed as above
>  - Numpy 1.21.5 install fails
>  - All pip installs fail on M2

Re. cartopy, the most recent version requires >=3.9 but previous versions (e.g. [0.21.0](https://pypi.org/project/Cartopy/0.21.0/)) do
support 3.8
Re. Numpy and M2 problems, I think these are now resolved since you downgraded to a version of python supported by MapReader - is this correct?
