### 2024-03-11 #6434 @emdupre
👋  Hi everyone, happy Monday !

@geekysquirrel I just wanted to check-in, as I noticed you have not yet created your reviewer checklist. If you are hitting any issues in doing so, please don't hesitate to let me know !

I also wanted to explicitly carry over your note from the pre-review issue :
> Hi all, not sure where the best place to say this is but I've had an initial play and I've run into some issues with the installation that required a bit of fiddling so I think the other reviewers might appreciate a heads up:

> The README suggests to install using python 3.8, which works for the most part but when it gets to installing Cartopy, it fails because the latest version requires minimum 3.9. I know it's not a dependency but without it, some of the worked examples are not as useful, so I reckon it should be installed. I originally tried with 3.11 which didn't work but I could get things running with 3.9.

> Also I think it'd be good to update the docs to say that some of the examples (and indeed the tool and its dependencies) require quite a bit of disk space and downloads, so in case your machine is a bit older and/or you're not on a fast connection I'd set some time aside to run the downloads - one of them took almost an hour for me.

I believe this is now being discussed in https://github.com/Living-with-machines/MapReader/issues/364#issue-2168770591, but please feel free to either chime in there or open an additional issue as appropriate !
