### 2024-03-17 Living-with-machines/MapReader#369 @geekysquirrel
Unit tests fail
**Describe the bug**
Tests fail due to dependencies not being installed during normal installation (`timm`, `transformers`).
If the test environment is different from the runtime environment, I would expect instructions on how to install and run tests but I can't find separate instructions on running tests. That said even after installing all missing dependencies, tests still fail.

**To Reproduce**
`pytest .`

**Expected behavior**
tests should pass

**Screenshots**
![Screenshot from 2024-03-17 18-46-53](https://github.com/Living-with-machines/MapReader/assets/15160279/24b5f8a0-3294-4b60-9ac5-494f162a5cf9)

 - OS: Ubuntu
 - Version 22.04
