### 2024-03-18 Living-with-machines/MapReader#369 @rwood-97
Hi, thanks for flagging this.

Currently `timm` and `transformers` are in our dev dependencies and installed using `pip install mapreader[dev]` (PyPI) or `pip install ".[dev]" (if installing from source).
These are used for loading models when testing the classifier but aren't required if you load a model from torch so are in the dev dependencies vs requirements. In the [docs](https://mapreader.readthedocs.io/en/latest/User-guide/Classify/Train.html#initialize-classifiercontainer) we then say to install them separately.

Happy to take advice on this of whether this seems like a sensible set up or would be better for them to be requirements.
Also happy to add this to the docs.
