### 2024-03-18 Living-with-machines/MapReader#369 @rwood-97
Also re. tests failing, can you confirm mapreader version so I can check this?
