### 2024-03-18 Living-with-machines/MapReader#369 @rwood-97
Hi, just thinking this through and suspect if you have installed mapreader 1.1.0 but are testing with most recent version of mapreader repo then this is likely the problem. I will try build in some backwards compatibility and deprecation warnings to fix this but it should also be resolved by checking out the commit corresponding to your mapreader version.
For 1.1.0 [[8997b4d](https://github.com/Living-with-machines/MapReader/commit/8997b4d7fe7c82d5579ffd13daaa90e7253fc0c2) should be the right commit
