### 2024-03-18 #6434 @kmcdono2
@geekysquirrel - thanks, we are working on the authorship issue to highlight the immense work from Rosie (we have dealt with this differently for different pubs, we just need to put our heads together and come up with our plan for this one).

For install, this is essentially resolved, just awaiting a final confirmation from @PipGrylls. You should be able to proceed.
