### 2024-03-25 1c3d5f3
readd square cuts option
