### 2024-03-27 Living-with-machines/MapReader#364 @emdupre
👋  Hi everyone, I just wanted to check in on this as it appears to be a blocker for the JOSS review.

I've tested `pip install mapreader==1.1.0` in two clean conda environments on an M1 Macbook:

python 3.9:
```
Python 3.9.19 (main, Mar 21 2024, 12:07:41)
[Clang 14.0.6 ] :: Anaconda, Inc. on darwin
>>> import mapreader
>>>
```

and python 3.10:
```
Python 3.10.14 (main, Mar 21 2024, 11:21:31)
[Clang 14.0.6 ] on darwin
>>> import mapreader
>>>
```

As shown above, it appears to be correctly installed and imported in both environments.

To review the checklist in the first issue, I would thus make the following suggestions (though @PipGrylls should please correct any misunderstandings on my part !):

1. Document the supported Python versions directly in the README as [in the `setup.py`](https://github.com/Living-with-machines/MapReader/blob/13e0f8816027ba71b1c4e895ff11101e30854cc8/setup.py#L29). As Python 3.7 is already deprecated and 3.8 will be deprecated later this year, it would be good to remove Python 3.7 support and note a timeline for removing 3.8 support.
2.  I would consider additional `conda` support as an optional, additional improvement in installation options, given that `pip` installation is [the JOSS installation criteria for Python packages](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions). However, the installation instructions could be clarified to reduce the likelihood of users mixing `conda` and `pip` installations for dependencies within the same environment.
3. re: cartopy, I would encourage the authors to note this as an optional dependency in the `setup.py` file with the `extras_requires` field, allowing users to additionally pip install this and other associated, optional dependencies. This would be the same structure you have taken with [the `dev` requirements here](https://github.com/Living-with-machines/MapReader/blob/13e0f8816027ba71b1c4e895ff11101e30854cc8/setup.py#L63-L70), but could be titled something like `plotting` so users would run a command like `pip install 'mapreader[plotting]'`.
