### 2024-03-28 #6434 @emdupre
Hi all,

I've followed up on https://github.com/Living-with-machines/MapReader/issues/364 with my understanding of the current status.

If there are any other blockers, of course, please don't hesitate to know ! And thank you to @PipGrylls @geekysquirrel @jordibc for your time so far.
