+++ setup.py
@@ -67,10 +67,10 @@ setuptools.setup(
             "transformers<5.0.0",
             "black>=23.7.0,<24.0.0",
             "flake8>=6.0.0,<7.0.0",
-        ]
+        ],
         "geo": [
             "cartopy>=0.20.0"
-        ]
+        ],
     },  
     classifiers=[
         "Development Status :: 3 - Alpha",
