+++ docs/source/Install.rst
@@ -74,6 +74,16 @@ For example, if you would like to use venv, open your terminal and use the follo
 Step 2: Install MapReader
 --------------------------
 
+Cartopy
+~~~~~~~~
+
+`Cartopy <https://scitools.org.uk/cartopy/docs/latest/>`__ is used by MapReader's download subpackage to plot maps and geospatial data.
+It is not essential to the functioning of MapReader and so is not listed as a required dependancy.
+
+If you would like to use MapReader's plotting functions, you will need to ensure you have installed the `required dependencies <https://scitools.org.uk/cartopy/docs/latest/installing.html#required-dependencies>`__ for cartopy.
+
+Then, to install cartopy with your mapreader installation, you should install mapreader with the `geo` optional dependencies (see notes below).
+
 Method 1: Install from `PyPI <https://pypi.org/project/mapreader/>`_
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
@@ -85,6 +95,9 @@ This is probably the easiest way to install MapReader.
    .. code-block:: bash
 
       pip install mapreader
+
+
+.. note:: Replace `pip install mapreader` with `pip install "mapreader[geo]"` if you would like to install cartopy with your mapreader installation.
 
 Method 2: Install from source
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -106,6 +119,9 @@ This method will create a ``MapReader`` directory on your machine which will con
 
       cd MapReader
       pip install -v -e .
+
+
+.. note:: Replace `pip install -v -e .` with `pip install -v -e ".[geo]" if you would like to install cartopy with your mapreader installation.
 
 ..
    Method 3: Install via conda (**EXPERIMENTAL**)
@@ -163,13 +179,3 @@ If you are using an M1 mac and are having issues installing MapReader due to an 
 
       brew install openblas
       OPENBLAS="$(brew --prefix openblas)" pip install mapreader
-
-Cartopy
-~~~~~~~~
-
-`Cartopy <https://scitools.org.uk/cartopy/docs/latest/>`__ is used by MapReader's download subpackage to plot maps and geospatial data.
-It is not essential to the functioning of MapReader and so is not listed as a dependancy.
-
-If you would like to use MapReader's plotting functions, you should download cartopy by following the instructions `here <https://scitools.org.uk/cartopy/docs/latest/installing.html>`__.
-
-.. note:: There are some required dependencies which you may need to install first. These are listed `here <https://scitools.org.uk/cartopy/docs/latest/installing.html#required-dependencies>`__.
