### 2024-04-02 Living-with-machines/MapReader#364 @rwood-97
I have addressed these in #337 :

- [x] Document the supported Python versions directly in the README as [in the setup.py](https://github.com/Living-with-machines/MapReader/blob/13e0f8816027ba71b1c4e895ff11101e30854cc8/setup.py#L29). As Python 3.7 is already deprecated and 3.8 will be deprecated later this year, it would be good to remove Python 3.7 support and note a timeline for removing 3.8 support.
- [ ] I would consider additional conda support as an optional, additional improvement in installation options, given that pip installation is [the JOSS installation criteria for Python packages](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions). However, the installation instructions could be clarified to reduce the likelihood of users mixing conda and pip installations for dependencies within the same environment.
- [x] re: cartopy, I would encourage the authors to note this as an optional dependency in the setup.py file with the extras_requires field, allowing users to additionally pip install this and other associated, optional dependencies. This would be the same structure you have taken with [the dev requirements here](https://github.com/Living-with-machines/MapReader/blob/13e0f8816027ba71b1c4e895ff11101e30854cc8/setup.py#L63-L70), but could be titled something like plotting so users would run a command like pip install 'mapreader[plotting]'.

For the conda install, we have a ticket #162 related to this and will try do this as soon as possible.

Can I double check whether you are happy for me to close this issue?
