### 2024-04-02 Living-with-machines/MapReader#364 @emdupre
Thank you, @rwood-97  ! I think you meant to link https://github.com/Living-with-machines/MapReader/pull/377, which looks like it's addressing the two noted points.

I'll let @PipGrylls confirm that they're happy to close this issue !
