### 2024-04-04 Living-with-machines/MapReader#364 @PipGrylls
The above isn’t a blocker to JOSS, however further to this and what would IMO be a blocker:

You need to update the documentation and setup.py to list python 3.9 as the minimum python version. Cartopy requires python 3.9 so the installation instructions on your documentation are incorrect.
