### 2024-04-04 #6434 @PipGrylls
I’m happy to tick off the installation with the caveats that the minimum python version be bumped to 3.9 and the troubleshooting be updated to include the exports that I linked in the [issue](https://github.com/Living-with-machines/MapReader/issues/364)

I have also ticked off the tests, which run with warnings that should be addressed asap. Lots of deprecation warning from numpy and pandas warnings https://pandas.pydata.org/pandas-docs/stable/indexing.html#attribute-access. Not strictly blocking but also not good.

Currently the documentation for worked examples is unclear. The [docs link to ](https://github.com/Living-with-machines/MapReader/tree/main/worked_examples) I think a README here explaining how to launch the examples and a list of examples is whats needed. Running the annotation notebooks and the worked_examples/geospatial/classification_one_inch_maps/Pipeline.ipynb notebook seems like this is well done with just some signposting needed.
