### 2024-04-04 #6434 @PipGrylls
As far a functionality goes the sphinx docs seem comprehensive and the pipeline shoes everything in one place. I haven’t tested everything exhaustively but it satisfies a cursory pass. So I’ve checked that off.
