### 2024-04-05 Living-with-machines/MapReader#364 @rwood-97
Hi, sorry, what error you are getting when installing cartopy using python 3.8? Although I can see that the most recent version of cartopy does not support 3.8, earlier versions do so it should install fine. I've just done a test on my laptop (M1 mac) and have also asked someone in the office with an M2 mac to test and had no problems for either (both of us used python 3.8). I also cannot reproduce the error above.

Commands I would suggest to follow are:

step 1. set up env
```
conda create -n mapreader python=3.8
conda activate mapreader
```

step 2. (a) install cartopy dependencies if you haven't done so already (noted [here](https://mapreader.readthedocs.io/en/latest/Install.html#cartopy) in our docs)
```
brew install geos
pip install pyshp
pip install shapely --no-binary shapely
```

step 2. (b) install mapreader
```
pip install "mapreader[geo]"
```

You can also run without cartopy if it is causing you issues. It's not a required dependency for mapreader and is just for visualisation purposes.

Re. the cffi error, could you open a new ticket for this with the full output log as its hard to troubleshoot without further info. I'm unsure if this is something we (MapReader team) can solve since cffi isn't one of our direct dependencies but they have some info on installing [here](https://cffi.readthedocs.io/en/stable/installation.html). It might be a case of needing to run `brew doctor` or similar to update some of the packages on your laptop.

I've moved #229 to our upcoming list of things to do and will try address this asap.
I am going to close this ticket for now since both you and @emdupre  have installed successfully but happy to address other issues in a new ticket.
