+++ .github/dependabot.yml
@@ -13,4 +13,6 @@ updates:
       - "rwood-97"
     labels:
       - "dependencies"
+    target-branch: 
+      - "dev_dependencies"
     
