+++ setup.py
@@ -62,7 +62,7 @@ setuptools.setup(
     extras_require={
         "dev": [
             "pytest<9.0.0",
-            "pytest-cov>=4.1.0,<5.0.0",
+            "pytest-cov>=4.1.0,<6.0.0",
             "timm<1.0.0",
             "transformers<5.0.0",
             "black>=23.7.0,<25.0.0",
