### 2024-04-05 fab54f6
update allowed python versions
