+++ setup.py
@@ -26,7 +26,7 @@ setuptools.setup(
     packages=setuptools.find_packages(),
     include_package_data=True,
     platforms="OS Independent",
-    python_requires=">=3.8, <3.11",
+    python_requires=">=3.9",
     install_requires=[
         "matplotlib>=3.5.0,<4.0.0",
         "numpy>=1.21.5,<2.0.0",
@@ -68,10 +68,8 @@ setuptools.setup(
             "black>=23.7.0,<25.0.0",
             "flake8>=6.0.0,<8.0.0",
         ],
-        "geo": [
-            "cartopy>=0.20.0"
-        ],
-    },  
+        "geo": ["cartopy>=0.20.0"],
+    },
     classifiers=[
         "Development Status :: 3 - Alpha",
         "License :: OSI Approved :: MIT License",
