+++ setup.py
@@ -38,14 +38,13 @@ setuptools.setup(
         "nest-asyncio>=1.5.1,<2.0.0",
         "scikit-image>=0.18.3",
         "scikit-learn>=1.0.1,<2.0.0",
-        "torch>=1.10.0,<2.0.0",
+        "torch>=1.10.0",
         "torchvision>=0.11.1,<0.17.3",
         "jupyter>=1.0.0,<2.0.0",
         "ipykernel>=6.5.1,<7.0.0",
         "ipywidgets>=8.0.0,<9.0.0",
         "ipyannotate==0.1.0-beta.0",
         "Cython>=0.29.24,<0.30.0",
-        # "proj>=0.2.0,<0.3.0",
         "PyYAML>=6.0,<7.0",
         "tensorboard>=2.7.0,<3.0.0",
         "parhugin>=0.0.3,<0.0.4",
