### 2024-04-07 Living-with-machines/MapReader#364 @jordibc
Hi @rwood-97 ! Not sure about the error that @PipGrylls finds, but in case it helps, the one I find when installing cartopy on my computer (AMD Ryzen 7, latest Debian GNU/Linux) with python3.8 after:

```sh
conda create -n mapreader python=3.8
conda activate mapreader
pip install "mapreader[geo]"
```

is the following (with removed output as `[...]`):

```
[...]
Collecting cartopy>=0.20.0 (from mapreader[geo])
  Using cached Cartopy-0.21.1.tar.gz (10.9 MB)
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... done
[...]
Building wheels for collected packages: cartopy
  Building wheel for cartopy (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Building wheel for cartopy (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [288 lines of output]
[...]
      creating build/temp.linux-x86_64-cpython-38/lib/cartopy
      gcc -pthread -B /data/dotstuff/local/mambaforge/envs/mapreader2/compiler_compat -Wno-unused-result -Wsign-compare -DNDEBUG -fwrapv -O2 -Wall -fPIC -O2 -isystem /data/dotstuff/local/mambaforge/envs/mapreader2/include -fPIC -O2 -isystem /data/dotstuff/local/mambaforge/envs/mapreader2/include -fPIC -I/data/dotstuff/local/mambaforge/envs/mapreader2/include -I./lib/cartopy -I/tmp/pip-build-env-zaedxhik/overlay/lib/python3.8/site-packages/numpy/core/include -I/data/dotstuff/local/mambaforge/envs/mapreader2/include/python3.8 -c lib/cartopy/trace.cpp -o build/temp.linux-x86_64-cpython-38/lib/cartopy/trace.o
      lib/cartopy/trace.cpp:767:10: fatal error: geos_c.h: No such file or directory
        767 | #include "geos_c.h"
            |          ^~~~~~~~~~
      compilation terminated.
      error: command '/usr/bin/gcc' failed with exit code 1
      [end of output]

  ERROR: Failed building wheel for cartopy
Failed to build cartopy
ERROR: Could not build wheels for cartopy, which is required to install pyproject.toml-based projects
```

When I use instead a conda environment with python3.9 (`conda create -n mapreader python=3.9`) it correctly installs [cartopy-0.22.0](https://pypi.org/project/Cartopy/0.22.0/) and it works.
