### 2024-04-07 #6434 @jordibc
Hi!

I've finished looking into all the parts and checked all the boxes since I find everything basically satisfactory. MapReader feels very round and mature. I'd like to comment on some minor things that it'd be nice to address though.

### Installation

As @PipGrylls and @geekysquirrel mentioned, it wouldn't work with python 3.8 for me either. (I added some comments about it in https://github.com/Living-with-machines/MapReader/issues/364 .) When using python 3.9 or 3.10 it worked. I second the request to see it fixed in the doc.

### Tests

With respect the automated tests, maybe I missed it, but other than in https://mapreader.readthedocs.io/en/latest/Contribution-guide/Code.html I couldn't find how to run them. I think it may be better to have them mentioned in the main `README.md` too (not just for development, since they are useful to check that one has a working installation), and probably mentioning that `pip install "mapreader[geo,dev]"` would get all that is needed for a full installation.

Also, even if they all pass with no errors (tried with both with python 3.9 and 3.10), they do produce quite a few warnings (3775). It would also be nice if external calls could be mocked, as @geekysquirrel mentioned.

### Notebooks

With respect to the notebooks, I agree with @PipGrylls that some signposting would be very valuable.

They all run correctly, though I needed to do a couple of minor changes. In `worked_examples/geospatial/classification_one_inch_maps/Pipeline.ipynb`, I had to change `./models_tutorial/checkpoint_10.pkl` -> `./models_tutorial/checkpoint_8.pkl`. In `worked_examples/non-geospatial/classification_plant_phenotype/Pipeline.ipynb` I had to change `patch-300-300-350-350-#2014-06-06_plant001_rgb.png#.png` -> `patch-200-200-250-250-#2014-06-06_plant001_rgb.png#.png`. And in `classification_plant_phenotype` I had to change the file `phenotype_#kasra#-[hash]` with a different hash value. (There may have been very few other minor changes I had to do at some point too.)

I also noticed that a couple of `csv` files are actually tab-separated (so, `tsv`):

```
./non-geospatial/classification_plant_phenotype/annotations_phenotype_open_access/phenotype_test_#kasra#.csv
./geospatial/workshop_june_2023/annotations/buildings_#workshop#.csv
```

For consistency, they would probably be better changed into actual csv files. Or at least, renamed with the suffix "tsv".

It's also a bit unusual to find "#" in the names of files. If there's no particular good reason for that, I'd suggest using another name scheme.

One curiosity, I see that the ROC AUC goes up to 100. I suppose this is percentage. I'm more used to see it go up to 1, but if it is percentage, I think adding the "%" would help.

### Documentation

In https://mapreader.readthedocs.io/en/latest/Worked-examples/Worked-examples.html it says:

> MapReader can also for non-geospatial images

which is just missing "can also *be used* for" or similar.

Also, in section "Classification of plant phenotypes", where it says "In our plant phenotypes example [...]. It can be found here." the link in "here" is missing an initial "h" (it's linking to "ttps://github.com/Living-with-machines/MapReader/blob/main/worked_examples/non-geospatial/classification_plant_phenotype/Pipeline.ipynb").

In https://mapreader.readthedocs.io/en/latest/User-guide/Download.html the section "Downloader" just says:

> TBC

Well... :)

As I said, they are minor comments. I think it's all basically okay, and actually a very nice piece of software. Congrats!
