### 2024-04-08 Living-with-machines/MapReader#385 @rwood-97
Update tests - fix warnings and add docs for testing
As we are updating our dependencies and supported python versions it would be good to address the large number of deprecation warnings that are produced when running tests.
As part of this work, we should ensure there is clear documentation of how to run tests after installing mapreader. Potentially on the Installation page of the docs or as a page just below this in the TOC tree.
