+++ CHANGELOG.md
@@ -1,0 +1,6 @@
+# Changelog
+
+## v1.2.0 (2024-04-08)
+
+- Remove support for python 3.8
+- Update dependencies and add cartopy as required dependency
