### 2024-04-08 Living-with-machines/MapReader#369 @rwood-97
Hi,

I can't reproduce this. My steps were:
- set up a new conda env with python 3.9
- `pip install "mapreader[geo, dev]"`  to get mapreader v1.1.5
- `python -m pytest .` gives me 188 passed

Things I would suggest (sorry if these are obvious and you have tried them already):
- `git checkout main` and then `git pull` to check you have most recent version of test files
- `git clean -di` to clean git repo just incase
- `find . -name '__pycache__' -exec rm -rf {} +` to remove pycache
- set up new clean conda env and start installation again

Then also, googling the error suggests it could be to do with a corrupted checkpoint file, e.g. [this](https://stackoverflow.com/questions/71617570/pytorchstreamreader-failed-reading-zip-archive-failed-finding-central-directory) and [this](https://github.com/pytorch/pytorch/issues/31620). I guess this is related to the resnet18 torch model.
