### 2024-04-08 3152729
update changelog with python version table
