+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+# Python versions
+
+The following table shows which versions of MapReader are compatible with which versions of python:
+
++----------+------------+
+| Python   | MapReader  |
++==========+============+
+| 3.9+     | v1.2.0     |
+| 3.8-3.11 | <= v.1.1.5 |
++----------+------------+
+
+
 ## v1.2.0 (2024-04-08)
 
 - Remove support for python 3.8
