### 2024-04-09 f42b1c0
update installation instructions
