### 2024-04-09 #6434 @kmcdono2
@jordibc @PipGrylls  and @geekysquirrel - thanks for all these notes!

I wanted to say a few words about the README and design of names for notebooks. There is a long history to the origins of these notebooks, and we can and should do better in linking up what is referred to in the documentation and what is visible in the directory of worked examples. We'll get on this asap!
