### 2024-04-09 Living-with-machines/MapReader#386 @kmcdono2
fix documentation and directory for worked examples
Per JOSS review:
- [x] think about naming conventions for worked examples @kmcdono2
- [x] move annotation and post processing notebooks to geospatial directory @rwood-97
- [x] update readme in geospatial @kmcdono2
   - [x] add list of worked examples to README for this directory and explain origins/task @kmcdono2
- [x] add/update non-geospatial README @kmcdono2
   - [x] add list of worked examples to README for this directory and explain origins/task @kmcdono2
- [ ] add this list to documentation page for worked examples @rwood-97
- [x] add explanation that you need to clone the repo to run the worked examples @rwood-97
