### 2024-04-10 e1fbd4e
add deepsolo runner
