+++ mapreader/text_spot/runner.py
@@ -1,0 +1,512 @@
+from __future__ import annotations
+
+import os
+import pathlib
+import pickle
+
+import adet
+import geopandas as geopd
+import numpy as np
+import pandas as pd
+from adet.config import get_cfg
+from detectron2.engine import DefaultPredictor
+from PIL import Image
+from shapely import LineString, Polygon
+
+
+class DeepSoloRunner:
+    def __init__(
+        self,
+        patch_df: pd.DataFrame = None,
+        parent_df: pd.DataFrame = None,
+        cfg_file: str
+        | pathlib.Path = "./DeepSolo/configs/R_50/IC15/finetune_150k_tt_mlt_13_15_textocr.yaml",
+        weights_file: str
+        | pathlib.Path = "./ic15_res50_finetune_synth-tt-mlt-13-15-textocr.pth",
+        device: str = "cpu",
+    ) -> None:
+        # first assert we are using the deep solo version of adet
+        assert adet.__version__ == "0.2.0-deepsolo"
+
+        # setup the dataframes
+        self.patch_df = patch_df
+        self.parent_df = parent_df
+
+        # set up predictions as dictionaries
+        self.patch_predictions = {}
+        self.parent_predictions = {}
+        self.geo_predictions = {}
+
+        # setup the config
+        cfg = get_cfg()  # get a fresh new config
+        cfg.merge_from_file(cfg_file)
+        cfg.MODEL.WEIGHTS = weights_file
+        cfg.MODEL.DEVICE = device
+
+        self.voc_size = cfg.MODEL.TRANSFORMER.VOC_SIZE
+        self.use_customer_dictionary = cfg.MODEL.TRANSFORMER.CUSTOM_DICT
+        if self.voc_size == 96:
+            self.CTLABELS = [
+                " ",
+                "!",
+                '"',
+                "#",
+                "$",
+                "%",
+                "&",
+                "'",
+                "(",
+                ")",
+                "*",
+                "+",
+                ",",
+                "-",
+                ".",
+                "/",
+                "0",
+                "1",
+                "2",
+                "3",
+                "4",
+                "5",
+                "6",
+                "7",
+                "8",
+                "9",
+                ":",
+                ";",
+                "<",
+                "=",
+                ">",
+                "?",
+                "@",
+                "A",
+                "B",
+                "C",
+                "D",
+                "E",
+                "F",
+                "G",
+                "H",
+                "I",
+                "J",
+                "K",
+                "L",
+                "M",
+                "N",
+                "O",
+                "P",
+                "Q",
+                "R",
+                "S",
+                "T",
+                "U",
+                "V",
+                "W",
+                "X",
+                "Y",
+                "Z",
+                "[",
+                "\\",
+                "]",
+                "^",
+                "_",
+                "`",
+                "a",
+                "b",
+                "c",
+                "d",
+                "e",
+                "f",
+                "g",
+                "h",
+                "i",
+                "j",
+                "k",
+                "l",
+                "m",
+                "n",
+                "o",
+                "p",
+                "q",
+                "r",
+                "s",
+                "t",
+                "u",
+                "v",
+                "w",
+                "x",
+                "y",
+                "z",
+                "{",
+                "|",
+                "}",
+                "~",
+            ]
+        elif self.voc_size == 37:
+            self.CTLABELS = [
+                "a",
+                "b",
+                "c",
+                "d",
+                "e",
+                "f",
+                "g",
+                "h",
+                "i",
+                "j",
+                "k",
+                "l",
[…]
