### 2024-04-10 c46b45e
add run all method
