### 2024-04-10 a8194dc
update for if patch_df not passed
