+++ .github/workflows/publish-to-test-pypi.yml
@@ -2,10 +2,6 @@
 name: Publish all Python 🐍 distributions 📦 to Test PyPI
 
 on:
-  pull_request:
-    types: [review_requested]
-    branches:
-      - main
   push:
     branches:
       - main
