### 2024-04-10 Living-with-machines/MapReader#389 @rwood-97
Fix typos in docs
https://github.com/openjournals/joss-reviews/issues/6434#issuecomment-2041626766

Fix typos mentioned here
