### 2024-04-10 #6434 @rwood-97
Hi, quick update from me:

## Installation
I've just merged [#384](https://github.com/Living-with-machines/MapReader/pull/384) which updates our minimum python version to 3.9 and adds cartopy as a required dependency.
This should simplify installation and fix the above issues with installing cartopy (I now realise its to do with a C++ library that is required for older versions of cartopy but not for the most recent version and so had been dropped from the cartopy docs).

# Tests
I've created [#385](https://github.com/Living-with-machines/MapReader/issues/385) to address the warnings in tests and to add ensure documentation on how to run them is clearly signposted. Will try and get on this asap.
We also have [#183](https://github.com/Living-with-machines/MapReader/issues/183) re. the mocking of external calls to NLS tileserver and will move this up our to do list.

# Docs /worked examples
Re. worked examples, we've just created [#386](https://github.com/Living-with-machines/MapReader/issues/386) to focus on this. I will add some info on how to clone the repo and run the notebooks for users that want to run the worked examples vs just use them for guidance.
We also have an old ticket [#234](https://github.com/Living-with-machines/MapReader/issues/234) about this with an idea for creating a tool for doing this.
I've also just created [#389](https://github.com/Living-with-machines/MapReader/issues/389) for the typos so I don't forget.
