+++ mapreader/__init__.py
@@ -12,6 +12,8 @@ from mapreader.classify.datasets import PatchContextDataset
 from mapreader.classify.classifier import ClassifierContainer
 from mapreader.classify import custom_models
 
+from mapreader.text_spot.runner import DeepSoloRunner
+
 from mapreader.process import process
 
 from mapreader.annotate.annotator import Annotator
