### 2024-04-12 #6434 @emdupre
Thank you very much to @PipGrylls, @geekysquirrel, and @jordibc for your careful reviews and for summarizing your current concerns here in-thread ! I really appreciate your efforts and time in supporting this JOSS review 💐

Thank you, too, for @kmcdono2 and @rwood-97 for being responsive to these points and for linking the associated issues in-thread ! I will monitor these to confirm how revisions are progressing. Please also feel free to tag me directly, either in these issues or on this global review thread.

Thank you all again for your work to date, and please don't hesitate to ping me here or via email if you have any questions or concerns at this point.
