+++ mapreader/text_spot/dptext_detr_runner.py
@@ -1,0 +1,400 @@
+from __future__ import annotations
+
+import os
+import pathlib
+
+try:
+    import adet
+except ImportError:
+    raise ImportError(
+        "Please install DPText-DETR from the following link: https://github.com/rwood-97/DPText-DETR"
+    )
+
+import geopandas as geopd
+import numpy as np
+import pandas as pd
+from adet.config import get_cfg
+
+try:
+    from detectron2.engine import DefaultPredictor
+except ImportError:
+    raise ImportError("Please install Detectron2")
+
+import matplotlib.patches as patches
+import matplotlib.pyplot as plt
+from PIL import Image
+from shapely import Polygon
+
+# first assert we are using the deep solo version of adet
+assert adet.__version__ == "0.2.0-dptext-detr"
+
+
+class DPTextDETRRunner:
+    def __init__(
+        self,
+        patch_df: pd.DataFrame = None,
+        parent_df: pd.DataFrame = None,
+        cfg_file: str
+        | pathlib.Path = "./DPText-DETR/configs/DPText_DETR/ArT/R_50_poly.yaml",
+        weights_file: str | pathlib.Path = "./art_final.pth",
+        device: str = "cpu",
+    ) -> None:
+        # first assert we are using the deep solo version of adet
+        assert adet.__version__ == "0.2.0-dptext-detr"
+
+        # setup the dataframes
+        self.patch_df = patch_df
+        self.parent_df = parent_df
+
+        # set up predictions as dictionaries
+        self.patch_predictions = {}
+        self.parent_predictions = {}
+        self.geo_predictions = {}
+
+        # setup the config
+        cfg = get_cfg()  # get a fresh new config
+        cfg.merge_from_file(cfg_file)
+        cfg.MODEL.WEIGHTS = weights_file
+        cfg.MODEL.DEVICE = device
+
+        # setup the predictor
+        self.predictor = DefaultPredictor(cfg)
+
+    def run_on_images(
+        self,
+        img_paths: str | pathlib.Path | list,
+        return_dataframe: bool = False,
+    ) -> dict | pd.DataFrame:
+        """Run the model on a list of images.
+
+        Parameters
+        ----------
+        img_paths : str, pathlib.Path or list
+            A list of image paths to run the model on.
+        return_dataframe : bool, optional
+            Whether to return the predictions as a pandas DataFrame, by default False
+
+        Returns
+        -------
+        dict or pd.DataFrame
+            A dictionary of predictions for each image or a DataFrame if `as_dataframe` is True.
+        """
+
+        if isinstance(img_paths, (str, pathlib.Path)):
+            img_paths = [img_paths]
+
+        for img_path in img_paths:
+            _ = self.run_on_image(img_path, return_outputs=False)
+
+        if return_dataframe:
+            return self._dict_to_dataframe(self.patch_predictions, geo=False)
+        return self.patch_predictions
+
+    def run_on_image(
+        self,
+        img_path: str | pathlib.Path,
+        return_outputs=False,
+        return_dataframe: bool = False,
+    ) -> dict | pd.DataFrame:
+        """Run the model on a single image.
+
+        Parameters
+        ----------
+        img_path : str or pathlib.Path
+            The path to the image to run the model on.
+        return_outputs : bool, optional
+            Whether to return the outputs direct from the model, by default False
+        return_dataframe : bool, optional
+            Whether to return the predictions as a pandas DataFrame, by default False
+
+        Returns
+        -------
+        dict or pd.DataFrame
+            The predictions for the image or the outputs from the model if `return_outputs` is True.
+        """
+        # load image
+        img = Image.open(img_path).convert("RGB")
+        img_array = np.array(img)
+
+        # run inference
+        outputs = self.predictor(img_array)
[…]
