### 2024-04-16 7ea1405
fix model weights warning
