### 2024-04-22 #6434 @emdupre
👋  Hi @rwood-97 @kmcdono2 ! I just wanted to check in to see how this response is going for you.

I see that https://github.com/Living-with-machines/MapReader/pull/396 is currently open (in draft) to address https://github.com/Living-with-machines/MapReader/issues/385. I'm therefore assuming that this is still a work-in-progress, but please let me know if you have any specific concerns that I should be aware of !
