### 2024-04-23 d7caa8d
rename text_spot to spot_text
