### 2024-04-23 ad0f59c
add docs for spotting text
