+++ docs/source/User-guide/Spot-text.rst
@@ -1,0 +1,191 @@
+Spot text
+=========
+
+MapReader now contains two new methods of spotting text on maps:
+
+- ``DPTextDETRRunner`` - This is used to detect text on maps using `DPTextDETR<https://github.com/ymy-k/DPText-DETR/tree/main>`__ and outputs bounding boxes and scores.
+- ``DeepSoloRunner`` - This is used to detect and recognise text on maps using `DeepSolo<https://github.com/ViTAE-Transformer/DeepSolo/tree/main>`__ and outputs bounding boxes, text and scores.
+
+Install dependencies
+--------------------
+
+To run these, you will need to install the required dependencies.
+
+.. note::
+
+    We have our own forks of these repos to enable them to work on CPU!
+
+For DPTextDETR, you need to install the following:
+
+- Detectron2 - main repo available [here](https://github.com/facebookresearch/detectron2)
+- DPText-DETR - our fork available [here](https://github.com/rwood-97/DPText-DETR)
+
+For DeepSolo, you need to install the following:
+
+- Detectron2 - main repo available [here](https://github.com/facebookresearch/detectron2)
+- DeepSolo - our fork available [here](https://github.com/rwood-97/DeepSolo)
+
+
+.. note::
+
+    Since both the DPText-DETR and DeepSolo repos are built ontop of `AdelaiDet<https://github.com/aim-uofa/AdelaiDet>`__, you won't be able to install both at the same. To get around this, you can set up two different conda environments, one for each.
+
+
+Set-up the runner
+-----------------
+
+Once you have installed the dependencies, you can set up your chosen "runner".
+
+You will need to choose a model configuration and download the corresponding model weights.
+
+- e.g. for the ``DPTextDETRRunner``, if you choose the "ArT/R_50_poly.yaml", you should download the "art_final.pth" model weights file from the DPTextDETR repo.
+- e.g. for the ``DeepSoloRunner``, if you choose the "R_50/IC15/finetune_150k_tt_mlt_13_15_textocr.yaml", you should download the "ic15_res50_finetune_synth-tt-mlt-13-15-textocr.pth" model weights file from the DeepSolo repo.
+
+You will also need to load your patch and parent dataframes.
+Assuming you have saved them, as shown in the `Load </User-guide/Load>` user guide, you can load them like so:
+
+.. code-block:: python
+
+    import pandas as pd
+
+    patch_df = pd.read_csv("patch_df.csv")
+    parent_df = pd.read_csv("parent_df.csv")
+
+You can then instantiate your runner.
+
+For the DPTextDETRRunner, use:
+
+.. code-block:: python
+
+    from map_reader.spot_text import DPTextDETRRunner
+
+    my_runner = DeepSoloRunner(
+        patch_df,
+        parent_df,
+        cfg_file = "DPText-DETR/configs/DPText_DETR/ArT/R_50_poly.yaml",
+        weights_file = "./art_final.pth",
+    )
+
+or, for the DeepSoloRunner, use:
+
+.. code-block:: python
+
+    from map_reader.spot_text import DeepSoloRunner
+
+    my_runner = DeepSoloRunner(
+        patch_df,
+        parent_df,
+        cfg_file = "DeepSolo/configs/R_50/IC15/finetune_150k_tt_mlt_13_15_textocr.yaml",
+        weights_file = "./ic15_res50_finetune_synth-tt-mlt-13-15-textocr.pth"
+    )
+
+You'll need to adjust the paths to the config and weights files to match your own set-up.
+
+Run the runner
+--------------
+
+You can then run the runner on all patches in your patch dataframe:
+
+.. code-block:: python
+
+    patch_preds = my_runner.run_all()
+
+By default, this will return a dictionary containing all the predictions for each patch.
[…]
