### 2024-04-24 Living-with-machines/MapReader#386 @kmcdono2
NEW REVISED `README` from `worked_examples/geospatial`:

# MapReader and geospatial images (maps or earth observation imagery)

## Worked Examples

| Task Name  | Input Type | Link to Input Data | Brief Description of What Notebook Does | Output Type | Created By |
| ------------- | ------------- | ------------- | ------------- | ------------- | ------------- |
| Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  |
| Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  |

Would you like to add a new Worked Example? Open a ticket and pitch the contribution to the team!
