+++ worked_examples/non-geospatial/README.md
@@ -1,7 +1,5 @@
 # MapReader and non-geospatial images
 
-[TOC]
-
 ## Worked Examples
 
 - [classification_plant_phenotype](./classification_plant_phenotype)
@@ -10,9 +8,16 @@
   * **Data access:** locally stored
   * **Annotations** are done on plant patches (i.e., slices of each plant image).
   * **Classifier:** train/fine-tuned PyTorch CV models.
+ 
+  * 
 - [classification_mnist](./classification_mnist)
   * **Goal:** train/fine-tune PyTorch CV classifiers on MNIST.
   * **Dataset:** Example images taken from http://yann.lecun.com/exdb/mnist/.
   * **Data access:** locally stored
   * **Annotations** are done on whole MNIST images, **not** on patches/slices of those images.
   * **Classifier:** train/fine-tuned PyTorch CV models.
+
+| Task Name  | Input Type | Link to Input Data | Brief Description of What Notebook Does | Output Type | Created By | Annotations | Model |
+| ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- |
+| Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  |
+| Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  |
