### 2024-04-29 197e4dd
update docs for install
