### 2024-04-29 6237341
update where to find configs and weights
