+++ tests/test_sheet_downloader.py
@@ -1,16 +1,20 @@
 from __future__ import annotations
 
 import os
+import re
 from ast import literal_eval
 from pathlib import Path
 
 import pandas as pd
 import pytest
+from PIL import Image
 from pytest import approx
 from shapely.geometry import LineString, MultiPolygon, Polygon
 
 from mapreader import SheetDownloader
 from mapreader.download.data_structures import GridBoundingBox
+from mapreader.download.tile_loading import TileDownloader
+from mapreader.download.tile_merging import TileMerger
 
 
 @pytest.fixture
@@ -20,14 +24,14 @@ def sample_dir():
 
 @pytest.fixture
 def sheet_downloader(sample_dir):
-    test_json = f"{sample_dir}/test_json.json"  # contains 4 features
+    test_json = f"{sample_dir}/test_json.json"  # contains 6 features
     download_url = "https://geo.nls.uk/maps/os/1inch_2nd_ed/{z}/{x}/{y}.png"
     return SheetDownloader(test_json, download_url)
 
 
 def test_init(sheet_downloader):
     sd = sheet_downloader
-    assert len(sd) == 4
+    assert len(sd) == 6
     assert sd.crs == "EPSG:4326"
 
 
@@ -121,7 +125,7 @@ def test_get_minmax_latlon(sheet_downloader, capfd):
     out, _ = capfd.readouterr()
     assert (
         out
-        == "[INFO] Min lat: 51.49344796, max lat: 54.2089733 \n[INFO] Min lon: -2.62837527, max lon: -0.16093917\n"
+        == "[INFO] Min lat: 51.49344796, max lat: 54.2089733 \n[INFO] Min lon: -4.7682, max lon: -0.16093917\n"
     )
 
 
@@ -280,36 +284,41 @@ def test_query_by_string_key_errors(sheet_downloader):
 # download
 
 
-def test_download_all(sheet_downloader, tmp_path):
-    sd = sheet_downloader
-    sd.get_grid_bb(10)
-    assert sd.grid_bbs is True
-    maps_path = tmp_path / "test_maps/"
-    metadata_fname = "test_metadata.csv"
-    sd.download_all_map_sheets(maps_path, metadata_fname)
-    assert os.path.exists(f"{maps_path}/map_102352861.png")
-    assert os.path.exists(f"{maps_path}/{metadata_fname}")
-    df = pd.read_csv(f"{maps_path}/{metadata_fname}", sep=",", index_col=0)
-    assert len(df) == 4
-    assert list(df.columns) == [
-        "name",
-        "url",
-        "coordinates",
-        "crs",
-        "published_date",
-        "grid_bb",
-    ]
-    assert list(df["name"]) == [
-        "map_101602026.png",
-        "map_101602038.png",
-        "map_102352861.png",
-        "map_91617032.png",
-    ]
-    # test coords
-    assert literal_eval(df.loc[0, "coordinates"]) == approx(
-        (-1.0546875, 53.33087298301705, -0.703125, 53.54030739150021), rel=1e-6
-    )
+@pytest.fixture(scope="function")
+def mock_response(monkeypatch):
+    def mock_download_tiles(self, *args, **kwargs):
+        os.makedirs(self.temp_folder, exist_ok=True)
+        return
 
+    monkeypatch.setattr(TileDownloader, "download_tiles", mock_download_tiles)
+
+    def mock_merge(self, *args, **kwargs):
+        os.makedirs(self.output_folder, exist_ok=True)
+
+        merged_image = Image.new("RGBA", (10, 10))
+
+        if kwargs["file_name"] is None:
+            file_name = self._get_output_name(kwargs["grid_bb"])
+        else:
[…]
