### 2024-05-03 04248fc
remove re.search
