+++ mapreader/annotate/annotator.py
@@ -27,7 +27,7 @@ _CENTER_LAYOUT = widgets.Layout(
 )
 
 
-class Annotator(pd.DataFrame):
+class Annotator:
     """
     Annotator class for annotating patches with labels.
 
@@ -227,27 +227,28 @@ class Annotator(pd.DataFrame):
                 delimiter=delimiter,
             )
 
-        # initiate as a DataFrame
-        super().__init__(patch_df)
-
         ## pixel_bounds = x0, y0, x1, y1
-        self["min_x"] = self.pixel_bounds.apply(lambda x: x[0])
-        self["min_y"] = self.pixel_bounds.apply(lambda x: x[1])
-        self["max_x"] = self.pixel_bounds.apply(lambda x: x[2])
-        self["max_y"] = self.pixel_bounds.apply(lambda x: x[3])
+        patch_df["min_x"] = patch_df["pixel_bounds"].apply(lambda x: x[0])
+        patch_df["min_y"] = patch_df["pixel_bounds"].apply(lambda x: x[1])
+        patch_df["max_x"] = patch_df["pixel_bounds"].apply(lambda x: x[2])
+        patch_df["max_y"] = patch_df["pixel_bounds"].apply(lambda x: x[3])
 
         # Sort by sortby column if provided
         if isinstance(sortby, str):
-            if sortby in self.columns:
+            if sortby in patch_df.columns:
                 self._sortby = sortby
                 self._ascending = ascending
             else:
-                raise ValueError(f"[ERROR] {sortby} is not a column in the DataFrame.")
+                raise ValueError(
+                    f"[ERROR] {sortby} is not a column in the patch DataFrame."
+                )
         elif sortby is not None:
             raise ValueError("[ERROR] ``sortby`` must be a string or None.")
         else:
             self._sortby = None
             self._ascending = True
+
+        self.patch_df = patch_df
 
         self._labels = labels
         self.label_col = label_col
@@ -298,6 +299,9 @@ class Annotator(pd.DataFrame):
 
         # Setup queue
         self._queue = []
+
+    def __len__(self):
+        return len(self.patch_df)
 
     @staticmethod
     def _load_dataframes(
@@ -499,7 +503,7 @@ class Annotator(pd.DataFrame):
 
             return True
 
-        queue_df = self.copy(deep=True)
+        queue_df = self.patch_df.copy(deep=True)
         queue_df = queue_df[queue_df[self.label_col].isna()]  # only unlabelled
         queue_df["eligible"] = queue_df.apply(check_eligibility, axis=1)
 
@@ -562,20 +566,20 @@ class Annotator(pd.DataFrame):
 
         ix = self._queue[self.current_index]
 
-        min_x = self.at[ix, "min_x"]
-        min_y = self.at[ix, "min_y"]
+        min_x = self.patch_df.at[ix, "min_x"]
+        min_y = self.patch_df.at[ix, "min_y"]
 
         # cannot assume all patches are same size
         try:
-            height, width, _ = self.at[ix, "shape"]
+            height, width, _ = self.patch_df.at[ix, "shape"]
         except KeyError:
-            im_path = self.at[ix, self.patch_paths_col]
+            im_path = self.patch_df.at[ix, self.patch_paths_col]
             im = Image.open(im_path)
             height = im.height
             width = im.width
 
-        current_parent = self.at[ix, "parent_id"]
-        parent_frame = self.query(f"parent_id=='{current_parent}'")
+        current_parent = self.patch_df.at[ix, "parent_id"]
+        parent_frame = self.patch_df.query(f"parent_id=='{current_parent}'")
 
         deltas = list(range(-self.surrounding, self.surrounding + 1))
[…]
