+++ worked_examples/geospatial/README.md
@@ -1,40 +1,29 @@
-# MapReader and geospatial images
+# MapReader and geospatial images 
+## Example of using MapReader with scanned maps or earth observation imagery
 
-- [Tutorials](#tutorials)
-  - [classification_one_inch_maps_001](./classification_one_inch_maps_001)
-- [Why use MapReader?](#why-use-mapreader)
-- [Guidance for specific user groups](#guidance-for-specific-user-groups)
+## Worked Examples 
 
-## Tutorials
+These subdirectories contain jupyter notebooks that use MapReader for different tasks.
 
-- [classification_one_inch_maps_001](./classification_one_inch_maps_001)
-  * **Goal:** train/fine-tune PyTorch CV classifiers on historical maps.
-  * **Dataset:** from National Library of Scotland: [OS one-inch, 2nd edition layer](https://mapseries-tilesets.s3.amazonaws.com/1inch_2nd_ed/index.html).
-  * **Data access:** tileserver
-  * **Annotations** are done on map patches (i.e., slices of each map).
-  * **Classifier:** train/fine-tuned PyTorch CV models.
-
-## Why use MapReader?
-
-MapReader enables quick, flexible research with large map corpora. It is based on the patchwork method, e.g. scanned map sheets are preprocessed to divide up the content of the map into a grid of squares. Using image classificaion at the level of each patch allows users to define classes (labels) of features on maps related to their research questions.
-
-### You might be interested in using MapReader if:
-- you have access to a large corpus of georeferenced maps
-- you want to quickly test different labels to help refine your research question before/without committing to manual vector data creation
-- your maps were created before surveying accuracy reached modern standards, and therefore you do not want to create overly precise geolocated data based on the content of those maps
-
-### MapReader is well-suited for finding spatial phenomena that:
-- have a homogeneous visual signal across many maps
-- may not correspond to typical categories of map features that are traditionally digitized as vector data in a GIS
+| Worked Example Name  | Task Type | Input Type | Link to Input Data Source | Brief Description of What Notebook Does | Output Type | Created By |
+| ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- |
+| classification_one_inch_maps  | Patch Classification| Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb63k1885/ | Classify patches from OS 1-inch maps | MapReader patch output  | Kasra Hosseini, Rosie Wood |
+| context_classification_one_inch_maps | Patch Classification: Context Model | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb63k1885/ | Classify patches from OS 1-inch maps using context models | Rosie Wood |
+|text_spotting_one_inch_maps | Text Spotting | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb63k1885/ | Detect & Recognize text on OS 1-inch maps | Rosie Wood |
+| workshop_april_2024 | Patch Classification | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb10k1888/ | Classify patches from OS 6-inch 2nd edition maps | Rosie Wood |
+| workshop_june_2023 | Patch Classification | Historical Maps | TBC | Classify patches from OS maps | Rosie Wood, Katie McDonough |
+<!--| Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  |-->
 
 
-## Guidance for specific user groups
+## Worked Example Required Files
 
-### "I want to search the visual contents of a large set of maps to help me answer a question about the past."
+These files are located within the worked_examples directory because they are required to run various notebooks provided here.
 
-- MapReader can help you find instances of spatial phenomena in a collection of maps that is too large for you to 'close read/view'.
[…]
