### 2024-05-03 Living-with-machines/MapReader#386 @kmcdono2
Naming conventions for worked examples:

Geospatial
- patch classification
- patch classification: context model
- text spotting

Non-Geospaial
- patch classification
