+++ worked_examples/geospatial/README.md
@@ -8,10 +8,10 @@ These subdirectories contain jupyter notebooks that use MapReader for different 
 | Worked Example Name  | Task Type | Input Type | Link to Input Data Source | Brief Description of What Notebook Does | Output Type | Annotations | Model | Related Paper | Created By |
 | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- |
 | classification_one_inch_maps  | Patch Classification| Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb63k1885/ | Classify patches from OS 1-inch maps | MapReader patch output | [add annotations] | [add model link] | (no paper)| Kasra Hosseini, Rosie Wood |
-| context_classification_one_inch_maps | Patch Classification: Context Model | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb63k1885/ | Classify patches from OS 1-inch maps using context models | [add annotations] | [add model link] | (no paper) |Rosie Wood |
-|text_spotting_one_inch_maps | Text Spotting | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb63k1885/ | Detect & Recognize text on OS 1-inch maps | [add annotations] | [add model link] | (no paper) | Rosie Wood |
-| workshop_april_2024 | Patch Classification | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb10k1888/ | Classify patches from OS 6-inch 2nd edition maps | [add annotations] | [add model link] | (no paper) | Rosie Wood |
-| workshop_june_2023 | Patch Classification | Historical Maps | TBC | Classify patches from OS maps | Rosie Wood, Katie McDonough |
+| context_classification_one_inch_maps | Patch Classification: Context Model | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb63k1885/ | Classify patches from OS 1-inch maps using context models | MapReader patch outputs | [add annotations] | [add model link] | (no paper) |Rosie Wood |
+|text_spotting_one_inch_maps | Text Spotting | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb63k1885/ | Detect & Recognize text on OS 1-inch maps | MapReader patch outputs | [add annotations] | [add model link] | (no paper) | Rosie Wood |
+| workshop_april_2024 | Patch Classification | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb10k1888/ | Classify patches from OS 6-inch 2nd edition maps | MapReader patch outputs | [add annotations] | [add model link] | (no paper) | Rosie Wood |
+| workshop_june_2023 | Patch Classification | Historical Maps | TBC | Classify patches from OS maps | MapReader patch outputs | Rosie Wood, Katie McDonough |
 <!--| Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  |-->
 
 
