### 2024-05-03 96723b0
Update non-geospatial README.md
