+++ worked_examples/non-geospatial/README.md
@@ -17,7 +17,8 @@
   * **Annotations** are done on whole MNIST images, **not** on patches/slices of those images.
   * **Classifier:** train/fine-tuned PyTorch CV models.
 
-| Task Name  | Input Type | Link to Input Data | Brief Description of What Notebook Does | Output Type | Created By | Annotations | Model |
-| ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- |
-| Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  |
-| Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  |
+| Worked Example Name  | Task Type | Input Type | Link to Input Data | Brief Description of What Notebook Does | Output Type | Created By | Annotations | Model | Paper |
+| ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | 
+| classification_plant_phenotype | Patch Classification | Plant Images | TBC | Fine-tune PyTorch CV classifiers on plant patches in images  | MapReader patch classifications in csv format | 
+ Evangeline Corcoran, Kasra Hosseini | TBC  | TBC  | TBC | 
+| classification_mnist  | Image Classification | MNIST  | http://yann.lecun.com/exdb/mnist/ | train/fine-tune PyTorch CV classifiers on MNIST | MapReader whole image classifications in csv format  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | 
