### 2024-05-03 Living-with-machines/MapReader#386 @kmcdono2
Ask @evangeline-corcoran to review updated non-geospatial worked examples README to add link to her paper, model, annotations, etc. when it's ready.
