### 2024-05-03 bdfbb62
Update Worked-examples.rst
