### 2024-05-03 Living-with-machines/MapReader#386 @kmcdono2
@rwood-97 - just needs for you to do the last few bits in the bullet points at the top, and approve my pull requests! Let me know if there are decisions to take about prose in the docs worked examples page.
