### 2024-05-07 2e2e8e9
update geospatial readme table
