+++ worked_examples/geospatial/README.md
@@ -5,14 +5,15 @@
 
 These subdirectories contain jupyter notebooks that use MapReader for different tasks.
 
-| Worked Example Name  | Task Type | Input Type | Link to Input Data Source | Brief Description of What Notebook Does | Output Type | Annotations | Model | Related Paper | Created By |
-| ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- |
-| classification_one_inch_maps  | Patch Classification| Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb63k1885/ | Classify patches from OS 1-inch maps | MapReader patch output | [add annotations] | [add model link] | (no paper)| Kasra Hosseini, Rosie Wood |
-| context_classification_one_inch_maps | Patch Classification: Context Model | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb63k1885/ | Classify patches from OS 1-inch maps using context models | MapReader patch outputs | [add annotations] | [add model link] | (no paper) |Rosie Wood |
-|text_spotting_one_inch_maps | Text Spotting | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb63k1885/ | Detect & Recognize text on OS 1-inch maps | MapReader patch outputs | [add annotations] | [add model link] | (no paper) | Rosie Wood |
-| workshop_april_2024 | Patch Classification | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb10k1888/ | Classify patches from OS 6-inch 2nd edition maps | MapReader patch outputs | [add annotations] | [add model link] | (no paper) | Rosie Wood |
-| workshop_june_2023 | Patch Classification | Historical Maps | TBC | Classify patches from OS maps | MapReader patch outputs | [add annotations] | [add model link] | (no paper) | Rosie Wood, Katie McDonough |
-<!--| Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  | Content Cell  |-->
+| Worked Example Name  | Task Type | Input Type | Link to Input Data Source | Brief Description of What Notebook Does | Output Type | Model | Related Paper | Created By |
+| ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- |
+| classification_one_inch_maps  | Patch Classification| Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb63k1885/ | Classify patches from OS 1-inch maps | MapReader patch output | None (created in notebook) | (no paper)| Kasra Hosseini, Rosie Wood |
+| context_classification_one_inch_maps | Patch Classification: Context Model | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb63k1885/ | Classify patches from OS 1-inch maps using context models | MapReader patch outputs | None (created in notebook) | (no paper) | Rosie Wood |
+|text_spotting_one_inch_maps | Text Spotting | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb63k1885/ | Detect & Recognize text on OS 1-inch maps | MapReader patch outputs | https://github.com/rwood-97/DeepSolo/tree/dev and https://github.com/rwood-97/DPText-DETR | (no paper) | Rosie Wood |
+| workshop_april_2024 | Patch Classification | Historical Maps | https://cloud.maptiler.com/tiles/uk-osgb10k1888/ | Classify patches from OS 6-inch 2nd edition maps | MapReader patch outputs | None (created in notebook), also uses https://huggingface.co/Livingwithmachines/mr_resnest101e_finetuned_OS_6inch_2nd_ed_railspace | (no paper) | Rosie Wood |
[…]
