### 2024-05-07 b9c029d
update instructions for worked examples
