+++ worked_examples/non-geospatial/README.md
@@ -2,10 +2,10 @@
 
 ### Worked Examples
 
-| Worked Example Name | Task Type | Input Type | Link to Input Data | Brief Description of What Notebook Does | Output Type | Created By | Annotations | Model | Paper |
-| ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- |
-| classification_plant_phenotype | Patch Classification | Plant Images | https://www.plant-phenotyping.org/datasets-download | Fine-tune PyTorch CV classifiers on plant patches in images  | MapReader patch classifications in csv format | Evangeline Corcoran, Kasra Hosseini | TBC  | TBC  | TBC | 
-| classification_mnist  | Image Classification | MNIST  | http://yann.lecun.com/exdb/mnist/ | train/fine-tune PyTorch CV classifiers on MNIST | MapReader whole image classifications in csv format  | Kasra Hosseini  | TBC  | TBC  | (no paper)  | 
+| Worked Example Name | Task Type | Input Type | Link to Input Data | Brief Description of What Notebook Does | Output Type | Created By | Model | Paper |
+| ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | ------------- | 
+| classification_plant_phenotype | Patch Classification | Plant Images | https://www.plant-phenotyping.org/datasets-download | Fine-tune PyTorch CV classifiers on plant patches in images  | MapReader patch classifications in csv format | Evangeline Corcoran, Kasra Hosseini  | TBC  | TBC | 
+| classification_mnist  | Image Classification | MNIST  | http://yann.lecun.com/exdb/mnist/ | train/fine-tune PyTorch CV classifiers on MNIST | MapReader whole image classifications in csv format  | Kasra Hosseini  | None  | (no paper)  | 
 
 #### How can I contribute a worked example?
 Open a ["Documentation Update"](https://github.com/Living-with-machines/MapReader/issues/new?assignees=&labels=documentation&projects=&template=documentation_update.md&title=) ticket and pitch your contribution to the team!
