### 2024-05-08 #6434 @rwood-97
Updates from us:

# Installation
- [x] [#384](https://github.com/Living-with-machines/MapReader/pull/384) is done. This updates our minimum python version to 3.9 and adds cartopy as a required dependency.

# Tests
- [x] [#385](https://github.com/Living-with-machines/MapReader/issues/385) is done. This addresses the warnings in tests and adds documentation on how to run them.
- [x]  [#183](https://github.com/Living-with-machines/MapReader/issues/183) is done. This adds mocking of external requests to NLS tileserver in our tests.

# Docs /worked examples
- [x] [#386](https://github.com/Living-with-machines/MapReader/issues/386) is done. This adds some info on how to clone the repo, run the notebooks for users that want to run the worked examples vs just use them for guidance.
- [x] [#389](https://github.com/Living-with-machines/MapReader/issues/389) is done. This fixes the typos mentioned in https://github.com/openjournals/joss-reviews/issues/6434#issuecomment-2041626766

Was there anything else we needed to fix/update? If not, I think this is all done from our end.
