### 2024-05-09 Living-with-machines/MapReader#369 @rwood-97
hey, could you try these again with the most recent mapreader?
