### 2024-05-12 #6434 @jordibc
Hi @rwood-97 ,

This looks good, thanks for the changes! Alas, for me the tests still produce thousands of errors -- I just tried with the updated branch main, and with both python 3.9 and 3.10. If you cannot reproduce it, I'll be happy to share the warnings. Just for reference, this is with: `conda create -n mapreader python=3.0`, `conda activate mapreader`, `pip install "mapreader[dev]"`, and `python -m pytest .`.

A minor comment is that I still think it would be nice to mention how to run the tests in the main README.md (and not only in https://mapreader.readthedocs.io/en/latest/Contribution-guide/Code.html). Probably mentioning that `pip install "mapreader[dev]"` would get all that is needed for a full installation. It could all be mentioned for example just before "New users should refer to the Installation instructions and Input guidance for help with the initial set up of MapReader" (so they are directed there if they need more detailed installation instructions). If you don't think that's a good change to make, please just let me know.

Other than that, it's all good!
