### 2024-05-12 #6434 @jordibc
Sorry, one more comment about the updates. Other things (of little importance) that are not addressed are the ones that I'm copying below. Please just let me know if you don't want to change it:

---

I also noticed that a couple of `csv` files are actually tab-separated (so, `tsv`):

```
./non-geospatial/classification_plant_phenotype/annotations_phenotype_open_access/phenotype_test_#kasra#.csv
./geospatial/workshop_june_2023/annotations/buildings_#workshop#.csv
```

For consistency, they would probably be better changed into actual csv files. Or at least, renamed with the suffix "tsv".

It's also a bit unusual to find "#" in the names of files. If there's no particular good reason for that, I'd suggest using another name scheme.

One curiosity, I see that the ROC AUC goes up to 100. I suppose this is percentage. I'm more used to see it go up to 1, but if it is percentage, I think adding the "%" would help.
