### 2024-05-13 b7c9981
update test instructions
