+++ docs/source/Install.rst
@@ -137,6 +137,29 @@ Step 3: Add virtual python environment to notebooks
 
 .. note:: if you have used a different name for your python virtual environment replace the ``mapreader`` with whatever name you have used.
 
+
+Running tests
+-------------
+
+To run the tests for MapReader, you will need to have installed the **dev dependencies** as described above.
+
+Also, if you have followed the "Install from PyPI" instructions, you will need to clone the MapReader repository to access the tests. i.e.:
+
+.. code-block:: bash
+
+   git clone https://github.com/Living-with-machines/MapReader.git
+
+You can then run the tests using from the root of the MapReader directory using the following commands:
+
+.. code-block:: bash
+
+   cd path/to/MapReader # change this to your path, e.g. cd ~/MapReader
+   conda activate mapreader
+   python -m pytest -v
+
+If all tests pass, this means that MapReader has been installed and is working as expected.
+
+
 Troubleshooting
 ----------------
 
@@ -167,16 +190,3 @@ If you are using an M1 mac and are having issues installing MapReader due to an 
 
       brew install openblas
       OPENBLAS="$(brew --prefix openblas)" pip install mapreader
-
-
-Running tests
--------------
-
-To run the tests for MapReader, you will need to have installed the dev dependencies as described above.
-You can then run the tests by running the following command from the root of the MapReader directory:
-
-.. code-block:: bash
-
-   python -m pytest -v
-
-If all tests pass, this means that MapReader has been installed correctly.
