### 2024-05-13 #6434 @rwood-97
> This looks good, thanks for the changes! Alas, for me the tests still produce thousands of errors -- I just tried with the updated branch main, and with both python 3.9 and 3.10. If you cannot reproduce it, I'll be happy to share the warnings. Just for reference, this is with: `conda create -n mapreader python=3.0`, `conda activate mapreader`, `pip install "mapreader[dev]"`, and `python -m pytest .`.

Hi, I just tried to reproduce this and wonder what directory you are running `python-m pytest .` from?
I accidentally just ran this from my home directory (as opposed to MapReader directory) and it was collecting tests and logging errors for ages and so wonder if this is the issue. When I then did `cd MapReader` and ran again it all passed.

> A minor comment is that I still think it would be nice to mention how to run the tests in the main README.md (and not only in https://mapreader.readthedocs.io/en/latest/Contribution-guide/Code.html). Probably mentioning that `pip install "mapreader[dev]"` would get all that is needed for a full installation. It could all be mentioned for example just before "New users should refer to the Installation instructions and Input guidance for help with the initial set up of MapReader" (so they are directed there if they need more detailed installation instructions). If you don't think that's a good change to make, please just let me know.

Hi, we've also added how to run tests in the [Install](https://mapreader.readthedocs.io/en/latest/Install.html) instructions page of the docs. We'd like to keep everything in the docs vs. README as much as possible so its all in one place. Let me know if you think this needs more fleshing out though.

> I also noticed that a couple of `csv` files are actually tab-separated (so, `tsv`):
>
> ```
> ./non-geospatial/classification_plant_phenotype/annotations_phenotype_open_access/phenotype_test_#kasra#.csv
> ./geospatial/workshop_june_2023/annotations/buildings_#workshop#.csv
> ```
>
> For consistency, they would probably be better changed into actual csv files. Or at least, renamed with the suffix "tsv".

Sorry I missed this one, have just changed it now.

> It's also a bit unusual to find "#" in the names of files. If there's no particular good reason for that, I'd suggest using another name scheme.

The # is to do with being able to split the names back out (i.e. with str.split("#")) later on.

> One curiosity, I see that the ROC AUC goes up to 100. I suppose this is percentage. I'm more used to see it go up to 1, but if it is percentage, I think adding the "%" would help.

Again, sorry I missed this one, have just updated all the notebooks to have `(%)` in their y-labels.
