### 2024-05-13 #6434 @jordibc
Thanks! Everything looks good!

> Hi, I just tried to reproduce this and wonder what directory you are running `python-m pytest .` from? I accidentally just ran this from my home directory (as opposed to MapReader directory) and it was collecting tests and logging errors for ages and so wonder if this is the issue. When I then did `cd MapReader` and ran again it all passed.

I was running from the main MapReader directory, but it seems I was running with an installation from PyPI (`pip install "mapreader[dev]"`, instead of using `-U` to update, or a fresh `pip install -e ".[dev]"`).

Indeed all those thousands of warnings are gone, great! With python 3.10 I got a few warnings (91) that I don't think are important but I'm copying just for reference:

```
tests/test_annotator.py: 18 warnings
tests/test_classify/test_datasets.py: 9 warnings
tests/test_geo_pipeline.py: 1 warning
tests/test_load/test_geo_utils.py: 1 warning
tests/test_load/test_images.py: 55 warnings
  /data/dotstuff/local/mambaforge/envs/mapreader/lib/python3.10/site-packages/rasterio/__init__.py:304: NotGeoreferencedWarning: Dataset has no geotransform, gcps, or rpcs. The identity matrix will be returned.
    dataset = DatasetReader(path, driver=driver, sharing=sharing, **kwargs)

tests/test_classify/test_classifier.py::test_init_resnet18_hf
tests/test_classify/test_classifier.py::test_infer_hf_models
  /data/dotstuff/local/mambaforge/envs/mapreader/lib/python3.10/site-packages/huggingface_hub/file_download.py:1132: FutureWarning: `resume_download` is deprecated and will be removed in version 1.0.0. Downloads always resume when possible. If you want to force a new download, use `force_download=True`.
    warnings.warn(

tests/test_classify/test_classifier.py::test_init_resnet18_hf
tests/test_classify/test_classifier.py::test_infer_hf_models
  /data/dotstuff/local/mambaforge/envs/mapreader/lib/python3.10/site-packages/transformers/models/convnext/feature_extraction_convnext.py:28: FutureWarning: The class ConvNextFeatureExtractor is deprecated and will be removed in version 5 of Transformers. Please use ConvNextImageProcessor instead.
    warnings.warn(

tests/test_classify/test_classifier.py::test_calculate_add_metrics
tests/test_classify/test_classifier.py::test_calculate_add_metrics
tests/test_classify/test_classifier.py::test_calculate_add_metrics
  /data/dotstuff/local/mambaforge/envs/mapreader/lib/python3.10/site-packages/sklearn/metrics/_classification.py:1509: UndefinedMetricWarning: Recall is ill-defined and being set to 0.0 in labels with no true samples. Use `zero_division` parameter to control this behavior.
    _warn_prf(average, modifier, f"{metric.capitalize()} is", len(result))

tests/test_geo_pipeline.py::test_pipeline
tests/test_geo_pipeline.py::test_pipeline
tests/test_geo_pipeline.py::test_pipeline
  /data/dotstuff/local/mambaforge/envs/mapreader/lib/python3.10/site-packages/sklearn/metrics/_classification.py:1509: UndefinedMetricWarning: Precision is ill-defined and being set to 0.0 in labels with no predicted samples. Use `zero_division` parameter to control this behavior.
    _warn_prf(average, modifier, f"{metric.capitalize()} is", len(result))

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
======================================================== 189 passed, 94 warnings
```

> Hi, we've also added how to run tests in the [Install](https://mapreader.readthedocs.io/en/latest/Install.html) instructions page of the docs. We'd like to keep everything in the docs vs. README as much as possible so its all in one place. Let me know if you think this needs more fleshing out though.

Yes, keeping it in one place which is easily found makes total sense to me. Thanks!

> The # is to do with being able to split the names back out (i.e. with str.split("#")) later on.

I see, thanks for explaining.

> > One curiosity, I see that the ROC AUC goes up to 100. I suppose this is percentage. I'm more used to see it go up to 1, but if it is percentage, I think adding the "%" would help.
>
> Again, sorry I missed this one, have just updated all the notebooks to have `(%)` in their y-labels.

No worries at all, it was totally minor. I see it is all changed and I'm happy to sign off on this.

Thanks again for your prompt replies and congrats on a great work!
