### 2024-05-13 #6434 @rwood-97
Great, very pleased to know its working!

Re. warnings, the rasterio ones are all to do with loading a non-georeferenced image (i.e. png file) in rasterio. Its part of a try/except where we try get georeferencing information from the image (e.g. if you loaded a tiff it could have it but not always) so we do it for all images. We probably should work out a way to suppress these warnings.
The other ones I also get but hadn't worked out/got around to addressing and don't (yet) seem to be a problem.
:)
