### 2024-05-13 #6434 @emdupre
Thank you @jordibc for reviewing these changes and @rwood-97 for your quick responses !

@PipGrylls, @geekysquirrel, please let us know when you also are able to re-review this submission for _MapReader_.

And thank you again to @ everyone for your work on this review to date !
