### 2024-05-14 d500aac
add device argument to docs
