+++ docs/source/User-guide/Classify/Infer.rst
@@ -10,6 +10,7 @@ To initialize your ``ClassifierContainer()`` for inference, you will need to def
 
 - ``model`` - The model (classifier) you would like to use.
 - ``labels_map`` - A dictionary mapping your labels to their indices (e.g. ``{0: "no_railspace", 1: "railspace"}``). This labels map should be the same as that used when training/fine-tuning the classifier.
+- ``device`` - The device you would like to use for inference (e.g. ``"cuda"``, ``"mps"`` or ``"cpu"``).
 
 There are a number of options for the ``model`` argument:
 
@@ -29,7 +30,9 @@ There are a number of options for the ``model`` argument:
             my_model = torch.load("./models/model_checkpoint_6.pkl")
             labels_map = {0: "no_railspace", 1: "railspace"}
 
-            my_classifier = ClassifierContainer(my_model, labels_map)
+            device = 'cuda' if torch.cuda.is_available() else 'mps' if torch.backends.mps.is_available() else 'cpu'
+
+            my_classifier = ClassifierContainer(my_model, labels_map, device=device)
 
         .. admonition:: Advanced usage
             :class: dropdown
@@ -56,13 +59,16 @@ There are a number of options for the ``model`` argument:
         .. code-block:: python
 
             #EXAMPLE
+            import torch
             from transformers import AutoFeatureExtractor, AutoModelForImageClassification
 
             extractor = AutoFeatureExtractor.from_pretrained("davanstrien/autotrain-mapreader-5000-40830105612")
             my_model = AutoModelForImageClassification.from_pretrained("davanstrien/autotrain-mapreader-5000-40830105612")
             labels_map = {0: "no_railspace", 1: "railspace"}
 
-            my_classifier = ClassifierContainer(my_model, labels_map)
+            device = 'cuda' if torch.cuda.is_available() else 'mps' if torch.backends.mps.is_available() else 'cpu'
+
+            my_classifier = ClassifierContainer(my_model, labels_map, device=device)
 
         .. note:: You will need to install the `transformers <https://github.com/huggingface/transformers>`__ library to do this (``pip install transformers``).
 
@@ -73,10 +79,14 @@ There are a number of options for the ``model`` argument:
 
             #EXAMPLE
             import timm
+            import torch
 
             my_model = timm.create_model("hf_hub:timm/resnest101e.in1k", pretrained=True, num_classes=len(annotated_images.labels_map))
+            labels_map = {0: "no_railspace", 1: "railspace"}
 
-            my_classifier = ClassifierContainer(my_model, annotated_images.labels_map, dataloaders)
+            device = 'cuda' if torch.cuda.is_available() else 'mps' if torch.backends.mps.is_available() else 'cpu'
+
+            my_classifier = ClassifierContainer(my_model, labels_map, device=device)
 
         .. note:: You will need to install the `timm <https://huggingface.co/docs/timm/index>`__ library to do this (``pip install timm``).
 
