### 2024-05-14 b58331c
add f-scores per class to docs
