### 2024-05-15 0a69fae
update pixel stats calculation
