+++ mapreader/spot_text/deepsolo_runner.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import os
 import pathlib
 import pickle
+from itertools import combinations
 
 try:
     import adet
@@ -213,6 +214,7 @@ class DeepSoloRunner:
         self,
         patch_df: pd.DataFrame = None,
         return_dataframe: bool = False,
+        min_ioa: float = 0.7,
     ) -> dict | pd.DataFrame:
         """Run the model on all images in the patch dataframe.
 
@@ -222,6 +224,8 @@ class DeepSoloRunner:
             Dataframe containing patch information, by default None.
         return_dataframe : bool, optional
             Whether to return the predictions as a pandas DataFrame, by default False
+        min_ioa : float, optional
+            The minimum intersection over area to consider two polygons the same, by default 0.7
 
         Returns
         -------
@@ -236,7 +240,7 @@ class DeepSoloRunner:
         img_paths = patch_df["image_path"].to_list()
 
         patch_predictions = self.run_on_images(
-            img_paths, return_dataframe=return_dataframe
+            img_paths, return_dataframe=return_dataframe, min_ioa=min_ioa
         )
         return patch_predictions
 
@@ -244,6 +248,7 @@ class DeepSoloRunner:
         self,
         img_paths: str | pathlib.Path | list,
         return_dataframe: bool = False,
+        min_ioa: float = 0.7,
     ) -> dict | pd.DataFrame:
         """Run the model on a list of images.
 
@@ -253,6 +258,8 @@ class DeepSoloRunner:
             A list of image paths to run the model on.
         return_dataframe : bool, optional
             Whether to return the predictions as a pandas DataFrame, by default False
+        min_ioa : float, optional
+            The minimum intersection over area to resolve overlapping polygons, by default 0.7
 
         Returns
         -------
@@ -264,7 +271,7 @@ class DeepSoloRunner:
             img_paths = [img_paths]
 
         for img_path in img_paths:
-            _ = self.run_on_image(img_path, return_outputs=False)
+            _ = self.run_on_image(img_path, return_outputs=False, min_ioa=min_ioa)
 
         if return_dataframe:
             return self._dict_to_dataframe(self.patch_predictions, geo=False)
@@ -275,6 +282,7 @@ class DeepSoloRunner:
         img_path: str | pathlib.Path,
         return_outputs=False,
         return_dataframe: bool = False,
+        min_ioa: float = 0.7,
     ) -> dict | pd.DataFrame:
         """Run the model on a single image.
 
@@ -286,6 +294,8 @@ class DeepSoloRunner:
             Whether to return the outputs direct from the model, by default False
         return_dataframe : bool, optional
             Whether to return the predictions as a pandas DataFrame, by default False
+        min_ioa : float, optional
+            The minimum intersection over area to resolve overlapping polygons, by default 0.7
 
         Returns
         -------
@@ -304,16 +314,17 @@ class DeepSoloRunner:
         if return_outputs:
             return outputs
 
-        self.get_patch_predictions(outputs)
+        self.get_patch_predictions(outputs, min_ioa=min_ioa)
 
         if return_dataframe:
-            return self._dict_to_dataframe(self.patch_predictions, patch=True)
+            return self._dict_to_dataframe(self.patch_predictions, geo=False)
         return self.patch_predictions
 
     def get_patch_predictions(
         self,
         outputs: dict,
         return_dataframe: bool = False,
+        min_ioa: float = 0.7,
     ) -> dict | pd.DataFrame:
         """Post process the model outputs to get patch predictions.
 
@@ -323,6 +334,8 @@ class DeepSoloRunner:
[…]
