### 2024-05-16 c65ff77
move common functions to base class
