+++ mapreader/spot_text/deepsolo_runner.py
@@ -1,18 +1,15 @@
 from __future__ import annotations
 
-import os
 import pathlib
 import pickle
-from itertools import combinations
 
 try:
     import adet
 except ImportError:
     raise ImportError(
-        "Please install DeepSolo from the following link: https://github.com/rwood-97/DeepSolo"
+        "[ERROR] Please install DeepSolo from the following link: https://github.com/rwood-97/DeepSolo"
     )
 
-import geopandas as geopd
 import numpy as np
 import pandas as pd
 from adet.config import get_cfg
@@ -20,21 +17,20 @@ from adet.config import get_cfg
 try:
     from detectron2.engine import DefaultPredictor
 except ImportError:
-    raise ImportError("Please install Detectron2")
+    raise ImportError("[ERROR] Please install Detectron2")
 
-import matplotlib.patches as patches
-import matplotlib.pyplot as plt
-from PIL import Image
 from shapely import LineString, Polygon
 
 # first assert we are using the deep solo version of adet
 if adet.__version__ != "0.2.0-deepsolo":
     raise ImportError(
-        "Please install DeepSolo from the following link: https://github.com/rwood-97/DeepSolo"
+        "[ERROR] Please install DeepSolo from the following link: https://github.com/rwood-97/DeepSolo"
     )
 
+from .runner_base import Runner
 
-class DeepSoloRunner:
+
+class DeepSoloRunner(Runner):
     def __init__(
         self,
         patch_df: pd.DataFrame = None,
@@ -210,116 +206,6 @@ class DeepSoloRunner:
         # setup the predictor
         self.predictor = DefaultPredictor(cfg)
 
-    def run_all(
-        self,
-        patch_df: pd.DataFrame = None,
-        return_dataframe: bool = False,
-        min_ioa: float = 0.7,
-    ) -> dict | pd.DataFrame:
-        """Run the model on all images in the patch dataframe.
-
-        Parameters
-        ----------
-        patch_df : pd.DataFrame, optional
-            Dataframe containing patch information, by default None.
-        return_dataframe : bool, optional
-            Whether to return the predictions as a pandas DataFrame, by default False
-        min_ioa : float, optional
-            The minimum intersection over area to consider two polygons the same, by default 0.7
-
-        Returns
-        -------
-        dict or pd.DataFrame
-            A dictionary of predictions for each patch image or a DataFrame if `as_dataframe` is True.
-        """
-        if patch_df is None:
-            if self.patch_df is not None:
-                patch_df = self.patch_df
-            else:
-                raise ValueError("[ERROR] Please provide a `patch_df`")
-        img_paths = patch_df["image_path"].to_list()
-
-        patch_predictions = self.run_on_images(
-            img_paths, return_dataframe=return_dataframe, min_ioa=min_ioa
-        )
-        return patch_predictions
-
-    def run_on_images(
-        self,
-        img_paths: str | pathlib.Path | list,
-        return_dataframe: bool = False,
-        min_ioa: float = 0.7,
-    ) -> dict | pd.DataFrame:
-        """Run the model on a list of images.
-
-        Parameters
-        ----------
-        img_paths : str, pathlib.Path or list
-            A list of image paths to run the model on.
-        return_dataframe : bool, optional
-            Whether to return the predictions as a pandas DataFrame, by default False
-        min_ioa : float, optional
-            The minimum intersection over area to resolve overlapping polygons, by default 0.7
-
-        Returns
-        -------
-        dict or pd.DataFrame
-            A dictionary of predictions for each image or a DataFrame if `as_dataframe` is True.
-        """
-
-        if isinstance(img_paths, (str, pathlib.Path)):
-            img_paths = [img_paths]
-
-        for img_path in img_paths:
-            _ = self.run_on_image(img_path, return_outputs=False, min_ioa=min_ioa)
[…]

[…]
