### 2024-05-16 f3d872e
add deduplicate for parent images (e.g. if there is overlap between patches)
