### 2024-05-29 #6434 @emdupre
Hi everyone,

I just wanted to follow-up on this submission for _MapReader_.

Currently, it looks like @geekysquirrel 's testing concerns in https://github.com/Living-with-machines/MapReader/issues/369 is still open ; could you please confirm that this is waiting for reviewer input @rwood-97 @kmcdono2 ?

Otherwise, @PipGrylls [had raised this point](https://github.com/openjournals/joss-reviews/issues/6434#issuecomment-2036751956) on example usage:
> Currently the documentation for worked examples is unclear. The [docs link to ](https://github.com/Living-with-machines/MapReader/tree/main/worked_examples)this. I think a README here explaining how to launch the examples and a list of examples is whats needed. Running the annotation notebooks and the worked_examples/geospatial/classification_one_inch_maps/Pipeline.ipynb notebook seems like this is well done with just some signposting needed.

I believe https://github.com/Living-with-machines/MapReader/issues/386 is intended to address this, but it would be great to get confirmation from both the authors as well as @PipGrylls himself !

Otherwise, please let us know if you have any remaining comments @jordibc @geekysquirrel @PipGrylls , and thank you again for all of your work to date !
