### 2024-05-29 Living-with-machines/MapReader#369 @kmcdono2
I think this is resolved with the fix of using the most recent mapreader version. But would be great for @geekysquirrel to confirm.
