### 2024-05-29 #6434 @PipGrylls
I’ve installed mapreader and run the tests on a M2 Mac and it now works well.

My previous concern as having the recommended install pathway mixing conda and pip remains but isn't a blocker. In an ideal world there would be a:
- 'conda install’: using conda and the anaconda channels
- ‘pypi install’: using venv/virtual environments and PyPi
- 'source install’: using venv/virtual environments and the Github (this would be the way to run the tests)

I think finally I would recommend having the example notebooks being cloned from a repo that does not also contain the source code so people can clone only them. (you could then use submodules to make this part of the source module to avoid maintaining multiple versions)
