### 2024-05-31 #6434 @rwood-97
> I think finally I would recommend having the example notebooks being cloned from a repo that does not also contain the source code so people can clone only them. (you could then use submodules to make this part of the source module to avoid maintaining multiple versions)

I like this idea, will make an issue on the MapReader repo to address this at some point (probably won't happen soon but good to have in our backlog)
