### 2024-05-31 maps-as-data/MapReader#437 @rwood-97
Move worked examples to separate repo
As per suggestion in https://github.com/openjournals/joss-reviews/issues/6434#issuecomment-2136797648:

> I think finally I would recommend having the example notebooks being cloned from a repo that does not also contain the source code so people can clone only them. (you could then use submodules to make this part of the source module to avoid maintaining multiple versions)

Would be good to move worked examples to separate repo.
This could be under LwM org and linked to from the docs/ MapReader repo.

---

Related issues:
- [x] #425
