### 2024-06-12 Living-with-machines/MapReader#369 @emdupre
Hi everyone, I just wanted to follow-up that I've confirmed with @geekysquirrel via email that they're happy with the current status of the codebase and are OK to move forward.

I'll thus consider this issue closed for the purposes of the JOSS review !
