### 2024-06-12 #6434 @emdupre
Hi everyone,

Confirming that I've now heard back from all reviewers, and I will now go ahead and perform a few editorial checks to continue processing this submission. Thank you again @PipGrylls, @jordibc, @geekysquirrel for your reviews of _MapReader_ for JOSS ! 💐
