### 2024-06-17 #6434 @rwood-97
Hi, just wanted to check whether these are tasks we should do now or wait until the editor tasks are complete?
