### 2024-06-19 fa9a969
Update authors paper.md
