+++ .github/workflows/metrics.yml
@@ -1,0 +1,16 @@
+on: [push]
+
+jobs:
+  paper:
+    runs-on: ubuntu-latest
+    name: Metrics
+    steps:
+      - name: Metrics embed
+  
+  # You may pin to the exact commit or the version.
+  # uses: lowlighter/metrics@65836723097537a54cd8eb90f61839426b4266b6
+  uses: lowlighter/metrics@v3.34
+    with:
+      template: repository
+      user: living-with-machines
+      repo: mapreader
