### 2024-06-22 b40ca5c
Minor updates to software paper for JOSS review
