### 2024-06-22 #6434 @emdupre
Thank you for your patience, @rwood-97  ! I've now finished reviewing both the paper as well as the references and have a only a few remaining editorial requests on both :

- I've [opened a pull request](https://github.com/Living-with-machines/MapReader/pull/446) with minor edits to the text of the paper. Please review these and let me know if you have any concerns or questions !
- Could you please update the mapkurator citation to include the references they request in their documentation, [as described here ](https://knowledge-computing.github.io/mapkurator-doc/#/docs/introduction?id=cite) ?
- It looks like the Corcoran citation is listed as "in preparation." Is there a preprint or another source that could be directly cited, here ?
