### 2024-06-25 169072d
Adding in shortcuts for text detection for easier import
