### 2024-06-26 f369714
Adding URL to `Detectron2`
