### 2024-06-26 dd2414b
Adding in `logger` commands instead of `print`
