+++ mapreader/load/geo_utils.py
@@ -1,10 +1,14 @@
 #!/usr/bin/env python
 from __future__ import annotations
 
+import logging
+
 import numpy as np
 import rasterio
 from geopy.distance import geodesic, great_circle
 from pyproj import Transformer
+
+logger = logging.getLogger(__name__)
 
 
 def extractGeoInfo(image_path):
@@ -34,7 +38,7 @@ def extractGeoInfo(image_path):
         tiff_coord = tuple(tiff_src.bounds)
 
     print(f"[INFO] Shape: {tiff_shape}. \n[INFO] CRS: {tiff_proj}.")
-    print("[INFO] Coordinates: {:.4f} {:.4f} {:.4f} {:.4f}".format(*tiff_coord))
+    logger.info("Coordinates: {:.4f} {:.4f} {:.4f} {:.4f}".format(*tiff_coord))
 
     return tiff_shape, tiff_proj, tiff_coord
 
@@ -62,7 +66,7 @@ def reproject_geo_info(image_path, target_crs="EPSG:4326", calc_size_in_m=False)
     transformer = Transformer.from_crs(tiff_proj, target_crs, always_xy=True)
     coord = transformer.transform_bounds(*tiff_coord)
     print(f"[INFO] New CRS: {target_crs}")
-    print("[INFO] Reprojected coordinates: {:.4f} {:.4f} {:.4f} {:.4f}".format(*coord))
+    logger.info("Reprojected coordinates: {:.4f} {:.4f} {:.4f} {:.4f}".format(*coord))
 
     height, width, _ = tiff_shape
 
