+++ mapreader/spot_text/deepsolo_runner.py
@@ -33,6 +33,7 @@ if adet.__version__ != "0.2.0-deepsolo":
         "Please install DeepSolo from the following link: https://github.com/rwood-97/DeepSolo"
     )
 
+# Set up logging
 logger = logging.getLogger(__name__)
 
 
