+++ paper/paper.bib
@@ -63,14 +63,12 @@
 	pages = {103711},
 }
 
-@misc{mapkurator,
-	title = {{mapKurator}},
-	url = {https://knowledge-computing.github.io/mapkurator-doc/#/},
-	abstract = {Description},
-	urldate = {2023-11-23},
-	author = {Knowledge Computing Lab},
-	publisher = {Github},
-	journal = {Github repository},
+@inproceedings{mapkurator,
+  title={An automatic approach for generating rich, linked geo-metadata from historical map images},
+  author={Li, Zekun and Chiang, Yao-Yi and Tavakkol, Sasan and Shbita, Basel and Uhl, Johannes H and Leyk, Stefan and Knoblock, Craig A},
+  booktitle={Proceedings of the 26th ACM SIGKDD International Conference on Knowledge Discovery \& Data Mining},
+  pages={3290--3298},
+  year={2020}
 }
 
 @article{Petitpierre,
