### 2024-07-02 #6434 @kmcdono2
I reviewed your pull request and made 1 request to not edit as suggested, otherwise fine.

I have made the update to the mapkurator citation here: https://github.com/Living-with-machines/MapReader/pull/449

I have emailed Corcoran about whether her citation can be updated.
