### 2024-07-02 6d68559
paper update to mention text spotting
