+++ paper/paper.bib
@@ -106,5 +106,5 @@
 	journal = {Frontiers in Plant Science},
 	author = {Corcoran, E and Hosseini, K and Siles, L and Kurup, S and Ahnert, S},
 	year = {2023},
-	pages = {In preparation},
+	pages = {Under review},
 }
