### 2024-07-03 #6434 @rwood-97
Just updated the repo with a release for the JOSS paper. This is v1.3.3 and can be found [here](https://github.com/Living-with-machines/MapReader/releases/tag/v1.3.3).

Link for draft of zenodo release: https://zenodo.org/uploads/12636179?token=eyJhbGciOiJIUzUxMiJ9.eyJpZCI6ImU0MjI2ZDEzLThhZjUtNDZkZS1iMTFmLTQ0NDU3NTBkZGI4MCIsImRhdGEiOnt9LCJyYW5kb20iOiIzOTFiOWYzMWI0NWRkNThhYTJjNjc0M2YyMmViYTcxZCJ9.vAW3VwdfXQbzYqWhZySYN7KXDjoj8PHf7ETCW_-pqTtxtCyH9HjHEb94AAkMUR66qS390ASkgSpWpMLXiigepQ

Is this correct? I can edit if not
