### 2024-07-03 169fea3
Editing upcoming communityevents
