+++ docs/source/Events.rst
@@ -3,38 +3,53 @@ Events
 =======
 
 Community Calls
-----------------
+---------------
+
 We host online community calls every 2 months for friends of MapReader to meet up and discuss work in progress, feature requests, shared development tasks, or other issues and ideas. All are welcome. Details for joining are available in the Google notes doc for each call.
 
-*Upcoming calls*
+Upcoming calls
+~~~~~~~~~~~~~~
 
-- June 27, 2024, 4-5pm UK time - `Notes and zoom link <https://docs.google.com/document/d/1VCsFZgJjJ4XdmwzskRfsLdrZy4Xx1huMu-WZsj7wshE>`_
+- August 1, 2024, 4-5pm UK time
+  - `Sign up for our community calls: <https://forms.office.com/e/x9zGL0yQyp>`_
+  - `Notes and Zoom link for this month's meeting <https://docs.google.com/document/d/1UE2OjPE7OitoB8AsG4swk1k3pD44lW_YgMkaR4VxUXU/edit>`_
 
 - September 5, 2024, 4-5pm UK time
+  - `Sign up for our community calls: <https://forms.office.com/e/x9zGL0yQyp>`_
 
-*Past calls*
+Past calls
+~~~~~~~~~~
 
-- April 18, 2024, 4-5pm UK time - `Notes and zoom link <https://docs.google.com/document/d/1CJFzwQzg9ZgviKVr9T_yKAY_SIzC6tl-sEOyJlpWMnw/edit#heading=h.t71vmryvuznf>`_
+- April 18, 2024, 4-5pm UK time - `Notes and zoom link <https://docs.google.com/document/d/1CJFzwQzg9ZgviKVr9T_yKAY_SIzC6tl-sEOyJlpWMnw/edit>`_
 
-- February 15, 2024, 4-5pm UK time - `Notes <https://docs.google.com/document/d/155VlsYRbOEmmQDpDflWcTipFTfsKuPACWWmTKh6M820/edit#heading=h.t71vmryvuznf>`_
+- February 15, 2024, 4-5pm UK time - `Notes <https://docs.google.com/document/d/155VlsYRbOEmmQDpDflWcTipFTfsKuPACWWmTKh6M820/edit>`_
 
 
 
 Workshops
-----------------
+---------
 
 Thanks to the new AHRC-supported project `Data/Culture <https://www.turing.ac.uk/research/research-projects/dataculture-building-sustainable-communities-around-arts-and-humanities>`_ based at The Alan Turing Institute, we are offering a series of hands-on workshops to support a growing user community around MapReader during 2024.
 
+Upcoming workshops
+~~~~~~~~~~~~~~~~~~
 
-**Spring Data/Culture Workshop: MapReader**
+**MapReader Workshop at Spatial Humanities 2024**
 
-April 30-May 1, 2024 at The Alan Turing Institute, London, UK
+September 25, 2024 in Bamberg, Germany
 
-- 2-day workshop for historians and other historically-inclined researchers about learning to use MapReader. 
+- 1/2 day workshop co-located with the Spatial Humanities conference
+- `More details <https://spathum.uni-bamberg.de/>`_
 
-- Bursaries available.
 
-- Workshop webpage: https://www.turing.ac.uk/events/spring-dataculture-workshop-search-inside-maps-mapreader 
+**Fall Data/Culture Workshop: MapReader + Newspapers**
+
+September 30-October 2 at The Alan Turing Institute, London, UK (and partially online/hybrid)
+
+
+
+Past workshops
+~~~~~~~~~~~~~~
 
 **Summer Data/Culture Workshop: MapReader + Newspapers**
 
@@ -46,18 +61,16 @@ June 5-7, 2024 in Lancaster, UK
 
 - Bursaries available.
 
[…]
