### 2024-07-03 e2bf35c
Redoing list of community calls
