+++ docs/source/index.rst
@@ -1,9 +1,4 @@
-.. MapReader documentation master file, created by
-   sphinx-quickstart on Wed Jan 25 11:03:05 2023.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-MapReader
+apReader
 =========
 
 **A computer vision pipeline for exploring and analyzing images at scale**
