+++ docs/source/index.rst
@@ -1,4 +1,4 @@
-apReader
+MapReader
 =========
 
 **A computer vision pipeline for exploring and analyzing images at scale**
