### 2024-07-07 2fc24c2
Updating tables for workshops
