+++ docs/source/Events.rst
@@ -5,10 +5,12 @@ Events
 Community Calls
 ---------------
 
-We host online community calls every 2 months for friends of MapReader to meet up and discuss work in progress, feature requests, shared development tasks, or other issues and ideas. All are welcome. Details for joining are available in the Google notes doc for each call.
+We host online community calls every 2 months for friends of MapReader to meet
+up and discuss work in progress, feature requests, shared development tasks,
+or other issues and ideas. All are welcome. Details for joining are available
+in the Google notes doc for each call.
 
-Upcoming calls
-~~~~~~~~~~~~~~
+.. ----- Upcoming community calls ---------------------------------------------
 
 .. list-table:: Upcoming community calls
    :widths: 25 25 50
@@ -24,8 +26,7 @@ Upcoming calls
      - `Sign up <https://forms.office.com/e/x9zGL0yQyp>`_
      - *Coming soon*
 
-Past calls
-~~~~~~~~~~
+.. ----- Past community calls -------------------------------------------------
 
 .. list-table:: Past community calls
    :widths: 25 50
@@ -33,57 +34,80 @@ Past calls
 
    * - Date
      - Notes and Zoom link
+
    * - April 18, 2024, 4-5pm UK time
-     - `Notes and Zoom link <https://docs.google.com/document/d/1CJFzwQzg9ZgviKVr9T_yKAY_SIzC6tl-sEOyJlpWMnw/edit>`_
+     - `April 2024 document <https://docs.google.com/document/d/1CJFzwQzg9ZgviKVr9T_yKAY_SIzC6tl-sEOyJlpWMnw/edit>`_
+
    * - February 15, 2024, 4-5pm UK time
-     - `Notes and Zoom link <https://docs.google.com/document/d/155VlsYRbOEmmQDpDflWcTipFTfsKuPACWWmTKh6M820/edit>`_
+     - `February 2024 document <https://docs.google.com/document/d/155VlsYRbOEmmQDpDflWcTipFTfsKuPACWWmTKh6M820/edit>`_
 
 
 Workshops
 ---------
 
-Thanks to the new AHRC-supported project `Data/Culture <https://www.turing.ac.uk/research/research-projects/dataculture-building-sustainable-communities-around-arts-and-humanities>`_ based at The Alan Turing Institute, we are offering a series of hands-on workshops to support a growing user community around MapReader during 2024.
+Thanks to the new AHRC-supported project
+`Data/Culture <https://www.turing.ac.uk/research/research-projects/dataculture-building-sustainable-communities-around-arts-and-humanities>`_
+based at The Alan Turing Institute, we are offering a series of hands-on
+workshops to support a growing user community around MapReader during 2024.
 
-Upcoming workshops
-~~~~~~~~~~~~~~~~~~
+.. ----- Upcoming workshops ---------------------------------------------------
 
-**MapReader Workshop at Spatial Humanities 2024**
+.. list-table:: Upcoming workshops
+   :header-rows: 1
 
-September 25, 2024 in Bamberg, Germany
+   * - Date and location
+     - Workshop name
+     - Description
+     - More details
 
-- 1/2 day workshop co-located with the Spatial Humanities conference
-- `More details <https://spathum.uni-bamberg.de/>`_
+   * - Sept 25, 2024
+     - MapReader Workshop
+       
+       at Spatial Humanities 2024
+     - 1/2 day workshop co-located with the Spatial Humanities conference.
+     - `More details <https://spathum.uni-bamberg.de/>`_
 
-
-**Fall Data/Culture Workshop: MapReader + Newspapers**
-
[…]
