### 2024-07-08 dfe4ef1
`pre-commit` and `ruff` are currently not used
