### 2024-07-09 9019178
Moving `pre–commit` and `ruff` files back after @rwood-97 comment
