+++ .pre-commit-config.yaml
@@ -1,0 +1,36 @@
+repos:
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: "v4.4.0"
+    hooks:
+      - id: check-added-large-files
+      - id: check-case-conflict
+      - id: check-merge-conflict
+      - id: check-symlinks
+      - id: check-yaml
+      - id: debug-statements
+      - id: end-of-file-fixer
+      - id: mixed-line-ending
+      - id: name-tests-test
+        args: ["--pytest-test-first"]
+      - id: requirements-txt-fixer
+      - id: trailing-whitespace
+
+  - repo: https://github.com/psf/black-pre-commit-mirror
+    rev: "23.9.1"
+    hooks:
+      - id: black-jupyter
+        language_version: python3.9
+
+  - repo: https://github.com/astral-sh/ruff-pre-commit
+    rev: "v0.0.290"
+    hooks:
+      - id: ruff
+        args: ["--fix", "--show-fixes"]
+
+exclude: |
+  (?x)^(
+    conda/.*|
+    worked_examples/.*|
+    .*/__init__.py|
+    tests/test_import.py
+  )$
