+++ .ruff.toml
@@ -1,0 +1,20 @@
+select = [
+  "E", "F", "W", # flake8
+  "B",           # flake8-bugbear
+  "I",           # isort
+  "UP"           # pyupgrade
+#  "N",           # pep8 naming
+]
+ignore = [
+  "E501",   # Line too long
+  "E722",   # Do not use bare 'except'
+  "B904",   # Within an 'except' clause ...
+]
+unfixable = [
+  "F841", # Removes unused variables
+]
+flake8-unused-arguments.ignore-variadic-names = true
+isort.required-imports = ["from __future__ import annotations"]
+
+# Assume Python 3.9
+target-version = "py39"
