### 2024-07-09 Living-with-machines/MapReader#437 @kallewesterling
I have taken a first stab at this here:

https://github.com/Living-with-machines/MapReader-examples

It is currently private, but what's nice about it is that it's containerised, which means that you could load it like this, after building the image and pushing it to docker hub:

```
docker run -p 8888:8888 kallewesterling/mapreader-examples
```
