### 2024-07-09 Living-with-machines/MapReader#437 @kallewesterling
It has a new script called [`download-examples.py`](https://github.com/Living-with-machines/MapReader-examples/blob/main/download-examples.py) which will read [a YAML file](https://github.com/Living-with-machines/MapReader-examples/blob/main/download-examples.yaml) with all the pre-requirements for each notebook, and run the downloader for those particular queries.
