### 2024-07-09 Living-with-machines/MapReader#437 @kallewesterling
The Python file has this example of a `download-examples.yaml` file, but here they are for ease of access:

```yaml
- notebook: path/to/notebook.ipynb
  metadata: path/to/metadata.json
  path_save: maps
  zoom: 14
  queries:
    coordinates:
      - -4.33, 55.90
      - -3.25, 51.93
- notebook: path/to/notebook2.ipynb
  metadata: path/to/metadata.json
  path_save: maps
  zoom: 14
  queries:
    polygons:
      latlons:
        - 55.6, -3.5, 56, -2.8
      modes:
        - within
        - intersects
    coordinates:
      - -4.5, 55.4
    lines:
      latlons:
        - xy1: 56.5, -5
          xy2: 57.0, -4.5
    wfs_ids:
      - 12
      - 30
      - 37
      - 38
    string:
      - literal: "Stirling"
      - property: IMAGEURL
        literal: 74487492
```
