### 2024-07-09 1862490
Splitting up "beginners' info" into separate pages and adding more content
