### 2024-07-09 070bbcb
Add `sphinx-autobuild` for ease of developing documentation Fixes #456
