### 2024-07-09 94b7d9a
Adding a fallback option of waiting 5 seconds and trying again Fixes #442
