+++ mapreader/download/sheet_downloader.py
@@ -618,6 +618,7 @@ class SheetDownloader:
         else:
             print(f'[WARNING] Download of "{img_path}" was unsuccessful.')
 
+        # Try to remove the temporary folder
         try:
             shutil.rmtree(DEFAULT_TEMP_FOLDER)
         except PermissionError:
@@ -628,6 +629,7 @@ class SheetDownloader:
             # try again
             time.sleep(5)
             shutil.rmtree(DEFAULT_TEMP_FOLDER)
+
         return img_path
 
     def _save_metadata(
