+++ mapreader/download/downloader.py
@@ -108,11 +108,13 @@ class Downloader:
             map_name = self.merger._get_output_name(grid_bb)
         self.downloader.download_tiles(grid_bb, download_in_parallel=False)
         success = self.merger.merge(grid_bb, map_name)
+
         if success:
             print(f'[INFO] Downloaded "{map_name}.png"')
         else:
             print(f'[WARNING] Download of "{map_name}.png" was unsuccessful.')
 
+        # Try to remove the temporary folder
         try:
             shutil.rmtree(DEFAULT_TEMP_FOLDER)
         except PermissionError:
