### 2024-07-10 Living-with-machines/MapReader#437 @rwood-97
Some comments from me (I can make these issues, or do a PR, on the other repo if you like):
- I think in the main README it should be clear that you can just clone and run the notebooks in the repo, i.e. you don't have to use docker. My instinct is that most users will not be familiar/comfortable using docker so will want to use them just by cloning and running on their laptop so we should have very clear instructions for how to do this.
- We might need to update some links in the notebook READMEs
- Could we change the name of the `test_data` function since `test_` is used by pytest to find test functions. I am thinking that a longer term goal for this repo could be to have tests to make sure the notebooks work with any updates to the MapReader code. i.e. tests would run every time there is a version update on the MapReader repo.

I'm just in the process of building the docker image so will let you know any other comments when thats done
