### 2024-07-10 Living-with-machines/MapReader#437 @rwood-97
Okay on the docker:
- Potentially good to use python 3.12
- Some of the notebooks require specific MapReader versions (e.g. I put in the 2024 notebook that it should be 1.3.2 to stop people having old MapReader versions when running), is there a way round this in docker or maybe we only add notebooks which don't have this)
- A couple of file paths need updating I think - I guess we need to make sure file paths are same for people running normally and via docker, so the file structure needs to be the same for both.
