### 2024-07-10 087365b
add note about opening the url manually if needed
