+++ docs/source/Contribution-guide/Documentation.rst
@@ -112,6 +112,10 @@ To do this, navigate to the ``MapReader/docs`` directory and run:
     make livehtml
 
 This will build the documentation and open a new tab in your browser with the documentation.
+
+.. note::
+    If a new tab does not open automatically, you can navigate to ``http://127.0.0.1:8000`` in your browser to view the live documentation.
+
 The ``livehtml`` command will automatically update the documentation as you make changes to the files.
 
 When you are finished
