### 2024-07-12 #6434 @emdupre
Hi all,

Thanks for sending these updates, @kmcdono2 and @rwood-97 !

Confirming that I also cannot see the Zenodo draft and so cannot check the archive. Please let me know when this is accessible (usually by publishing the record).

On my side, I'm happy to update or close the PR if you'd like to make the remaining changes to the software paper.
Of note, for the Corcoran reference, the bibtex entry will need to be updated to the following :

```
@unpublished{Corcoran,
	title = {Automated extraction of dynamic phenotype data from whole plant images collected under controlled conditions},
	author = {Corcoran, E and Hosseini, K and Siles, L and Kurup, S and Ahnert, S},
	note = {Under review},
}
```
