### 2024-07-15 7acc97c
Adding all the commits into CHANGELOG for easier writing of bullet points per release
