+++ CHANGELOG.md
@@ -12,7 +12,1680 @@ The following table shows which versions of MapReader are compatible with which 
 +----------+------------+
 
 
-## v1.2.0 (2024-04-08)
+## [v1.2.0](https://github.com/Living-with-machines/MapReader/releases/tag/v1.3.4) (2024-04-08)
+
+<!--
+-->
 
 - Remove support for python 3.8
 - Update dependencies and add cartopy as required dependency
+
+---
+
+## [v1.3.4](https://github.com/Living-with-machines/MapReader/releases/tag/v1.3.4) (2024-07-05)
+
+<!--
+add823a Merge pull request #435 from Living-with-machines/dev_text_spotting
+1d76188 add test for overlap
+6b2f282 fix ioa figure
+a25251c update docs
+1ee4fdf Missing M
+2942935 add parent deduplication
+9d67955 Removing placeholder text
+e2bf35c Redoing list of community calls
+169fea3 Editing upcoming communityevents
+-->
+
+-
+
+## [v1.3.3](https://github.com/Living-with-machines/MapReader/releases/tag/v1.3.3) (2024-07-03)
+
+<!--
+cfc90e7 Merge pull request #450 from Living-with-machines/katie-paper-final-read
+8cde900 rosie's edit
+bd3d5f6 Merge pull request #449 from Living-with-machines/km-citation-update
+382ec44 corcoran update
+6d68559 paper update to mention text spotting
+42ab95d update mapkurator citation
+c6d078b Merge pull request #443 from Living-with-machines/rwood-97-patch-1
+fa9a969 Update authors paper.md
+c3cc3bd Merge branch 'main' into dev_text_spotting
+6f19e41 Merge branch 'main' of https://github.com/Living-with-machines/MapReader
+d2af7ab fix workshop notebook part3
+5b5db89 update June date
+8ad642d Merge pull request #438 from Living-with-machines/text-on-maps-viz
+27420d2 clear outputs, filter for one parent map to save time, add a bit of text descriptions
+2e6f997 Merge branch 'main' into text-on-maps-viz
+1fd75c9 Merge pull request #424 from Living-with-machines/workshop_feedback
+5040a2d updates from Katie feedback
+011516c Update Events.rst
+b0e04ec add reference for multilingual models
+fd3ebda add link to NLS website
+50d5a46 add text exploration notebook
+816e6f9 update mapreader version
+646f8f0 ensure compatibility with geopands 1.0.0a1 pre-release
+b16a771 ignore workshop outputs
+6f49ee0 update uk viz
+b9784b1 small updates to plots
+e5b8470 add data viz notebooks
+cc73913 Add data/culture grant no. to paper
+c238d9b add method for creating overlapping patches
+733c483 rename workshops for 2024
+f3d872e add deduplicate for parent images (e.g. if there is overlap between patches)
+c65ff77 move common functions to base class
+5d34bdf add deduplicate code to both runners
+-->
+
+-
+
+## [v1.3.2](https://github.com/Living-with-machines/MapReader/releases/tag/v1.3.2) (2024-05-17)
+
+<!--
+282b873 Merge pull request #434 from Living-with-machines/426-calc-pixel-stats
+0a69fae update pixel stats calculation
+ed2af0b Merge pull request #433 from Living-with-machines/430-add-id
+5c63c92 fix add_id bug
+20bdbd0 Merge pull request #431 from Living-with-machines/64-commandline
+db75647 Merge pull request #432 from Living-with-machines/400-review-labels
+ca42977 fix review_labels
+2b3b985 fix command line script
+b58331c add f-scores per class to docs
+caa6547 add printing f-scores per class
+acd9a7f fix typo
[…]
