+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The following table shows which versions of MapReader are compatible with which 
 ## [v1.3.4](https://github.com/Living-with-machines/MapReader/releases/tag/v1.3.4) (2024-07-05)
 
 <!--
+PRs: #435
 add823a Merge pull request #435 from Living-with-machines/dev_text_spotting
 1d76188 add test for overlap
 6b2f282 fix ioa figure
@@ -41,6 +42,8 @@ e2bf35c Redoing list of community calls
 ## [v1.3.3](https://github.com/Living-with-machines/MapReader/releases/tag/v1.3.3) (2024-07-03)
 
 <!--
+PRs: #450, #449, #443, #438, #428, #424
+
 cfc90e7 Merge pull request #450 from Living-with-machines/katie-paper-final-read
 8cde900 rosie's edit
 bd3d5f6 Merge pull request #449 from Living-with-machines/km-citation-update
@@ -81,6 +84,8 @@ c65ff77 move common functions to base class
 ## [v1.3.2](https://github.com/Living-with-machines/MapReader/releases/tag/v1.3.2) (2024-05-17)
 
 <!--
+PRs: #434, #433, #431, #432, #418, #415, #414, #416, #396, #413
+
 282b873 Merge pull request #434 from Living-with-machines/426-calc-pixel-stats
 0a69fae update pixel stats calculation
 ed2af0b Merge pull request #433 from Living-with-machines/430-add-id
@@ -136,6 +141,8 @@ fe370ee Merge pull request #413 from Living-with-machines/issue-398
 ## [v1.3.1](https://github.com/Living-with-machines/MapReader/releases/tag/v1.3.1) (2024-05-03)
 
 <!--
+PRs: #399
+
 3d7ab10 update tests
 dd216eb fix overwriting issue
 1eb5aeb update saving for duplicate sheet names
@@ -150,6 +157,8 @@ e81c617 Merge branch 'main' into april_workshop
 ## [v1.3.0](https://github.com/Living-with-machines/MapReader/releases/tag/v1.3.0) (2024-04-29)
 
 <!--
+PRs: #388, 395, 384
+
 f256ae9 Merge pull request #388 from Living-with-machines/ocr
 6237341 update where to find configs and weights
 197e4dd update docs for install
@@ -208,6 +217,8 @@ f42b1c0 update installation instructions
 ## [v1.2.0](https://github.com/Living-with-machines/MapReader/releases/tag/v1.2.0) (2024-04-08)
 
 <!--
+PRs: #383, #380, #381, #382, #379
+
 e9d119f update docs
 5a62b69 move cartopy to required and update version
 d5ab0b8 force int for randint
@@ -235,6 +246,8 @@ a284f4e Update pytest requirement from <8.0.0 to <9.0.0
 ## [v1.1.5](https://github.com/Living-with-machines/MapReader/releases/tag/v1.1.5) (2024-04-04)
 
 <!--
+PRs: #378, #377, #374, #373, #372, #366, #363, #362
+
 75e2824 Merge pull request #378 from Living-with-machines/JOSS_paper
 0a61a89 Merge pull request #377 from Living-with-machines/rw_docs
 c1b9223 Update Install.rst
@@ -275,6 +288,8 @@ f106c95 Update Events.rst
 ## [v1.1.4](https://github.com/Living-with-machines/MapReader/releases/tag/v1.1.4) (2024-02-23)
 
 <!--
+PRs: #357
+
 4ae6026 Merge pull request #357 from Living-with-machines/rw_docs
 1d1ca54 Update Worked-examples.rst
 deeb7cb update docs
[…]
