### 2024-07-15 c2eae4d
Moving 1.2.0 bullet points in CHANGELOG
