### 2024-07-17 fa0e76e
Editing language for consistency Adding ``csv`` instead of csv, Python instead of python, and Jupyter instead of jupyter
