+++ docs/source/About.rst
@@ -60,7 +60,7 @@ What skills/knowledge will I need to use MapReader?
 
 * Understanding of your map collection and knowledge of visual features you would like to identify within your maps
 * Basic understanding of how to use your terminal
-* Basic python
+* Basic Python
 * Basic understanding of machine learning and computer vision (CV) methodology
 
 What are the inputs and outputs of each stage in the MapReader pipeline?
