+++ docs/source/User-guide/Post-process.rst
@@ -10,7 +10,7 @@ The user can then choose how to relabel the patch (e.g. 'railspace' -> 'no').
 To run the post-processing code, you will need to have saved the predictions from your model in the format expected for the post-processing code.
 See the :doc:`/User-guide/Classify/Classify` docs for more on this.
 
-If you have your predictions saved in a csv file, you will first need to load them into a pandas DataFrame:
+If you have your predictions saved in a ``csv`` file, you will first need to load them into a Pandas DataFrame:
 
 .. code-block:: python
 

[…]
