### 2024-07-17 df9684b
Removing TODO list from front page
