### 2024-07-17 e31c839
Removing double messaging on front page
