+++ docs/source/index.rst
@@ -1,11 +1,6 @@
 MapReader
 =========
 
-**A computer vision pipeline for exploring and analyzing images at scale**
-
-
-------------
-
 .. include:: ../../README.md
    :parser: myst_parser.sphinx_
 
