+++ docs/source/conf.py
@@ -32,7 +32,7 @@ extensions = [
     "sphinx.ext.napoleon",
     "sphinx.ext.todo",
     "sphinx_togglebutton",
-    'sphinxemoji.sphinxemoji',
+    "sphinxemoji.sphinxemoji",
 ]
 
 templates_path = ["_templates"]
@@ -47,7 +47,7 @@ source_suffix = {
 
 autoapi_dirs = ["../../mapreader"]
 autoapi_type = "python"
-autoapi_root = "api"
+autoapi_root = "in-depth-resources/api"
 
 autoapi_options = [
     "members",
