+++ docs/source/User-guide/Classify/Train.rst
@@ -1,704 +1,0 @@
-Train/fine-tune a classifier
-==============================
-
-If you are new to computer vision/ machine learning, `see this tutorial for details on fine-tuning torchvision models <https://pytorch.org/tutorials/beginner/finetuning_torchvision_models_tutorial.html>`__.
-This will help you get to grips with the basic steps needed to train/fine-tune a model.
-
-Load annotations and prepare data
------------------------------------
-
-Load and check annotations
-~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-First, load in your annotations using:
-
-.. code-block:: python
-
-    from mapreader import AnnotationsLoader
-
-    annotated_images = AnnotationsLoader()
-    annotated_images.load(annotations = "./path/to/annotations.csv")
-
-For example, if you have set up your directory as recommended in our :doc:`Input Guidance </Input-guidance>`, and then saved your patches and annotations using the default settings:
-
-.. code-block:: python
-
-    #EXAMPLE
-    annotated_images = AnnotationsLoader()
-    annotated_images.load("./annotations/railspace_#rosie#.csv")
-
-.. admonition:: Advanced usage
-    :class: dropdown
-
-    Other arguments you may want to specify when loading your annotations include:
-
-    - ``delimiter`` - By default, this is set to "," so will assume your ``csv`` file is comma separated. You will need to specify the ``delimiter`` argument if your file is saved in another format.
-    - ``id_col``, ``patch_paths_col``, ``label_col`` - These are used to indicate the column headings for the columns which contain image IDs, patch file paths and labels respectively. By default, these are set to "image_id", "image_path" and "label".
-    - ``append`` - By default, this is ``False`` and so each call to the ``.load()`` method will overwrite existing annotations. If you would like to load a second ``csv`` file into your ``AnnotationsLoader`` instance you will need to set ``append=True``.
-
-To view the data loaded in from your ``csv`` as a dataframe, use:
-
-.. code-block:: python
-
-    annotated_images.annotations
-
-You will note a ``label_index`` column has been added to your dataframe.
-This column contains a numerical reference number for each label, which is needed when training your model.
-
-To see how your labels map to their label indices, call the ``annotated_images.labels_map`` attribute:
-
-.. code-block:: python
-
-    annotated_images.labels_map
-
-.. note:: This ``labels_map`` will be needed later.
-
-To view a sample of your annotated images use the ``show_sample()`` method.
-The ``label_to_show`` argument specifies which label you would like to show.
-
-For example, to show your "railspace" label:
-
-.. code-block:: python
-
-    #EXAMPLE
-    annotated_images.show_sample("railspace")
-
-.. todo:: update this pic
-
-.. image:: ../../figures/show_image_labels_10.png
-    :width: 400px
-
-
-By default, this will show you a sample of 9 images, but this can be changed by specifying ``num_sample``.
-
-When viewing your annotations, you may notice that you have mislabelled one of your images.
-The ``.review_labels()`` method, which returns an interactive tool for adjusting your annotations, provides an easy way to fix this:
-
-.. code-block:: python
-
-    annotated_images.review_labels()
-
-.. image:: ../../figures/review_labels.png
-    :width: 400px
-
-
-.. note:: To exit, type "exit", "end", or "stop" into the text box.
-
-Prepare datasets and dataloaders
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. todo:: "Most neural networks expect the images of a fixed size. Therefore, we will need to write some preprocessing code." Add note about this is why we resize and also comment on square images.
-
-Before using your annotated images to train your model, you will first need to:
-
-.. _ratios:
-
[…]
