+++ docs/source/in-depth-resources/index.rst
@@ -1,0 +1,13 @@
+In-Depth Resources
+==================
+
+..
+    TODO: Write intro to in-depth resources section
+
+.. toctree::
+   :maxdepth: 2
+
+   coding-basics
+   github-basics
+   worked-examples/index
+   api/mapreader/index
