+++ docs/source/community-and-contributions/index.rst
@@ -1,0 +1,13 @@
+Community and contributions
+===========================
+
+..
+    TODO: Write an intro to community and contributions here.
+
+.. toctree::
+   :maxdepth: 2
+
+   joining-the-community
+   events
+   contribution-guide/index
+   code-of-conduct-and-inclusivity

[…]
