+++ docs/source/in-depth-resources/api/mapreader/classify/index.rst
@@ -1,16 +1,0 @@
-mapreader.classify
-==================
-
-.. py:module:: mapreader.classify
-
-
-Submodules
-----------
-
-.. toctree::
-   :maxdepth: 1
-
-   /in-depth-resources/api/mapreader/classify/classifier/index
-   /in-depth-resources/api/mapreader/classify/custom_models/index
-   /in-depth-resources/api/mapreader/classify/datasets/index
-   /in-depth-resources/api/mapreader/classify/load_annotations/index
