+++ docs/source/in-depth-resources/api/mapreader/annotate/utils/index.rst
@@ -1,189 +1,0 @@
-mapreader.annotate.utils
-========================
-
-.. py:module:: mapreader.annotate.utils
-
-
-Functions
----------
-
-.. autoapisummary::
-
-   mapreader.annotate.utils.prepare_data
-   mapreader.annotate.utils.annotation_interface
-   mapreader.annotate.utils.prepare_annotation
-   mapreader.annotate.utils.save_annotation
-
-
-Module Contents
----------------
-
-.. py:function:: prepare_data(df, col_names = None, annotation_set = '001', label_col_name = 'label', redo = False, random_state = 'random', num_samples = 100)
-
-   Prepare data for image annotation by selecting a subset of images from a
-   DataFrame.
-
-   :param df: DataFrame containing the image data to be annotated.
-   :type df: pandas.DataFrame
-   :param col_names: List of column names to include in the output. Default columns are
-                     ``["image_path", "parent_id"]``.
-   :type col_names: list of str, optional
-   :param annotation_set: String specifying the annotation set. Default is ``"001"``.
-   :type annotation_set: str, optional
-   :param label_col_name: Column name containing the label information for each image. Default
-                          is ``"label"``.
-   :type label_col_name: str, optional
-   :param redo: If ``True``, all images will be annotated even if they already have a
-                label. If ``False`` (default), only images without a label will be
-                annotated.
-   :type redo: bool, optional
-   :param random_state: Seed for the random number generator used when selecting images to
-                        annotate. If set to ``"random"`` (default), a random seed will be used.
-   :type random_state: int or str, optional
-   :param num_samples: Maximum number of images to annotate. Default is ``100``.
-   :type num_samples: int, optional
-
-   :returns: A list of lists containing the selected image data, with each sublist
-             containing the specified columns plus the annotation set and a row
-             counter.
-   :rtype: list of list of str/int
-
-
-.. py:function:: annotation_interface(data, list_labels, list_colors = None, annotation_set = '001', method = 'ipyannotate', list_shortcuts = None)
-
-   Create an annotation interface for a list of patches with corresponding
-   labels.
-
-   :param data: List of patches to annotate.
-   :type data: list
-   :param list_labels: List of strings representing the labels for each annotation class.
-   :type list_labels: list
-   :param list_colors: List of strings representing the colors for each annotation class,
-                       by default ``["red", "green", "blue", "green"]``.
-   :type list_colors: list, optional
-   :param annotation_set: String representing the annotation set, specified in the yaml file or
-                          via function argument, by default ``"001"``.
-   :type annotation_set: str, optional
-   :param method: String representing the method for annotation, by default
-                  ``"ipyannotate"``.
-   :type method: Literal["ipyannotate", "pigeonxt"], optional
-   :param list_shortcuts: List of strings representing the keyboard shortcuts for each
-                          annotation class, by default ``None``.
-   :type list_shortcuts: list, optional
-
-   :returns: **annotation** -- The annotation object containing the toolbar, tasks and canvas for the
-             interface.
-   :rtype: Annotation
-
-   :raises SystemExit: If ``method`` parameter is not ``"ipyannotate"`` or ``pigeonxt``.
-
-   .. rubric:: Notes
-
-   This function creates an annotation interface using the ``ipyannotate``
-   library, which is a browser-based tool for annotating data.
-
-
[…]
