+++ docs/source/in-depth-resources/api/mapreader/download/data_structures/index.rst
@@ -1,32 +1,0 @@
-mapreader.download.data_structures
-==================================
-
-.. py:module:: mapreader.download.data_structures
-
-
-Classes
--------
-
-.. autoapisummary::
-
-   mapreader.download.data_structures.Coordinate
-   mapreader.download.data_structures.GridIndex
-   mapreader.download.data_structures.GridBoundingBox
-
-
-Module Contents
----------------
-
-.. py:class:: Coordinate(lat, lon)
-
-.. py:class:: GridIndex(x, y, z)
-
-.. py:class:: GridBoundingBox(cell1, cell2)
-
-   .. py:property:: covered_cells
-
-
-   .. py:property:: x_range
-
-
-   .. py:property:: y_range
