+++ docs/source/in-depth-resources/api/mapreader/utils/index.rst
@@ -1,14 +1,0 @@
-mapreader.utils
-===============
-
-.. py:module:: mapreader.utils
-
-
-Submodules
-----------
-
-.. toctree::
-   :maxdepth: 1
-
-   /in-depth-resources/api/mapreader/utils/compute_and_save_stats/index
-   /in-depth-resources/api/mapreader/utils/slice_parallel/index

[…]
