+++ README.md
@@ -1,3 +1,7 @@
+<style>
+  h1 {text-align: center;}
+</style>
+
 # MapReader
 
 <h2 align="center">A computer vision pipeline for exploring and analyzing images at scale</h2>
@@ -137,7 +141,3 @@ Maps above reproduced with the permission of the National Library of Scotland ht
 <!-- prettier-ignore-end -->
 
 <!-- ALL-CONTRIBUTORS-LIST:END -->
-
-<style>
-  h1 {text-align: center;}
-</style>
