### 2024-07-17 6694369
Dropping indices and tables from front page
