+++ docs/source/index.rst
@@ -14,9 +14,10 @@ MapReader
    in-depth-resources/index
    community-and-contributions/index
 
-Indices and tables
-==================
+..
+   Indices and tables
+   ==================
 
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
+   * :ref:`genindex`
+   * :ref:`modindex`
+   * :ref:`search`
