+++ docs/source/in-depth-resources/api/mapreader/spot_text/index.rst
@@ -1,0 +1,15 @@
+mapreader.spot_text
+===================
+
+.. py:module:: mapreader.spot_text
+
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 1
+
+   /in-depth-resources/api/mapreader/spot_text/deepsolo_runner/index
+   /in-depth-resources/api/mapreader/spot_text/dptext_detr_runner/index
+   /in-depth-resources/api/mapreader/spot_text/runner_base/index
