+++ docs/source/in-depth-resources/api/mapreader/index.rst
@@ -1,0 +1,2230 @@
+mapreader
+=========
+
+.. py:module:: mapreader
+
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 1
+
+   /in-depth-resources/api/mapreader/annotate/index
+   /in-depth-resources/api/mapreader/classify/index
+   /in-depth-resources/api/mapreader/download/index
+   /in-depth-resources/api/mapreader/load/index
+   /in-depth-resources/api/mapreader/process/index
+   /in-depth-resources/api/mapreader/spot_text/index
+   /in-depth-resources/api/mapreader/utils/index
+
+
+Classes
+-------
+
+.. autoapisummary::
+
+   mapreader.MapImages
+   mapreader.SheetDownloader
+   mapreader.Downloader
+   mapreader.AnnotationsLoader
+   mapreader.PatchDataset
+   mapreader.PatchContextDataset
+   mapreader.ClassifierContainer
+   mapreader.Annotator
+
+
+Functions
+---------
+
+.. autoapisummary::
+
+   mapreader.loader
+   mapreader.load_patches
+   mapreader.create_polygon_from_latlons
+   mapreader.create_line_from_latlons
+   mapreader.print_version
+
+
+Package Contents
+----------------
+
+.. py:class:: MapImages(path_images = None, file_ext = False, tree_level = 'parent', parent_path = None, **kwargs)
+
+   Class to manage a collection of image paths and construct image objects.
+
+   :param path_images: Path to the directory containing images (accepts wildcards). By
+                       default, ``False``
+   :type path_images: str or None, optional
+   :param file_ext: The file extension of the image files to be loaded, ignored if file types are specified in ``path_images`` (e.g. with ``"./path/to/dir/*png"``).
+                    By default ``False``.
+   :type file_ext: str or False, optional
+   :param tree_level: Level of the image hierarchy to construct. The value can be
+                      ``"parent"`` (default) and ``"patch"``.
+   :type tree_level: str, optional
+   :param parent_path: Path to parent images (if applicable), by default ``None``.
+   :type parent_path: str, optional
+   :param \*\*kwargs: Additional keyword arguments to be passed to the ``_images_constructor``
+                      method.
+   :type \*\*kwargs: dict, optional
+
+   .. attribute:: path_images
+
+      List of paths to the image files.
+
+      :type: list
+
+   .. attribute:: images
+
+      A dictionary containing the constructed image data. It has two levels
+      of hierarchy, ``"parent"`` and ``"patch"``, depending on the value of
+      the ``tree_level`` parameter.
+
+      :type: dict
+
+
+   .. py:method:: add_metadata(metadata, index_col = 0, delimiter = ',', columns = None, tree_level = 'parent', ignore_mismatch = False)
+
+      Add metadata information to the images dictionary.
+
+      :param metadata: Path to a ``csv`` (or similar), ``xls`` or ``xlsx`` file or a pandas DataFrame that contains the metadata information.
+      :type metadata: str or pandas.DataFrame
+      :param index_col: Column to use as the index when reading the file and converting into a panda.DataFrame.
+                        Accepts column indices or column names.
+                        By default ``0`` (first column).
+
+                        Only used if a file path is provided as the ``metadata`` parameter.
+                        Ignored if ``columns`` parameter is passed.
+      :type index_col: int or str, optional
+      :param delimiter: Delimiter used in the ``csv`` file, by default ``","``.
+
+                        Only used if a ``csv`` file path is provided as
+                        the ``metadata`` parameter.
+      :type delimiter: str, optional
+      :param columns: List of columns indices or names to add to MapImages.
+                      If ``None`` is passed, all columns will be used.
[…]
