### 2024-07-17 808b6ad
Simplifying README somewhat (also to get rid of Sphinx warnings)
