+++ README.md
@@ -1,40 +1,41 @@
-<style>
-  h1 {text-align: center;}
-</style>
-
 # MapReader
 
-<h2 align="center">A computer vision pipeline for exploring and analyzing images at scale</h2>
+> [!IMPORTANT] A computer vision pipeline for exploring and analyzing images at scale
 
-<p align="center">
-    <a href="https://pypi.org/project/mapreader/">
-        <img alt="PyPI" src="https://img.shields.io/pypi/v/MapReader">
-    </a>
-    <a href="https://github.com/Living-with-machines/MapReader/blob/main/LICENSE">
-        <img alt="License" src="https://img.shields.io/badge/License-MIT-yellow.svg">
-    </a>
-    <a href="https://github.com/Living-with-machines/MapReader/actions/workflows/mr_ci.yml/badge.svg">
-        <img alt="Integration Tests badge" src="https://github.com/Living-with-machines/MapReader/actions/workflows/mr_ci.yml/badge.svg">
-    </a>
-    <a href="https://zenodo.org/badge/latestdoi/430661738"><img src="https://zenodo.org/badge/430661738.svg" alt="DOI"></a>
-    <br/>
-    <a href="https://codecov.io/github/Living-with-machines/MapReader" >
-    <img src="https://codecov.io/github/Living-with-machines/MapReader/graph/badge.svg?token=38GQ3O1GB5"/>
-    </a>
-</p>
+<!-- ALL-CONTRIBUTORS-BADGE:START - Do not remove or modify this section -->
+[![All Contributors](https://img.shields.io/badge/all_contributors-8-orange.svg?style=flat-square)](#contributors)
+<!-- ALL-CONTRIBUTORS-BADGE:END -->
 
+![PyPI](https://img.shields.io/pypi/v/MapReader)
+![License](https://img.shields.io/badge/License-MIT-yellow.svg)
+![Integration Tests badge](https://github.com/Living-with-machines/MapReader/actions/workflows/mr_ci.yml/badge.svg)
+![DOI](https://zenodo.org/badge/430661738.svg)
+![CodeCov](https://codecov.io/github/Living-with-machines/MapReader/graph/badge.svg?token=38GQ3O1GB5)
+
+## Table of Contents
+
+- [MapReader](#mapreader)
+  - [Table of Contents](#table-of-contents)
+  - [What is MapReader?](#what-is-mapreader)
+  - [Overview](#overview)
+    - [MapReader pipeline](#mapreader-pipeline)
+  - [Documentation](#documentation)
+  - [What is included in this repo?](#what-is-included-in-this-repo)
+  - [How to cite MapReader](#how-to-cite-mapreader)
+  - [Acknowledgements](#acknowledgements)
+  - [Contributors](#contributors)
+
+---
 
 ## What is MapReader?
 
-<div align="center">
-    <figure>
-    <img src="https://raw.githubusercontent.com/Living-with-machines/MapReader/main/figs/river_banner_8bit.png"
-        alt="Annotated Map with Prediction Outputs"
-        width="70%">
-    </figure>
-</div>
-
 MapReader is an end-to-end computer vision (CV) pipeline for exploring and analyzing images at scale.
+
+<figure align="center">
+  <img src="https://raw.githubusercontent.com/Living-with-machines/MapReader/main/figs/river_banner_8bit.png"
+      alt="Annotated Map with Prediction Outputs"
+      width="70%">
+</figure>
 
 MapReader was developed in the [Living with Machines](https://livingwithmachines.ac.uk/) project to analyze large collections of historical maps but is a _**generalizable**_ computer vision pipeline which can be applied to _**any images**_ in a wide variety of domains.
 
[…]
