### 2024-07-17 caaf4ac
Dropping unnecessary API docs files from repo
