+++ docs/source/in-depth-resources/api/mapreader/download/sheet_downloader/index.rst
@@ -1,395 +1,0 @@
-mapreader.download.sheet_downloader
-===================================
-
-.. py:module:: mapreader.download.sheet_downloader
-
-
-Classes
--------
-
-.. autoapisummary::
-
-   mapreader.download.sheet_downloader.SheetDownloader
-
-
-Module Contents
----------------
-
-.. py:class:: SheetDownloader(metadata_path, download_url)
-
-   A class to download map sheets using metadata.
-
-
-   .. py:method:: get_polygons()
-
-      For each map in metadata, creates a polygon from map geometry and saves to ``features`` dictionary.
-
-
-
-   .. py:method:: get_grid_bb(zoom_level = 14)
-
-      For each map in metadata, creates a grid bounding box from map polygons and saves to ``features`` dictionary.
-
-      :param zoom_level: The zoom level to use when creating the grid bounding box.
-                         Later used when downloading maps, by default 14.
-      :type zoom_level: int, optional
-
-
-
-   .. py:method:: extract_wfs_id_nos()
-
-      For each map in metadata, extracts WFS ID numbers from WFS information and saves to ``features`` dictionary.
-
-
-
-   .. py:method:: extract_published_dates(date_col = None)
-
-      For each map in metadata, extracts publication date and saves to ``features`` dictionary.
-
-      :param date_col: A key or list of keys which map to the metadata field containing the publication date.
-                       Multilayer keys should be passed as a list. e.g.:
-
-                       - "key1" will extract ``self.features[i]["key1"]``
-                       - ["key1","key2"] will search for ``self.features[i]["key1"]["key2"]``
-
-                       If  None, ["properties"]["WFS_TITLE"] will be used as keys. Date will then be extracted by regex searching for "Published: XXX".
-                       By default None.
-      :type date_col: str or list, optional
-
-
-
-   .. py:method:: get_merged_polygon()
-
-      Creates a multipolygon representing all maps in metadata.
-
-
-
-   .. py:method:: get_minmax_latlon()
-
-      Prints minimum and maximum latitudes and longitudes of all maps in metadata.
-
-
-
-   .. py:method:: query_map_sheets_by_wfs_ids(wfs_ids, append = False, print = False)
-
-      Find map sheets by WFS ID numbers.
-
-      :param wfs_ids: The WFS ID numbers of the maps to download.
-      :type wfs_ids: Union[list, int]
-      :param append: Whether to append to current query results list or, if False, start a new list.
-                     By default False
-      :type append: bool, optional
-      :param print: Whether to print query results or not.
-                    By default False
-      :type print: bool, optional
-
-
-
-   .. py:method:: query_map_sheets_by_polygon(polygon, mode = 'within', append = False, print = False)
-
-      Find map sheets which are found within or intersecting with a defined polygon.
-
-      :param polygon: shapely Polygon
-      :type polygon: Polygon
-      :param mode: The mode to use when finding maps.
-                   Options of ``"within"``, which returns all map sheets which are completely within the defined polygon,
-                   and ``"intersects""``, which returns all map sheets which intersect/overlap with the defined polygon.
-                   By default "within".
-      :type mode: str, optional
-      :param append: Whether to append to current query results list or, if False, start a new list.
-                     By default False
-      :type append: bool, optional
-      :param print: Whether to print query results or not.
-                    By default False
-      :type print: bool, optional
-
-      .. rubric:: Notes
-
-      Use ``create_polygon_from_latlons()`` to create polygon.
-
-
-
-   .. py:method:: query_map_sheets_by_coordinates(coords, append = False, print = False)
-
-      Find maps sheets which contain a defined set of coordinates.
[…]
