+++ docs/source/in-depth-resources/api/mapreader/utils/compute_and_save_stats/index.rst
@@ -1,35 +1,0 @@
-mapreader.utils.compute_and_save_stats
-======================================
-
-.. py:module:: mapreader.utils.compute_and_save_stats
-
-
-Attributes
-----------
-
-.. autoapisummary::
-
-   mapreader.utils.compute_and_save_stats.list_dirs
-   mapreader.utils.compute_and_save_stats.myproc
-   mapreader.utils.compute_and_save_stats.list_jobs
-
-
-Functions
----------
-
-.. autoapisummary::
-
-   mapreader.utils.compute_and_save_stats.save_stats
-
-
-Module Contents
----------------
-
-.. py:function:: save_stats(one_dir)
-
-.. py:data:: list_dirs
-
-.. py:data:: myproc
-
-.. py:data:: list_jobs
-   :value: []
