+++ docs/source/in-depth-resources/api/mapreader/spot_text/runner_base/index.rst
@@ -1,137 +1,0 @@
-mapreader.spot_text.runner_base
-===============================
-
-.. py:module:: mapreader.spot_text.runner_base
-
-
-Classes
--------
-
-.. autoapisummary::
-
-   mapreader.spot_text.runner_base.Runner
-
-
-Module Contents
----------------
-
-.. py:class:: Runner
-
-   .. py:method:: run_all(patch_df = None, return_dataframe = False, min_ioa = 0.7)
-
-      Run the model on all images in the patch dataframe.
-
-      :param patch_df: Dataframe containing patch information, by default None.
-      :type patch_df: pd.DataFrame, optional
-      :param return_dataframe: Whether to return the predictions as a pandas DataFrame, by default False
-      :type return_dataframe: bool, optional
-      :param min_ioa: The minimum intersection over area to consider two polygons the same, by default 0.7
-      :type min_ioa: float, optional
-
-      :returns: A dictionary of predictions for each patch image or a DataFrame if `as_dataframe` is True.
-      :rtype: dict or pd.DataFrame
-
-
-
-   .. py:method:: run_on_images(img_paths, return_dataframe = False, min_ioa = 0.7)
-
-      Run the model on a list of images.
-
-      :param img_paths: A list of image paths to run the model on.
-      :type img_paths: str, pathlib.Path or list
-      :param return_dataframe: Whether to return the predictions as a pandas DataFrame, by default False
-      :type return_dataframe: bool, optional
-      :param min_ioa: The minimum intersection over area to consider two polygons the same, by default 0.7
-      :type min_ioa: float, optional
-
-      :returns: A dictionary of predictions for each image or a DataFrame if `as_dataframe` is True.
-      :rtype: dict or pd.DataFrame
-
-
-
-   .. py:method:: run_on_image(img_path, return_outputs=False, return_dataframe = False, min_ioa = 0.7)
-
-      Run the model on a single image.
-
-      :param img_path: The path to the image to run the model on.
-      :type img_path: str or pathlib.Path
-      :param return_outputs: Whether to return the outputs direct from the model, by default False
-      :type return_outputs: bool, optional
-      :param return_dataframe: Whether to return the predictions as a pandas DataFrame, by default False
-      :type return_dataframe: bool, optional
-      :param min_ioa: The minimum intersection over area to consider two polygons the same, by default 0.7
-      :type min_ioa: float, optional
-
-      :returns: The predictions for the image or the outputs from the model if `return_outputs` is True.
-      :rtype: dict or pd.DataFrame
-
-
-
-   .. py:method:: convert_to_parent_pixel_bounds(patch_df = None, return_dataframe = False, deduplicate = False, min_ioa = 0.7)
-
-      Convert the patch predictions to parent predictions by converting pixel bounds.
-
-      :param patch_df: Dataframe containing patch information, by default None
-      :type patch_df: pd.DataFrame, optional
-      :param return_dataframe: Whether to return the predictions as a pandas DataFrame, by default False
-      :type return_dataframe: bool, optional
-      :param deduplicate: Whether to deduplicate the parent predictions, by default False.
-                          Depending on size of parent images, this can be slow.
-      :type deduplicate: bool, optional
-      :param min_ioa: The minimum intersection over area to consider two polygons the same, by default 0.7
-                      This is only used if `deduplicate` is True.
-      :type min_ioa: float, optional
-
-      :returns: A dictionary of predictions for each parent image or a DataFrame if `as_dataframe` is True.
-      :rtype: dict or pd.DataFrame
-
-      :raises ValueError: If `patch_df` is not available.
-
-
-
-   .. py:method:: convert_to_coords(parent_df = None, return_dataframe = False)
-
[…]

[…]
