### 2024-07-17 25fa0f7
Consolidating all figures into one directory
