### 2024-07-17 4d2343a
Moving all images to `_static` directory
