### 2024-07-17 0096b72
Changing README images to `_static` directory
