+++ docs/source/index.rst
@@ -3,6 +3,7 @@ MapReader
 
 .. include:: ../../README.md
    :parser: myst_parser.sphinx_
+   :start-after: <!--- sphinx-include --->
 
 .. toctree::
    :hidden:
