### 2024-07-17 e74f5df
Adding some easier access to other docs, addressing Add missing/edit ambiguous docstrings #463
