+++ mapreader/load/images.py
@@ -160,7 +160,9 @@ class MapImages:
         **kwargs: dict,
     ) -> None:
         """
-        Constructs image data from the given image path and parent path and adds it to the ``MapImages`` instance's ``images`` attribute.
+        Constructs image data from the given image path and parent path and
+        adds it to the :class:`~.load.images.MapImages` instance's ``images``
+        attribute.
 
         Parameters
         ----------
@@ -186,7 +188,11 @@ class MapImages:
 
         Notes
         -----
-        This method assumes that the ``images`` attribute has been initialized on the MapImages instance as a dictionary with two levels of hierarchy, ``"parent"`` and ``"patch"``. The image data is added to the corresponding level based on the value of ``tree_level``.
+        This method assumes that the ``images`` attribute has been initialized
+        on the :class:`~.load.images.MapImages` instance as a dictionary with
+        two levels of hierarchy, ``"parent"`` and ``"patch"``. The image data
+        is added to the corresponding level based on the value of
+        ``tree_level``.
         """
 
         if tree_level not in ["parent", "patch"]:
@@ -256,7 +262,8 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
     @staticmethod
     def _convert_image_path(inp_path: str) -> tuple[str, str, str]:
         """
-        Convert an image path into an absolute path and find basename and directory name.
+        Convert an image path into an absolute path and find basename and
+        directory name.
 
         Parameters
         ----------
@@ -283,7 +290,7 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
         ignore_mismatch: bool | None = False,
     ) -> None:
         """
-        Add metadata information to the images dictionary.
+        Add metadata information to the ``images`` dictionary property.
 
         Parameters
         ----------
@@ -325,10 +332,13 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
 
         Notes
         ------
-        Your metadata file must contain an column which contains the image IDs (filenames) of your images.
-        This should have a column name of either ``name`` or ``image_id``.
+        Your metadata file must contain an column which contains the image IDs
+        (filenames) of your images. This should have a column name of either
+        ``name`` or ``image_id``.
 
-        Existing information in your ``MapImages`` object will be overwritten if there are overlapping column headings in your metadata file/dataframe.
+        Existing information in your :class:`~.load.images.MapImages` object
+        will be overwritten if there are overlapping column headings in your
+        metadata file/dataframe.
         """
 
         if isinstance(metadata, pd.DataFrame):
@@ -508,7 +518,7 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
 
         Notes
         -----
-        The method runs :meth:`mapreader.load.images.MapImages._add_shape_id`
+        The method runs :meth:`~.load.images.MapImages._add_shape_id`
         for each image present at the ``tree_level`` provided.
         """
         print(f"[INFO] Add shape, tree level: {tree_level}")
@@ -547,7 +557,7 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
         Notes
         -----
         The method runs
-        :meth:`mapreader.load.images.MapImages._add_coord_increments_id`
+        :meth:`~.load.images.MapImages._add_coord_increments_id`
         for each image present at the parent level, which calculates
         pixel-wise delta longitude (``dlon``) and delta latitude (``dlat``)
         for the image and adds the data to it.
[…]
