### 2024-07-17 82efd76
Adding more edits that addresses #463
