+++ mapreader/download/sheet_downloader.py
@@ -646,7 +646,9 @@ class SheetDownloader:
 
             e.g. ``{"county": ["properties", "COUNTY"], "id": "id"}``
         **kwargs: dict, optional
-            Keyword arguments to pass to the ``extract_published_dates()`` method.
+            Keyword arguments to pass to the
+            :meth:`~.download.sheet_downloader.SheetDownloader.extract_published_dates`
+            method.
 
         Returns
         -------
@@ -655,8 +657,9 @@ class SheetDownloader:
 
         Notes
         -----
-        Default metadata items are: name, url, coordinates, crs, published_date, grid_bb.
-        Additional items can be added using ``metadata_to_save``.
+        Default metadata items are: ``name``, ``url``, ``coordinates``,
+        ``crs``, ``published_date``, ``grid_bb``. Additional items can be
+        added using the ``metadata_to_save`` argument.
         """
         metadata_cols = [
             "name",
@@ -739,7 +742,9 @@ class SheetDownloader:
         download_in_parallel : bool, optional
             Whether to download tiles in parallel, by default ``True``.
         **kwargs : dict, optional
-            Keyword arguments to pass to the ``_save_metadata()`` method.
+            Keyword arguments to pass to the
+            :meth:`~.download.sheet_downloader.SheetDownloader._save_metadata`
+            method.
         """
 
         for feature in tqdm(features):
@@ -786,7 +791,9 @@ class SheetDownloader:
         download_in_parallel : bool, optional
             Whether to download tiles in parallel, by default ``True``.
         **kwargs : dict, optional
-            Keyword arguments to pass to the ``_download_map_sheets()`` method.
+            Keyword arguments to pass to the
+            :meth:`~.download.sheet_downloader.SheetDownloader._download_map_sheets`
+            method.
         """
         if not self.grid_bbs:
             raise ValueError("[ERROR] Please first run ``get_grid_bb()``")
@@ -829,7 +836,9 @@ class SheetDownloader:
         download_in_parallel : bool, optional
             Whether to download tiles in parallel, by default ``True``.
         **kwargs : dict, optional
-            Keyword arguments to pass to the ``_download_map_sheets()`` method.
+            Keyword arguments to pass to the
+            :meth:`~.download.sheet_downloader.SheetDownloader._download_map_sheets`
+            method.
         """
 
         if not self.wfs_id_nos:
@@ -898,7 +907,9 @@ class SheetDownloader:
         download_in_parallel : bool, optional
             Whether to download tiles in parallel, by default ``True``.
         **kwargs : dict, optional
-            Keyword arguments to pass to the ``_download_map_sheets()`` method.
+            Keyword arguments to pass to the
+            :meth:`~.download.sheet_downloader.SheetDownloader._download_map_sheets`
+            method.
 
         Notes
         -----
@@ -976,7 +987,9 @@ class SheetDownloader:
         download_in_parallel : bool, optional
             Whether to download tiles in parallel, by default ``True``.
         **kwargs : dict, optional
-            Keyword arguments to pass to the ``_download_map_sheets()`` method.
+            Keyword arguments to pass to the
+            :meth:`~.download.sheet_downloader.SheetDownloader._download_map_sheets`
+            method.
         """
 
         if not isinstance(coords, tuple):
@@ -1036,7 +1049,9 @@ class SheetDownloader:
         download_in_parallel : bool, optional
             Whether to download tiles in parallel, by default ``True``.
         **kwargs : dict, optional
-            Keyword arguments to pass to the ``_download_map_sheets()`` method.
+            Keyword arguments to pass to the
+            :meth:`~.download.sheet_downloader.SheetDownloader._download_map_sheets`
+            method.
 
         Notes
[…]
