+++ mapreader/load/images.py
@@ -56,8 +56,8 @@ class MapImages:
     parent_path : str, optional
         Path to parent images (if applicable), by default ``None``.
     **kwargs : dict, optional
-        Additional keyword arguments to be passed to the ``_images_constructor``
-        method.
+        Keyword arguments to pass to the
+        :meth:`~.load.images.MapImages._images_constructor` method.
 
     Attributes
     ----------
