### 2024-07-24 #6434 @emdupre
Thank you, @rwood-97 !

On the archive, it looks like there's one missing ORCID for Rosie Wood (that is present in the software paper). Could you please add this ? You can usually do so by [editing the archive metadata](https://help.zenodo.org/docs/deposit/manage-records/#edit). If you hit any issues, of course, please let me know.

Otherwise, to move things along, I'll go ahead and modify my existing PR to incorporate the last reference edit.
