### 2024-07-25 acc1a49
Replacing remaining relevant print statements with logger.info/logger.warning
