+++ mapreader/annotate/annotator.py
@@ -31,6 +31,7 @@ _CENTER_LAYOUT = widgets.Layout(
 # Set up logging
 logger = logging.getLogger(__name__)
 
+
 class Annotator:
     """
     Annotator class for annotating patches with labels.
@@ -334,9 +335,9 @@ class Annotator:
             A tuple containing the parent dataframe and patch dataframe.
         """
         if patch_paths:
-            print(f"[INFO] Loading patches from {patch_paths}.")
+            logger.info(f"Loading patches from {patch_paths}.")
         if parent_paths:
-            print(f"[INFO] Loading parents from {parent_paths}.")
+            logger.info(f"Loading parents from {parent_paths}.")
 
         maps = load_patches(patch_paths=patch_paths, parent_paths=parent_paths)
         # Add pixel stats
@@ -344,7 +345,7 @@ class Annotator:
 
         try:
             maps.add_metadata(metadata_path, delimiter=delimiter)
-            print(f"[INFO] Adding metadata from {metadata_path}.")
+            logger.info(f"Adding metadata from {metadata_path}.")
         except ValueError:
             raise FileNotFoundError(
                 f"[INFO] Metadata file at {metadata_path} not found. Please specify the correct file path using the ``metadata_path`` argument."
@@ -749,7 +750,7 @@ class Annotator:
         self._queue = self.get_queue()
 
         if self._filter_for is not None:
-            print(f"[INFO] Filtering for: {self._filter_for}")
+            logger.info(f"Filtering for: {self._filter_for}")
 
         self.out = widgets.Output(layout=_CENTER_LAYOUT)
         display(self.box)
