### 2024-07-26 afea1b4
Merge main and fix conflicts around coding basics
