+++ docs/source/in-depth-resources/coding-basics.rst
@@ -1,52 +1,0 @@
-Coding Basics
-==============
-
-Using the terminal
--------------------
-
-A terminal is a command-line interface where you can type in commands and interact with your computer's operating system.
-It can be a powerful tool for programmers, allowing them to execute complex tasks with just a few keystrokes.
-However, it can also be intimidating for beginners who are not familiar with using it.
-
-Here are some resources to help you get started with using the terminal on your computer:
-
-- `Introduction to the command line <https://curriculum.dhinstitutes.org/workshops/command-line/>`__
-- `Introduction to the bash command line  <https://programminghistorian.org/en/lessons/intro-to-bash>`__
-- `Introduction to the Windows command line with PowerShell <https://programminghistorian.org/en/lessons/intro-to-powershell>`__
-- `The Unix shell <https://swcarpentry.github.io/shell-novice/>`__
-- more??
-
-Python packages
-----------------
-i.e. not a program/app that you can click and find in your start page/ launch pad.
-
-- Usually no GUI
-- ??
-
-Conda and Python virtual environments
---------------------------------------
-
-A virtual environment is a way to create an isolated environment in which you can install specific versions of packages and dependencies for your Python projects.
-This is useful because it allows you to avoid conflicts between different projects that may require different versions of the same package.
-In the installation instructions provided, there are two methods for setting up a virtual environment for MapReader: using Anaconda (also known as conda) or using venv, which is Python's native way of handling virtual environments.
-
-If you're new to virtual environments in Python, this tutorial provides a good introduction:
-
-- `Getting started with Python environments (using conda) <https://towardsdatascience.com/getting-started-with-python-environments-using-conda-32e9f2779307>`__
-- `Why you need Python environments and how to manage them with conda <https://www.freecodecamp.org/news/why-you-need-python-environments-and-how-to-manage-them-with-conda-85f155f4353c/>`__
-- `Virtual environments and packages <https://docs.python.org/3/tutorial/venv.html>`__
-- more??
-
-Jupyter notebooks
-------------------
-
-A Jupyter notebook is an interactive computational environment that allows you to write and run code, visualize data, and write narrative text all in the same place.
-It's a popular tool among data scientists and is commonly used for data analysis, machine learning, and scientific computing.
-
-If you're new to Jupyter notebooks, here's a great place to start:
-
-- `Introduction to Jupyter Notebooks <https://programminghistorian.org/en/lessons/jupyter-notebooks>`__
-- `How to use Jupyter Notebooks: A beginner's tutorial <https://www.dataquest.io/blog/jupyter-notebook-tutorial/>`__
-- more??
-
-You will notice that many of our "worked examples" are built in Jupyter notebooks so you can test run some of this code as soon as you have understood what the notebooks are, how to set them up on your computer, and how to run code in them.
