### 2024-07-26 8b4bfca
Fixing more print statements and testing side of things
