+++ mapreader/load/geo_utils.py
@@ -38,7 +38,8 @@ def extractGeoInfo(image_path):
         tiff_proj = tiff_src.crs.to_string()
         tiff_coord = tuple(tiff_src.bounds)
 
-    logger.info(f"Shape: {tiff_shape}. \n[INFO] CRS: {tiff_proj}.")
+    logger.info(f"Shape: {tiff_shape}.")
+    logger.info(f"CRS: {tiff_proj}.")
     logger.info("Coordinates: {:.4f} {:.4f} {:.4f} {:.4f}".format(*tiff_coord))
 
     return tiff_shape, tiff_proj, tiff_coord
@@ -96,8 +97,8 @@ def reproject_geo_info(image_path, target_crs="EPSG:4326", calc_size_in_m=False)
         mean_pixel_height = np.mean([right / height, left / height])
         mean_pixel_width = np.mean([bottom / width, top / width])
 
-        print(
-            f"[INFO] Size in meters of left/bottom/right/top: {left:.2f}/{bottom:.2f}/{right:.2f}/{top:.2f}"
+        logger.info(
+            f"Size in meters of left/bottom/right/top: {left:.2f}/{bottom:.2f}/{right:.2f}/{top:.2f}"
         )
         print(
             f"Each pixel is ~{mean_pixel_height:.3f} X {mean_pixel_width:.3f} meters (height x width)."
