+++ mapreader/load/images.py
@@ -200,8 +200,8 @@ class MapImages:
 
         if tree_level == "parent":
             if parent_path:
-                print(
-                    "[WARNING] Ignoring `parent_path` as `tree_level`  is set to 'parent'."
+                logger.warning(
+                    "Ignoring `parent_path` as `tree_level`  is set to 'parent'."
                 )
                 parent_path = None
             parent_id = None
@@ -382,8 +382,8 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
         if "name" in columns:
             image_id_col = "name"
             if "image_id" in columns:
-                print(
-                    "[WARNING] Both 'name' and 'image_id' columns exist! Using 'name' as index"  # noqa
+                logger.warning(
+                    "Both 'name' and 'image_id' columns exist! Using 'name' as index"  # noqa
                 )
         elif "image_id" in columns:
             image_id_col = "image_id"
@@ -393,8 +393,8 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
             )
 
         if any(metadata_df.duplicated(subset=image_id_col)):
-            print(
-                "[WARNING] Duplicates found in metadata. Keeping only first instance of each duplicated value"
+            logger.warning(
+                "Duplicates found in metadata. Keeping only first instance of each duplicated value"
             )
             metadata_df.drop_duplicates(subset=image_id_col, inplace=True, keep="first")
 
@@ -528,8 +528,8 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
 
         for parent_id in parent_list:
             if "grid_bb" not in self.parents[parent_id].keys():
-                print(
-                    f"[WARNING] No grid bounding box found for {parent_id}. Suggestion: run add_metadata or add_geo_info."  # noqa
+                logger.warning(
+                    f"No grid bounding box found for {parent_id}. Suggestion: run add_metadata or add_geo_info."  # noqa
                 )
                 continue
 
@@ -562,8 +562,8 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
 
         for parent_id in parent_list:
             if "coordinates" not in self.parents[parent_id].keys():
-                print(
-                    f"[WARNING] No coordinates found for {parent_id}. Suggestion: run add_metadata or add_geo_info."  # noqa
+                logger.warning(
+                    f"No coordinates found for {parent_id}. Suggestion: run add_metadata or add_geo_info."  # noqa
                 )
                 continue
 
@@ -633,8 +633,8 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
         for image_id in image_ids:
             if tree_level == "parent":
                 if "coordinates" not in self.parents[image_id].keys():
-                    print(
-                        f"[WARNING] 'coordinates' could not be found in {image_id}. Suggestion: run add_metadata or add_geo_info"  # noqa
+                    logger.warning(
+                        f"'coordinates' could not be found in {image_id}. Suggestion: run add_metadata or add_geo_info"  # noqa
                     )
                     continue
 
@@ -643,8 +643,8 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
 
                 if "coordinates" not in self.parents[parent_id].keys():
                     if parent_id not in already_checked_parent_ids:
-                        print(
-                            f"[WARNING] 'coordinates' could not be found in {parent_id} so center coordinates cannot be calculated for it's patches. Suggestion: run add_metadata or add_geo_info."  # noqa
+                        logger.warning(
[…]
