+++ mapreader/load/images.py
@@ -762,8 +762,10 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
         """
 
         if "coordinates" not in self.parents[image_id].keys():
-            logger.warning(
-                f"'coordinates' could not be found in {image_id}. Suggestion: run add_metadata or add_geo_info."
+            self._log_if_verbose(
+                f"'coordinates' could not be found in {image_id}. Suggestion: run add_metadata or add_geo_info.",
+                verbose,
+                type="warning",
             )
             return
 
@@ -800,8 +802,10 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
         parent_id = self.patches[image_id]["parent_id"]
 
         if "coordinates" not in self.parents[parent_id].keys():
-            logger.warning(
-                f"No coordinates found in  {parent_id} (parent of {image_id}). Suggestion: run add_metadata or add_geo_info."
+            self._log_if_verbose(
+                f"No coordinates found in  {parent_id} (parent of {image_id}). Suggestion: run add_metadata or add_geo_info.",
+                verbose,
+                type="warning",
             )
             return
 
@@ -875,8 +879,10 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
 
         if "coordinates" not in self.images[tree_level][image_id].keys():
             if tree_level == "parent":
-                logger.warning(
-                    f"No coordinates found for {image_id}. Suggestion: run add_metadata or add_geo_info."
+                self._log_if_verbose(
+                    f"No coordinates found for {image_id}. Suggestion: run add_metadata or add_geo_info.",
+                    verbose,
+                    type="warning",
                 )
                 return
 
@@ -884,7 +890,9 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
                 self._add_patch_coords_id(image_id, verbose)
 
         if "coordinates" in self.images[tree_level][image_id].keys():
-            logger.info(f"Reading 'coordinates' from {image_id}.")
+            self._log_if_verbose(
+                f"Reading 'coordinates' from {image_id}.", verbose, type="info"
+            )
 
             min_x, min_y, max_x, max_y = self.images[tree_level][image_id][
                 "coordinates"
@@ -966,11 +974,15 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
         mean_pixel_height = np.mean([right / height, left / height])
         mean_pixel_width = np.mean([bottom / width, top / width])
 
-        logger.info(
-            f"Size in meters of left/bottom/right/top: {left:.2f}/{bottom:.2f}/{right:.2f}/{top:.2f}"
+        self._log_if_verbose(
+            f"Size in meters of left/bottom/right/top: {left:.2f}/{bottom:.2f}/{right:.2f}/{top:.2f}",
+            verbose,
+            type="info",
         )
-        logger.info(
-            f"Each pixel is ~{mean_pixel_height:.3f} X {mean_pixel_width:.3f} meters (height x width)."
+        self._log_if_verbose(
+            f"Each pixel is ~{mean_pixel_height:.3f} X {mean_pixel_width:.3f} meters (height x width).",
+            verbose,
+            type="info",
         )  # noqa
 
         return size_in_m, mean_pixel_height, mean_pixel_width
@@ -1039,15 +1051,15 @@ See https://pillow.readthedocs.io/en/stable/handbook/concepts.html#modes for mor
             image_path = self.images[tree_level][image_id]["image_path"]
 
             try:
-                full_path = print(
+                full_path = logger.info(
                     os.path.relpath(image_path)
                 )  # TODO: This looks like it won't work
             except ValueError:  # if no rel path (e.g. mounted on different drives)
[…]
