+++ mapreader/classify/classifier.py
@@ -178,7 +178,7 @@ class ClassifierContainer:
             self.dataloaders = dataloaders if dataloaders else {}
 
         for set_name, dataloader in self.dataloaders.items():
-            print(f'[INFO] Loaded "{set_name}" with {len(dataloader.dataset)} items.')
+            logger.info(f'Loaded "{set_name}" with {len(dataloader.dataset)} items.')
 
     def generate_layerwise_lrs(
         self,
