### 2024-07-26 4409edb
Adding a logger to base class Runner
