+++ tests/test_geo_pipeline.py
@@ -24,7 +24,7 @@ def test_pipeline(tmp_path, sample_dir):
     print(f"{sample_dir}/test_json.json")
     my_ts = SheetDownloader(
         metadata_path=f"{sample_dir}/test_json.json",
-        download_url="https://geo.nls.uk/maps/os/1inch_2nd_ed/{z}/{x}/{y}.png",
+        download_url="https://mapseries-tilesets.s3.amazonaws.com/1inch_2nd_ed/{z}/{x}/{y}.png",
     )
 
     my_ts.extract_wfs_id_nos()
