### 2024-07-30 46ed60e
add occlusion analysis code
