+++ mapreader/process/occlusion_analysis.py
@@ -1,0 +1,301 @@
+#!/usr/bin/env python
+from __future__ import annotations
+
+import os
+from ast import literal_eval
+
+import cv2
+import numpy as np
+import pandas as pd
+import torch
+from PIL import Image
+from torch import nn
+from torchvision import transforms
+from tqdm import tqdm
+
+
+class OcclusionAnalyzer:
+    """A class for carrying out occlusion analysis on patches.
+
+    Parameters
+    ----------
+    patch_df : pd.DataFrame
+        The DataFrame containing patches and predictions.
+    model : str or nn.Module
+            The PyTorch model to add to the object.
+            If a string, this should be the path to a model checkpoint.
+    transform : str or callable
+        The transform to apply to the patches. Options of "default" or a torchvision transform.
+        Default transform
+    delimiter : str
+        The delimiter used in the patch_df csv file. By default, ",".
+    device : str
+        The device to use. By default, "default".
+    """
+
+    def __init__(
+        self,
+        patch_df: pd.DataFrame | str,
+        model: str | nn.Module,
+        transform: str | callable = "default",
+        delimiter: str = ",",
+        device: str = "default",
+    ):
+        if isinstance(patch_df, pd.DataFrame):
+            self.patch_df = patch_df
+
+        elif isinstance(patch_df, str):
+            if os.path.isfile(patch_df):
+                print(f'[INFO] Reading "{patch_df}".')
+                patch_df = pd.read_csv(patch_df, sep=delimiter, index_col=0)
+                # ensure tuple/list columns are read as such
+                patch_df = self._eval_df(patch_df)
+                self.patch_df = patch_df
+            else:
+                raise ValueError(f'[ERROR] "{patch_df}" cannot be found.')
+
+        else:
+            raise ValueError(
+                "[ERROR] Please pass ``patch_df`` as a string (path to csv file) or pd.DataFrame."
+            )
+
+        if patch_df.index.name != "image_id" and "image_id" in patch_df.columns:
+            patch_df.set_index("image_id", drop=True, inplace=True)
+
+        if any(col not in patch_df.columns for col in ["predicted_label", "pred"]):
+            raise ValueError(
+                "[ERROR] The patch dataframe should contain predicted labels."
+            )
+
+        if device == "default":
+            self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+        else:
+            self.device = torch.device(device)
+
+        if isinstance(model, str):
+            self.model = self._load_model(model)
+        elif isinstance(model, nn.Module):
+            self.model = model
+        else:
+            raise ValueError("[ERROR] Please pass ``model`` as a string or nn.Module.")
+        # set model to evaluation mode
+        self.model.eval()
+
+        if transform == "default":
+            mean = [0.485, 0.456, 0.406]
+            std = [0.229, 0.224, 0.225]
+            self.transform = transforms.Compose(
+                [
+                    transforms.Resize((224, 224)),
+                    transforms.ToTensor(),
+                    transforms.Normalize(mean, std),
+                ]
+            )
+        else:
+            self.transform = transform
+
+    def __len__(self):
+        return len(self.patch_df)
+
+    @staticmethod
+    def _eval_df(df):
+        for col in df.columns:
+            try:
+                df[col] = df[col].apply(literal_eval)
+            except (ValueError, TypeError, SyntaxError):
+                pass
+        return df
+
+    def _load_model(self, model_path: str) -> nn.Module:
+        model = torch.load(model_path, device=self.device)
+        return model
+
+    def add_criterion(
+        self, criterion: str | nn.modules.loss._Loss | None = "cross entropy"
+    ) -> None:
+        """
+        Add a loss criterion to the object.
+
+        Parameters
+        ----------
[…]
