### 2024-07-31 88c8b01
add maptext runner
