+++ mapreader/spot_text/maptext_runner.py
@@ -1,0 +1,491 @@
+from __future__ import annotations
+
+import pathlib
+import pickle
+
+try:
+    import adet
+except ImportError:
+    raise ImportError(
+        "[ERROR] Please install DeepSolo from the following link: https://github.com/rwood-97/DeepSolo"
+    )
+
+import numpy as np
+import pandas as pd
+from adet.config import get_cfg
+
+try:
+    from detectron2.engine import DefaultPredictor
+except ImportError:
+    raise ImportError("[ERROR] Please install Detectron2")
+
+from adet.utils.vitae_predictor import ViTAEPredictor
+from shapely import LineString, MultiPolygon, Polygon
+
+# first assert we are using the deep solo version of adet
+if adet.__version__ != "0.2.0-maptextpipeline":
+    raise ImportError(
+        "[ERROR] Please install MapTextPipeline from the following link: https://github.com/rwood-97/MapTextPipeline"
+    )
+
+from .runner_base import Runner
+
+
+class MapTextRunner(Runner):
+    def __init__(
+        self,
+        patch_df: pd.DataFrame = None,
+        parent_df: pd.DataFrame = None,
+        cfg_file: str
+        | pathlib.Path = "./MapTextPipeline/configs/ViTAEv2_S/rumsey/final_rumsey.yaml",
+        weights_file: str | pathlib.Path = "./rumsey-finetune.pth",
+        device: str = "cpu",
+    ) -> None:
+        # setup the dataframes
+        self.patch_df = patch_df
+        self.parent_df = parent_df
+
+        # set up predictions as dictionaries
+        self.patch_predictions = {}
+        self.parent_predictions = {}
+        self.geo_predictions = {}
+
+        # setup the config
+        cfg = get_cfg()  # get a fresh new config
+        cfg.merge_from_file(cfg_file)
+        cfg.MODEL.WEIGHTS = weights_file
+        cfg.MODEL.DEVICE = device
+
+        self.voc_size = cfg.MODEL.TRANSFORMER.VOC_SIZE
+        self.use_customer_dictionary = cfg.MODEL.TRANSFORMER.CUSTOM_DICT
+        if self.voc_size == 96:
+            self.CTLABELS = [
+                " ",
+                "!",
+                '"',
+                "#",
+                "$",
+                "%",
+                "&",
+                "'",
+                "(",
+                ")",
+                "*",
+                "+",
+                ",",
+                "-",
+                ".",
+                "/",
+                "0",
+                "1",
+                "2",
+                "3",
+                "4",
+                "5",
+                "6",
+                "7",
+                "8",
+                "9",
+                ":",
+                ";",
+                "<",
+                "=",
+                ">",
+                "?",
+                "@",
+                "A",
+                "B",
+                "C",
+                "D",
+                "E",
+                "F",
+                "G",
+                "H",
+                "I",
+                "J",
+                "K",
+                "L",
+                "M",
+                "N",
+                "O",
+                "P",
+                "Q",
+                "R",
+                "S",
+                "T",
+                "U",
+                "V",
+                "W",
+                "X",
+                "Y",
+                "Z",
+                "[",
+                "\\",
+                "]",
+                "^",
+                "_",
+                "`",
+                "a",
+                "b",
+                "c",
+                "d",
+                "e",
+                "f",
+                "g",
+                "h",
+                "i",
+                "j",
+                "k",
+                "l",
+                "m",
+                "n",
+                "o",
+                "p",
+                "q",
+                "r",
+                "s",
+                "t",
+                "u",
+                "v",
+                "w",
+                "x",
+                "y",
+                "z",
+                "{",
+                "|",
+                "}",
+                "~",
+            ]
+        elif self.voc_size == 37:
+            self.CTLABELS = [
+                "a",
+                "b",
+                "c",
[…]
