+++ setup.py
@@ -59,6 +59,7 @@ setuptools.setup(
         "pyogrio>=0.7.2",
         "cartopy>=0.22.0",
         "joblib>=1.4.0",
+        "opencv-python<5.0.0.0"
     ],
     extras_require={
         "dev": [
