+++ tests/test_post_processing/test_context_post_processing.py
@@ -10,7 +10,7 @@ from mapreader.process.context_post_process import ContextPostProcessor
 
 @pytest.fixture
 def sample_dir():
-    return Path(__file__).resolve().parent / "sample_files"
+    return Path(__file__).resolve().parent.parent / "sample_files"
 
 
 @pytest.fixture
@@ -114,6 +114,7 @@ def test_update_preds_railspace_railspace_building(patch_df, labels_map):
             "patch-1-7-2-8-#cropped_74488689.png#.png", "new_predicted_label"
         ]
         == "no"
+        
     )
     assert (
         patches.patch_df.loc[
