+++ tests/test_post_processing/test_occlusion.py
@@ -1,14 +1,14 @@
 from __future__ import annotations
 
 from pathlib import Path
+from typing import Callable
 
 import pandas as pd
 import pytest
 import timm
 import torch
-from torchvision import transforms, models
-from typing import Callable
 from PIL import Image
+from torchvision import models, transforms
 
 from mapreader.process.occlusion_analysis import OcclusionAnalyzer
 
@@ -22,15 +22,20 @@ def sample_dir():
 def patch_df(sample_dir):
     return pd.read_csv(f"{sample_dir}/post_processing_patch_df.csv", index_col=0)
 
+
 @pytest.fixture
 def model():
-    return timm.create_model("hf_hub:Livingwithmachines/mr_resnest101e_finetuned_OS_6inch_2nd_ed", pretrained=True)
+    return timm.create_model(
+        "hf_hub:Livingwithmachines/mr_resnest101e_finetuned_OS_6inch_2nd_ed",
+        pretrained=True,
+    )
 
 
 def test_init_dataframe(patch_df, model):
     analyzer = OcclusionAnalyzer(patch_df, model)
     assert isinstance(analyzer, OcclusionAnalyzer)
     assert len(analyzer) == 81
+
 
 def test_init_path(sample_dir, model):
     patch_df = f"{sample_dir}/post_processing_patch_df.csv"
@@ -39,30 +44,41 @@ def test_init_path(sample_dir, model):
     assert len(analyzer) == 81
     assert isinstance(analyzer.patch_df.iloc[0]["pixel_bounds"], tuple)
 
+
 def test_init_dataframe_transform(patch_df, model):
     transform = transforms.ToTensor()
-    analyzer = OcclusionAnalyzer(patch_df, model, transform = transform)
+    analyzer = OcclusionAnalyzer(patch_df, model, transform=transform)
     assert isinstance(analyzer, OcclusionAnalyzer)
     assert isinstance(analyzer.transform, Callable)
-    img = Image.new("RGB", (10,10))
+    img = Image.new("RGB", (10, 10))
     assert isinstance(analyzer.transform(img), torch.Tensor)
+
 
 def test_init_fake_path_error(model):
     patch_df = "fake_df.csv"
     with pytest.raises(ValueError, match="cannot be found"):
         OcclusionAnalyzer(patch_df, model)
 
+
 def test_init_error(model):
-    with pytest.raises(ValueError, match="as a string (path to csv file) or pd.DataFrame"):
-        OcclusionAnalyzer({"image_id":"patch"}, model)
+    with pytest.raises(
+        ValueError, match="as a string (path to csv file) or pd.DataFrame"
+    ):
+        OcclusionAnalyzer({"image_id": "patch"}, model)
+
 
 def test_init_dataframe_no_predictions(patch_df, model):
     patch_df.drop(columns="predicted_label", inplace=True)
-    with pytest.raises(ValueError, match="patch dataframe should contain predicted labels"):
+    with pytest.raises(
+        ValueError, match="patch dataframe should contain predicted labels"
+    ):
         OcclusionAnalyzer(patch_df, model)
     patch_df.drop(columns="pred", inplace=True)
-    with pytest.raises(ValueError, match="patch dataframe should contain predicted labels"):
+    with pytest.raises(
+        ValueError, match="patch dataframe should contain predicted labels"
+    ):
         OcclusionAnalyzer(patch_df, model)
+
 
 def test_init_reindex_dataframe(patch_df, model):
     patch_df.reset_index(inplace=True, drop=False)
@@ -110,5 +126,3 @@ def test_run_occlusion(patch_df, model):
     out = analyzer.run_occlusion("railspace", 2)
     assert len(out) == 2
     assert isinstance(out[0], Image)
-
-
