### 2024-07-31 dd87486
add more tests, make sure criterion is set when running occlusion
