### 2024-08-01 65b4ab6
change criterion to loss_fn, add docs
