+++ tests/test_post_processing/test_occlusion.py
@@ -101,33 +101,33 @@ def test_init_models_string(sample_dir, patch_df):
     assert isinstance(analyzer.model, torch.nn.Module)
 
 
-def test_add_criterion(patch_df, model):
+def test_add_loss_fn(patch_df, model):
     analyzer = OcclusionAnalyzer(patch_df, model)
-    for criterion, criterion_type in [
+    for loss_fn, loss_fn_type in [
         ("cross_entropy", torch.nn.CrossEntropyLoss),
         ("bce", torch.nn.BCELoss),
         ("mse", torch.nn.MSELoss),
     ]:
-        analyzer.add_criterion(criterion)  # loss function as str
-        assert isinstance(analyzer.criterion, criterion_type)
-    my_criterion = torch.nn.L1Loss()
-    analyzer.add_criterion(my_criterion)
-    assert isinstance(analyzer.criterion, torch.nn.L1Loss)
+        analyzer.add_loss_fn(loss_fn)  # loss function as str
+        assert isinstance(analyzer.loss_fn, loss_fn_type)
+    loss_fn = torch.nn.L1Loss()
+    analyzer.add_loss_fn(loss_fn)
+    assert isinstance(analyzer.loss_fn, torch.nn.L1Loss)
 
 
-def test_criterion_errors(patch_df, model):
+def test_loss_fn_errors(patch_df, model):
     analyzer = OcclusionAnalyzer(patch_df, model)
-    with pytest.raises(NotImplementedError, match="criterion can only be"):
-        analyzer.add_criterion("a fake criterion")
+    with pytest.raises(NotImplementedError, match="loss function can only be"):
+        analyzer.add_loss_fn("a fake loss_fn")
     with pytest.raises(ValueError, match="Please pass"):
-        analyzer.add_criterion(0.01)
+        analyzer.add_loss_fn(0.01)
 
 
 def test_run_occlusion(sample_dir, patch_df, model):
     img_path = f"{sample_dir}/patch-0-3045-145-3190-#map_100942121.png#.png"
     patch_df["image_path"] = img_path
     analyzer = OcclusionAnalyzer(patch_df, model)
-    analyzer.add_criterion()
+    analyzer.add_loss_fn()
     out = analyzer.run_occlusion("railspace", 2)
     assert len(out) == 2
     assert isinstance(out[0], Image.Image)
@@ -138,7 +138,7 @@ def test_run_occlusion_small_sample(sample_dir, patch_df, model):
     patch_df["image_path"] = img_path
     patch_df.loc[patch_df.iloc[0].name, "predicted_label"] = "fake_label"
     analyzer = OcclusionAnalyzer(patch_df, model)
-    analyzer.add_criterion()
+    analyzer.add_loss_fn()
     out = analyzer.run_occlusion("fake_label", 10)
     assert len(out) == 1
     assert isinstance(out[0], Image.Image)
@@ -148,7 +148,7 @@ def test_run_occlusion_save(sample_dir, patch_df, model, tmp_path):
     img_path = f"{sample_dir}/patch-0-3045-145-3190-#map_100942121.png#.png"
     patch_df["image_path"] = img_path
     analyzer = OcclusionAnalyzer(patch_df, model)
-    analyzer.add_criterion()
+    analyzer.add_loss_fn()
     analyzer.run_occlusion(
         "railspace", 2, save=True, path_save=f"{tmp_path}/occlusion/"
     )
@@ -156,7 +156,7 @@ def test_run_occlusion_save(sample_dir, patch_df, model, tmp_path):
     assert len(os.listdir(f"{tmp_path}/occlusion/")) == 2
 
 
-def test_run_occlusion_no_criterion_error(patch_df, model):
[…]
