+++ broken_files.txt
@@ -1,81 +1,0 @@
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-0-0-1-1-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-0-1-1-2-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-0-2-1-3-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-0-3-1-4-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-0-4-1-5-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-0-5-1-6-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-0-6-1-7-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-0-7-1-8-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-0-8-1-9-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-1-0-2-1-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-1-1-2-2-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-1-2-2-3-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-1-3-2-4-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-1-4-2-5-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-1-5-2-6-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-1-6-2-7-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-1-7-2-8-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-1-8-2-9-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-2-0-3-1-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-2-1-3-2-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-2-2-3-3-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-2-3-3-4-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-2-4-3-5-#cropped_74488689.png#.png
-/Users/rwood/LwM/mapreader/MapReader/tests/patches_1_pixel/patch-2-5-3-6-#cropped_74488689.png#.png
[…]
