### 2024-08-02 3a7d959
Resolving conflicts and fixing test errors
