### 2024-08-02 21d0495
Committing the old contribution guide for easier moving forward
