+++ docs/source/Contribution-guide/Contribution-guide.rst
@@ -1,0 +1,189 @@
+Contribution Guide
+===================
+
+Welcome! We are pleased to know that you're interested in contributing to MapReader!
+
+MapReader is a collaborative project, now expanding its community beyond the initial group in the `Living with Machines <https://livingwithmachines.ac.uk/>`_ project (The Alan Turing Institute).
+We welcome all contributions but **please** follow these guidelines to make sure your contributions can be easily integrated into the project.
+
+.. contents:: Table of Contents
+        :local:
+
+Joining the community
+----------------------
+
+Before you begin contributing to MapReader, we would love for you to join our community and there are many ways to do this:
+
+- Star the `MapReader GitHub repository <https://github.com/Living-with-machines/MapReader>`_.
+- Start creating Github issues and/or pull requests on our repository.
+- Join our `Slack workspace <https://forms.gle/dXjECHZQkwrZ3Xpt9>`_.
+
+You can also get in touch with the MapReader team personally by:
+
+- Tagging us on Github (find our Github handles `here <https://github.com/Living-with-machines/MapReader/blob/main/ways_of_working.md>`__).
+- Messaging us on Slack.
+- Emailing Katie McDonough at k.mcdonough@lancaster.ac.uk.
+
+Pre-requisites
+--------------
+
+Regardless of how you will be contributing to MapReader, you will need to:
+
+1. Ensure you have a `GitHub account <https://docs.github.com/en/get-started/signing-up-for-github/signing-up-for-a-new-github-account>`_ set up.
+2. Be able to use GitHub issues and pull requests - if you are unfamiliar with these, please look at our :doc:`GitHub guide </Contribution-guide/GitHub-guide>` before continuing.
+3. Set up a virtual python environment and install MapReader (as per our :doc:`Installation instructions </Install>`).
+4. Have read this guide.
+
+Ways to contribute
+-------------------
+
+We welcome contributions from community members of all skill levels and so have written three different guides for different types of contributions:
+
+- :doc:`Write a tutorial or worked example </Contribution-guide/Worked-examples>` - For those who are familiar with using MapReader and would like to showcase an example of how to use it, but are not comfortable with using Sphinx to write documentation or writing code.
+- :doc:`Add to or update the MapReader documentation </Contribution-guide/Documentation>` - For those who are familiar with using MapReader and either already comfortable with using Sphinx to write documentation or feel able to have a go at to learning it.
+- :doc:`Add to or update the MapReader code </Contribution-guide/Code>` - For those who are familiar with using MapReader and either already comfortable with writing code in Python or feel able to have a go at learning it.
+
+How to add yourself as a contributor
+------------------------------------
+
+We use the `all-contributors <https://allcontributors.org/>`__ bot to keep a record of contributors to the MapReader repo.
+To add yourself as a contributor to MapReader, comment on your issue or PR with the following:
+
+``@all-contributors please add @your_username_here for A, B, C``
+
+.. note:: Remember to change this to your GitHub username!
+
+for example, to add Katie McDonough for research, ideas and docs, you would write:
+
+``@all-contributors please add @kmcdono2 for research, ideas, docs``
+
+Types of contribution
+~~~~~~~~~~~~~~~~~~~~~
+
+.. list-table::
+    :widths: 30 30 40
+    :header-rows: 1
+
+    * - Emoji/Type
+      - Represents
+      - Comments
+    * - 💻 ``code``
+      - Code
+      -
+    * - 🔣 ``data``
+      - Data
+      -
+    * - 📖 ``doc``
+      - Documentation
+      - e.g. ReadtheDocs, Wiki, or other source of documentation
+    * - 🤔 ``ideas``
[…]
