### 2024-08-02 affadeb
enable loading dataframes from path
