+++ docs/source/User-guide/Spot-text.rst
@@ -1,19 +1,20 @@
 Spot text
 =========
 
-MapReader now contains two new methods of spotting text on maps:
+MapReader implements three new frameworks for spotting text on maps:
 
 - ``DPTextDETRRunner`` - This is used to detect text on maps using `DPTextDETR <https://github.com/ymy-k/DPText-DETR/tree/main>`__ and outputs bounding boxes and scores.
-- ``DeepSoloRunner`` - This is used to detect and recognise text on maps using `DeepSolo <https://github.com/ViTAE-Transformer/DeepSolo/tree/main>`__ and outputs bounding boxes, text and scores.
+- ``DeepSoloRunner`` - This is used to detect and recognize text on maps using `DeepSolo <https://github.com/ViTAE-Transformer/DeepSolo/tree/main>`__ and outputs bounding boxes, text and scores.
+- ``MapTextPipeline`` - This is used to detect and recognize text on maps using `MapTextPipeline <https://github.com/yyyyyxie/MapTextPipeline>`__ and outputs bounding boxes, text and scores.
+
+We recommend using the ``MapTextPipeline`` for most use cases as it has been used to train a model on a sample of David Rumsey maps and so should work best for map text spotting.
 
 Install dependencies
 --------------------
 
-To run these, you will need to install the required dependencies.
+To run text spotting with MapReader, you will need to install the required dependencies. These are listed below:
 
-.. note::
-
-    We have our own forks of these repos to enable them to work on CPU!
+.. note:: We have our own forks of the ``DPTextDETRRunner``, ``DeepSolo`` and ``MapTextPipeline`` repos to enable them to work on CPU. Please use our forks but remember to cite the original authors in your work!
 
 Detectron2
 ~~~~~~~~~~~
@@ -30,10 +31,6 @@ To install, run the following commands in your terminal:
     pip install .
 
 
-.. note::
-
-    Since both the DPText-DETR and DeepSolo repos are built ontop of `AdelaiDet <https://github.com/aim-uofa/AdelaiDet>`__, you won't be able to install both at the same. To get around this, you can set up two different conda environments, one for each.
-
 .. admonition:: Detectron2 issues for windows users
     :class: dropdown
 
@@ -43,8 +40,9 @@ To install, run the following commands in your terminal:
 
     Once this is done, retry installing Detectron2.
 
-You should then pick one of the following to install:
+You should then **pick one** of the following to install:
 
+.. note:: Since the DPText-DETR, DeepSolo and MapTextPipeline frameworks are built on top of `AdelaiDet <https://github.com/aim-uofa/AdelaiDet>`__, you won't be able to install them at the same. To get around this, you can set up three different conda environments, one for each framework. This will allow you to switch between them as needed.
 
 DPTextDETR
 ~~~~~~~~~~~
@@ -72,18 +70,31 @@ To install, run the following commands in your terminal:
     cd DeepSolo
     pip install .
 
+MapTextPipeline
+~~~~~~~~~~~~~~~
+
+Our fork for MapTextPipeline is available `here <https://github.com/rwood-97/MapTextPipeline>`__
+
+To install, run the following commands in your terminal:
+
+.. code:: bash
+
+    git clone https://github.com/rwood-97/MapTextPipeline.git
+    cd MapTextPipeline
+    pip install .
+
 
 Advice for patch size
 ---------------------
 
 When running the text spotting models, we recommend using a patch size of 1024x1024 pixels.
-This is the size used in the training of the models, and so should give the best results.
+This is the size used as input to the models, and so should give the best results.
 
[…]
