### 2024-08-02 12b774a
Replace most `assert` statements with more descriptive`raise` statements
