### 2024-08-02 fed74f9
add border option when annotating
