+++ mapreader/annotate/annotator.py
@@ -55,6 +55,8 @@ class Annotator:
         Name of the column in which labels are stored in patch DataFrame, by default "label"
     show_context : bool, optional
         Whether to show context when loading patches, by default False
+    border : bool, optional
+        Whether to add a border around the central patch when showing context, by default False
     auto_save : bool, optional
         Whether to automatically save annotations, by default True
     delimiter : str, optional
@@ -108,6 +110,7 @@ class Annotator:
         patch_paths_col: str = "image_path",
         label_col: str = "label",
         show_context: bool = False,
+        border: bool = False,
         auto_save: bool = True,
         delimiter: str = ",",
         sortby: str | None = None,
@@ -255,6 +258,7 @@ class Annotator:
         self.patch_paths_col = patch_paths_col
         self.annotations_file = annotations_file
         self.show_context = show_context
+        self.border = border
         self.auto_save = auto_save
         self.username = username
         self.task_name = task_name
@@ -452,7 +456,7 @@ class Annotator:
 
         self.navbox = widgets.VBox([widgets.HBox([prev_btn, next_btn])])
 
-    def get_queue(
+    def _get_queue(
         self, as_type: str | None = "list"
     ) -> list[int] | (pd.Index | pd.Series):
         """
@@ -520,7 +524,7 @@ class Annotator:
             return indices
         return queue_df
 
-    def get_context(self):
+    def _get_context(self):
         """
         Provides the surrounding context for the patch to be annotated.
 
@@ -531,15 +535,21 @@ class Annotator:
             context.
         """
 
-        def get_path(image_path, dim=True):
+        def get_square(image_path, dim=True, border=False):
             # Resize the image
             im = Image.open(image_path)
 
             # Dim the image
-            if dim in [True, "True"]:
+            if dim in ["True", True]:
                 im_array = np.array(im)
                 im_array = 256 - (256 - im_array) * 0.4  # lighten image
                 im = Image.fromarray(im_array.astype(np.uint8))
+
+            if border in ["True", True] and self.border:
+                w, h = im.size
+                im = ImageOps.expand(im, border=2, fill="red")
+                im = im.resize((w, h))
+
             return im
 
         def get_empty_square(patch_size: tuple[int, int]):
@@ -593,7 +603,10 @@ class Annotator:
 
         # derive ids from items
         ids = [x.index[0] if len(x.index) == 1 else None for x in items]
-        ids = [x != ix for x in ids]
+        # list of booleans, True if not the current patch, False if the current patch
+        # used for dimming the surrounding patches and adding a border to the current patch
+        dim_bools = [x != ix for x in ids]
+        border_bools = [x == ix for x in ids]
 
         # derive images from items
         image_paths = [
@@ -601,14 +614,16 @@ class Annotator:
         ]
 
         # zip them
-        image_list = list(zip(image_paths, ids))
+        image_list = list(zip(image_paths, dim_bools, border_bools))
 
         # split them into rows
         per_row = len(deltas)
         images = [
             [
-                get_path(x[0], dim=x[1]) if x[0] else get_empty_square((width, height))
-                for x in lst
+                get_square(image_path, dim=dim, border=border)
+                if image_path
+                else get_empty_square((width, height))
+                for image_path, dim, border in lst
             ]
             for lst in array_split(image_list, per_row)
         ]
@@ -641,6 +656,7 @@ class Annotator:
     def annotate(
         self,
         show_context: bool | None = None,
+        border: bool | None = None,
         sortby: str | None = None,
         ascending: bool | None = None,
         min_values: dict | None = None,
@@ -656,6 +672,9 @@ class Annotator:
         ----------
         show_context : bool or None, optional
             Whether or not to display the surrounding context for each image.
+            Default is None.
[…]
