### 2024-08-02 9bb9354
add ability to show patch info when annotating
