### 2024-08-05 08e00c9
add docs for `border`
