### 2024-08-05 aa2047f
add docs for `show_vals`
