### 2024-08-05 c3b1b47
General improvement of property methods Fixes #477
