+++ mapreader/download/sheet_downloader.py
@@ -19,7 +19,10 @@ from shapely.geometry import LineString, Point, Polygon, shape
 from shapely.ops import unary_union
 from tqdm.auto import tqdm
 
-from .downloader_utils import get_grid_bb_from_polygon, get_polygon_from_grid_bb
+from .downloader_utils import (
+    get_grid_bb_from_polygon,
+    get_polygon_from_grid_bb,
+)
 from .tile_loading import DEFAULT_TEMP_FOLDER, TileDownloader
 from .tile_merging import TileMerger
 
@@ -46,39 +49,51 @@ class SheetDownloader:
             should contain placeholders for the x coordinate (``x``), the y
             coordinate (``y``) and the zoom level (``z``).
         """
-        self.polygons = False
-        self.grid_bbs = False
-        self.wfs_id_nos = False
         self.published_dates = False
         self.found_queries = []
-        self.merged_polygon = None
+        self._merged_polygon = None
+        self._grid_bbs = None
+        self._polygons = None
 
         assert isinstance(
             metadata_path, str
         ), "[ERROR] Please pass metadata_path as string."
 
-        if os.path.isfile(metadata_path):
-            with open(metadata_path) as f:
-                self.metadata = json.load(f)
-                self.features = self.metadata["features"]
-                print(self.__str__())
-
-        else:
-            raise ValueError("[ERROR] Metadata file not found.")
-
-        if isinstance(download_url, str):
-            my_ts = [download_url]
-        elif isinstance(download_url, list):
-            my_ts = download_url
-        else:
+        if not isinstance(download_url, (str, list)):
             raise ValueError(
                 "[ERROR] Please pass ``download_url`` as string or list of strings."
             )
 
-        self.tile_server = my_ts
+        if not os.path.isfile(metadata_path):
+            raise ValueError("[ERROR] Metadata file not found.")
 
-        crs_string = CRS(self.metadata["crs"]["properties"]["name"])
-        self.crs = crs_string.to_string()
+        # Set up SheetDownloader.metadata + SheetDownloader.features
+        with open(metadata_path) as f:
+            self.metadata = json.load(f)
+            self.features = self.metadata["features"]
+            print(self.__str__())
+
+        # Set up SheetDownloader.tile_server
+        if isinstance(download_url, str):
+            download_url = [download_url]
+
+        self.tile_server = download_url
+
+        # Set up SheetDownloader.crs
+        self.crs = CRS(self.metadata["crs"]["properties"]["name"]).to_string()
+        self._test_crs()
+
+        # Set up SheetDownloader.polygons
+        _ = self.polygons
+
+        # Set up SheetDownloader.grid_bbs
+        _ = self.grid_bbs
+
+        # Set up SheetDownloader.merge_polygon
+        _ = self.merged_polygon
+
+        # Set up SheetDownloader.wfs_id_nos
+        _ = self.wfs_id_nos
 
     def __str__(self) -> str:
         info = f"[INFO] Metadata file has {self.__len__()} item(s)."
@@ -87,18 +102,37 @@ class SheetDownloader:
     def __len__(self) -> int:
         return len(self.features)
 
-    def get_polygons(self) -> None:
+    @property
+    def polygons(self) -> None:
         """
         For each map in metadata, creates a polygon from map geometry and saves to ``features`` dictionary.
         """
-        for feature in self.features:
-            polygon = shape(feature["geometry"])
-            map_name = feature["properties"]["IMAGE"]
-            if len(polygon.geoms) != 1:
-                f"[WARNING] Multiple geometries found in map {map_name}. Using first instance."
-            feature["polygon"] = polygon.geoms[0]
+        if not self._polygons:
[…]
