### 2024-08-05 510246f
Update reference, remove suggested phrasing
