+++ paper/paper.bib
@@ -102,11 +102,8 @@
 	pages = {1800},
 }
 
-
-@article{Corcoran,
+@unpublished{Corcoran,
 	title = {Automated extraction of dynamic phenotype data from whole plant images collected under controlled conditions},
-	journal = {Frontiers in Plant Science},
 	author = {Corcoran, E and Hosseini, K and Siles, L and Kurup, S and Ahnert, S},
-	year = {2023},
-	pages = {In preparation},
+	note = {Under review},
 }
