+++ tests/test_annotator.py
@@ -38,6 +38,138 @@ def test_init_with_dfs(load_dfs):
     assert isinstance(annotator.patch_df.iloc[0]["coordinates"], tuple)
 
 
+def test_init_defaults(load_dfs):
+    parent_df, patch_df, tmp_path = load_dfs
+    # test defaults
+    annotator = Annotator(
+        patch_df=patch_df,
+        parent_df=parent_df,
+        labels=["a", "b"],
+    )
+    assert len(annotator) == 9
+    assert annotator._labels == ["a", "b"]
+    assert annotator.show_context is False
+    assert annotator.border is False
+    assert annotator.auto_save is True
+    assert annotator._sortby is None
+    assert annotator._ascending is True
+    assert annotator.task_name == "task"
+    assert annotator._min_values == {}
+    assert annotator._max_values == {}
+    assert annotator._filter_for is None
+    assert annotator.surrounding == 1
+    assert annotator.max_size == 1000
+    assert annotator.resize_to is None
+
+    annotator.annotate()
+    assert annotator.show_context is False
+    assert annotator.border is False
+    assert annotator._sortby is None
+    assert annotator._ascending is True
+    assert annotator._min_values == {}
+    assert annotator._max_values == {}
+    assert annotator.surrounding == 1
+    assert annotator.max_size == 1000
+    assert annotator.resize_to is None
+    assert annotator.show_vals is None
+
+    annotator.annotate(
+        show_context=True,
+        border=True,
+        sortby="min_x",
+        ascending=False,
+        min_values={"min_x": 0},
+        max_values={"min_x": 10},
+        surrounding=2,
+        resize_to=300,
+        max_size=500,
+        show_vals=["label"],
+    )
+    assert annotator.show_context is True
+    assert annotator.border is True
+    assert annotator._sortby == "min_x"
+    assert annotator._ascending is False
+    assert annotator._min_values == {"min_x": 0}
+    assert annotator._max_values == {"min_x": 10}
+    assert annotator.surrounding == 2
+    assert annotator.max_size == 500
+    assert annotator.resize_to == 300
+    assert annotator.show_vals == ["label"]
+
+
+def test_init_optional_args(load_dfs):
+    parent_df, patch_df, tmp_path = load_dfs
+    # add optional arguments
+    annotator = Annotator(
+        patch_df=patch_df,
+        parent_df=parent_df,
+        labels=["a", "b"],
+        annotations_dir=f"{tmp_path}/annotations/",
+        show_context=True,
+        border=True,
+        auto_save=False,
+        sortby="min_x",
+        ascending=False,
+        username="rosie",
+        task_name="test",
+        surrounding=2,
+        max_size=500,
+        resize_to=300,
+    )
+    assert len(annotator) == 9
+    assert annotator._labels == ["a", "b"]
+    assert annotator.show_context is True
+    assert annotator.border is True
+    assert annotator.auto_save is False
+    assert annotator._sortby == "min_x"
+    assert annotator._ascending is False
+    assert annotator.username == "rosie"
+    assert annotator.task_name == "test"
+    assert annotator.surrounding == 2
+    assert annotator.max_size == 500
+    assert annotator.resize_to == 300
+
+    annotator.annotate()
+    assert annotator.show_context is True
+    assert annotator.border is True
+    assert annotator._sortby == "min_x"
+    assert annotator._ascending is False
[…]
