### 2024-08-06 #6434 @emdupre
Thank you, @rwood-97 !

It looks like there is one identified DOI that is not currently included in the bib file entry. From my review, it looks like this is indeed associated with the correct reference and should be added. Could you please include this ?

Otherwise, please let me know if you see any remaining issues with the current article proofs before we proceed !
