### 2024-08-06 48a5b24
Update paper.bib
