+++ paper/paper.bib
@@ -64,21 +64,20 @@
 }
 
 @inproceedings{mapkurator,
-author = {Li, Zekun and Chiang, Yao-Yi and Tavakkol, Sasan and Shbita, Basel and Uhl, Johannes H. and Leyk, Stefan and Knoblock, Craig A.},
-title = {An Automatic Approach for Generating Rich, Linked Geo-Metadata from Historical Map Images},
-year = {2020},
-isbn = {9781450379984},
-publisher = {Association for Computing Machinery},
-address = {New York, NY, USA},
-url = {https://doi.org/10.1145/3394486.3403381},
-doi = {10.1145/3394486.3403381},
-abstract = {Historical maps contain detailed geographic information difficult to find elsewhere covering long-periods of time (e.g., 125 years for the historical topographic maps in the US). However, these maps typically exist as scanned images without searchable metadata. Existing approaches making historical maps searchable rely on tedious manual work (including crowd-sourcing) to generate the metadata (e.g., geolocations and keywords). Optical character recognition (OCR) software could alleviate the required manual work, but the recognition results are individual words instead of location phrases (e.g., "Black'' and "Mountain'' vs. "Black Mountain''). This paper presents an end-to-end approach to address the real-world problem of finding and indexing historical map images. This approach automatically processes historical map images to extract their text content and generates a set of metadata that is linked to large external geospatial knowledge bases. The linked metadata in the RDF (Resource Description Framework) format support complex queries for finding and indexing historical maps, such as retrieving all historical maps covering mountain peaks higher than 1,000 meters in California. We have implemented the approach in a system called mapKurator. We have evaluated mapKurator using historical maps from several sources with various map styles, scales, and coverage. Our results show significant improvement over the state-of-the-art methods. The code has been made publicly available as modules of the Kartta Labs project at https://github.com/kartta-labs/Project.},
-booktitle = {Proceedings of the 26th ACM SIGKDD International Conference on Knowledge Discovery \& Data Mining},
-pages = {3290–3298},
-numpages = {9},
-keywords = {entity matching, geolocalization, historical map processing, information extraction, neural networks, text linking},
-location = {Virtual Event, CA, USA},
-series = {KDD '20}
+	author = {Li, Zekun and Chiang, Yao-Yi and Tavakkol, Sasan and Shbita, Basel and Uhl, Johannes H. and Leyk, Stefan and Knoblock, Craig A.},
+	title = {An Automatic Approach for Generating Rich, Linked Geo-Metadata from Historical Map Images},
+	year = {2020},
+	isbn = {9781450379984},
+	publisher = {Association for Computing Machinery},
+	address = {New York, NY, USA},
+	url = {https://doi.org/10.1145/3394486.3403381},
+	doi = {10.1145/3394486.3403381},
+	booktitle = {Proceedings of the 26th ACM SIGKDD International Conference on Knowledge Discovery \& Data Mining},
+	pages = {3290–3298},
+	numpages = {9},
+	keywords = {entity matching, geolocalization, historical map processing, information extraction, neural networks, text linking},
+	location = {Virtual Event, CA, USA},
+	series = {KDD '20}
 }
 
 @article{Petitpierre,
