### 2024-08-07 db8e022
Update unit test badge with version generated from github actions
