### 2024-08-07 b412400
Add python 3.12 to build matrix
