### 2024-08-07 34a8d79
Update codeql workflow with latest version from github
