### 2024-08-07 f6c0c12
Remove old version of codeql analysis workflow
