### 2024-08-07 1782d1d
Renaming unrecognised test functions for download.data_structures
