### 2024-08-07 6670442
Adding output to `_check_z` function
