### 2024-08-07 cd3a9f2
Move bdist wheel universal config to pyproject.toml
