+++ setup.cfg
@@ -1,4 +1,0 @@
-[bdist_wheel]
-universal = 1
-[metadata]
-description_file = README.md
