+++ pyproject.toml
@@ -2,6 +2,9 @@
 requires = ["setuptools", "setuptools-scm"]
 build-backend = "setuptools.build_meta"
 
+[tool.distutils.bdist_wheel]
+universal = true
+
 [project]
 name = "piffle"
 version = "0.6.0-dev"
