### 2024-08-07 5282ff2
Switch build to hatchling
