### 2024-08-07 26cffd3
Set dev version to 0.6.0-dev
