+++ piffle/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.5.1"
+__version__ = "0.6.0.dev0"
