+++ paper/paper.bib
@@ -109,14 +109,11 @@
 	pages = {1800},
 }
 
-@article{Corcoran,
-	author={Corcoran, Evangeline  and Siles, Laura  and Kurup, Smita  and Ahnert, Sebastian },
-	title={Automated extraction of pod phenotype data from micro-computed tomography},
-	journal={Frontiers in Plant Science},
-	volume={14},
-	year={2023},
-	url={https://www.frontiersin.org/journals/plant-science/articles/10.3389/fpls.2023.1120182},
-	doi={10.3389/fpls.2023.1120182},
-	issn={1664-462x},
+@unpublished{Corcoran,
+	author = {Corcoran, Evangeline  and Hosseini, Kasra and Siles, Laura  and Kurup, Smita  and Ahnert, Sebastian },
+	title = {Dynamic phenotyping of whole oilseed rape (Brassica napus) plants from images collected under controlled conditions},
+	journal = {Frontiers in Plant Science},
+	year = {2024},
+	note = {Under review},
 }
 
