+++ tests/test_sheet_downloader.py
@@ -42,7 +42,7 @@ def test_init(sheet_downloader):
 
 def test_init_errors(sample_dir):
     test_json = f"{sample_dir}/test_json.json"  # crs changed to EPSG:3857 (note: coordinates are wrong in file)
-        download_url = (
+    download_url = (
         "https://mapseries-tilesets.s3.amazonaws.com/1inch_2nd_ed/{z}/{x}/{y}.png"
     )
     with pytest.raises(ValueError, match="file not found"):
