+++ CHANGELOG.md
@@ -88,7 +88,114 @@ c65ff77 move common functions to base class
 ## [v1.3.2](https://github.com/Living-with-machines/MapReader/releases/tag/v1.3.2) (2024-05-17)
 
 <!--
-PRs: #434, #433, #431, #432, #418, #415, #414, #416, #396, #413
+PRs:
+
+    Pull request 434
+    ### Summary
+
+    Adds calculation of average across all channels for mean/std pixel stats.
+    Also makes sure image cropping is working for calc pixel stats.
+    Fixes #426
+
+    -------------------
+
+    Pull request 433
+    ### Summary
+
+    Fixes #430
+
+    -------------------
+
+    Pull request 431
+    ### Summary
+
+    Fixes #64
+
+    ### Describe your changes
+
+    Command line call now prints mapreader version
+
+    -------------------
+
+    Pull request 432
+    ### Summary
+
+    Fixes #400
+
+    ### Describe your changes
+
+    Previous code allowed you enter a new label when reviewing labels but would give you an error (even though it relabelled the patch fine). Now we give an error and only allow user to enter one of the existing labels when re-labelling/reviewing their annotations.
+
+    -------------------
+
+    Pull request 418
+    Addresses #386
+    -------------------
+
+    Pull request 415
+    ### Summary
+
+    Fixes non-geospatial README issues in #<386>
+
+    ### Describe your changes
+
+    Reformatted to standard fields per worked example, matching geospatial README.
+
+    -------------------
+
+    Pull request 414
+    ### Summary
+
+    Addresses "update geospatial README" in https://github.com/orgs/Living-with-machines/projects/7/views/1?pane=issue&itemId=59167565
+
+    Fixes part of  #<386>
+
+    ### Describe your changes
+
+    Changed README from content that is now mostly in documentation site to clear structure for list of worked examples and required files.
+
+    -------------------
+
+    Pull request 416
+    ### Summary
+
+    Fixes #389
+
+    ### Describe your changes
+
+    - updated all typos mentioned in https://github.com/openjournals/joss-reviews/issues/6434#issuecomment-2041626766
+
+    -------------------
+
+    Pull request 396
+    ### Summary
+
+    Fixes #385
+
+    ### Describe your changes
+
+    Update code and dependencies to ensure fix warnings in tests
+
+    To do:
+    - [x] Update docs re. how to run tests
+    - [x] Deal with annotator
+    - [ ] Zero division error ?
+
+    -------------------
+
+    Pull request 413
+    ### Summary
+
+    Fixes #398
+
+    ### Describe your changes
+
+    e.g. https://maps.nls.uk/view/101437802 goes over the boundary and so is registered twice in the NLS metadata but with two sets of coordinates.
+    Instead of erroring or overwriting the sheet, you will now get map_xxxx.png and map_xxxx_1.png saved and these will have separate metadata so separate coordinates.
+
+    This also adds a mock function when downloading maps and so fixes #183
+
+    -------------------
 
 282b873 Merge pull request #434 from Living-with-machines/426-calc-pixel-stats
 0a69fae update pixel stats calculation
@@ -145,7 +252,12 @@ fe370ee Merge pull request #413 from Living-with-machines/issue-398
 ## [v1.3.1](https://github.com/Living-with-machines/MapReader/releases/tag/v1.3.1) (2024-05-03)
 
 <!--
-PRs: #399
+PRs:
+    Pull request 399
+    ### Summary
+
+    Adding notebooks for the April/May workshop.
+
 
 3d7ab10 update tests
 dd216eb fix overwriting issue
[…]
