### 2024-08-09 #6434 @rwood-97
@emdupre We are considering moving the MapReader repo into its own MapReader org (as opposed to being under Living-with-machines). Would your preference be for us to do this before or after publicationo of the JOSS paper?
