### 2024-08-09 #6434 @emdupre
Hi @rwood-97 -- thank you for these updates. There are two remaining issues I see :

1. Given that you've updated the documentation significantly, it would make sense to also generate a new version of the software that includes these updates (and can be referenced in the paper). If you would like to do that at the same time as transferring the repo under a new org, that makes sense to me. Otherwise, the archive itself is the persistent record of the software.

2. On the software paper itself, it looks like the Cocoran reference was modified in the last pull request from the text [I had provided](https://github.com/openjournals/joss-reviews/issues/6434#issuecomment-2225873848). As this is currently an unpublished reference, we should include it as :

```
@unpublished{Corcoran,
	title = {Automated extraction of dynamic phenotype data from whole plant images collected under controlled conditions},
	author = {Corcoran, E and Hosseini, K and Siles, L and Kurup, S and Ahnert, S},
	note = {Under review},
}
```

rather than listing the current year or journal where it is currently under review. From the most recent changes it looks like the article title has since been updated as well !
