### 2024-08-12 #6434 @rwood-97
1. Okay thanks. I think the consensus among the team is that we want to take some time to make sure we do the transfer to the new org properly and so will do this after the JOSS paper. Should I still update the zenodo record with recent updates or would it be better to create a new record?
2. Ah sorry! I thought this could be updated to include the extra info but I will change it back. The title was changed just before submission so I will keep that as the new one.
