+++ tests/test_load/test_images.py
@@ -6,7 +6,6 @@ from random import randint
 
 import geopandas as geopd
 import pandas as pd
-import PIL
 import pytest
 from PIL import Image
 from pytest import approx
@@ -134,6 +133,7 @@ def test_init_png(sample_dir, image_id):
     assert isinstance(maps, MapImages)
     str(maps)
     len(maps)
+    assert not maps.georeferenced
 
 
 def test_init_png_grayscale(sample_dir):
@@ -151,6 +151,7 @@ def test_init_tiff(sample_dir):
     tiffs = MapImages(f"{sample_dir}/{image_id}")
     assert len(tiffs) == 1
     assert isinstance(tiffs, MapImages)
+    assert not tiffs.georeferenced
 
 
 def test_init_geotiff(sample_dir):
@@ -158,6 +159,7 @@ def test_init_geotiff(sample_dir):
     geotiffs = MapImages(f"{sample_dir}/{image_id}")
     assert len(geotiffs) == 1
     assert isinstance(geotiffs, MapImages)
+    assert geotiffs.georeferenced
 
 
 def test_init_parent_path(sample_dir, image_id, capfd):
@@ -187,7 +189,7 @@ def test_init_tiff_32bit_error(sample_dir):
 
 def test_init_non_image_error(sample_dir):
     file_name = "ts_downloaded_maps.csv"
-    with pytest.raises(PIL.UnidentifiedImageError, match="not an image"):
+    with pytest.raises(ValueError, match="Non-image file types detected"):
         MapImages(f"{sample_dir}/{file_name}")
 
 
@@ -217,6 +219,7 @@ def test_add_metadata(sample_dir, image_id, ts_metadata_keys):
     for maps in [maps_csv, maps_tsv, maps_xlsx]:
         assert all([k in maps.parents[image_id].keys() for k in ts_metadata_keys])
         assert isinstance(maps.parents[image_id]["coordinates"], tuple)
+        assert maps.georeferenced
         assert maps.parents[image_id]["coordinates"] == approx(
             (-4.83, 55.80, -4.21, 56.059), rel=1e-2
         )
@@ -302,21 +305,21 @@ def test_add_metadata_index_col(matching_metadata_dir):
 def test_add_metadata_columns(matching_metadata_dir, metadata_df):
     my_files_csv = MapImages(f"{matching_metadata_dir}/*png")
     my_files_csv.add_metadata(
-        f"{matching_metadata_dir}/metadata_df.csv", columns=["name", "coord"]
+        f"{matching_metadata_dir}/metadata_df.csv", usecols=["name", "coord"]
     )
 
     my_files_xlsx = MapImages(f"{matching_metadata_dir}/*png")
     my_files_xlsx.add_metadata(
-        f"{matching_metadata_dir}/metadata_df.xlsx", columns=["name", "coord"]
+        f"{matching_metadata_dir}/metadata_df.xlsx", usecols=["name", "coord"]
     )
 
     my_files_df = MapImages(f"{matching_metadata_dir}/*png")
-    my_files_df.add_metadata(metadata_df, columns=["name", "coord"])  # pass columns arg
+    my_files_df.add_metadata(metadata_df, usecols=["name", "coord"])  # pass columns arg
 
-    metadata_keys = ["parent_id", "image_path", "shape", "name", "coord"]
+    metadata_keys = set(["parent_id", "image_path", "shape", "name", "coord"])
     for my_files in [my_files_csv, my_files_xlsx, my_files_df]:
         for parent_id in my_files.list_parents():
-            assert list(my_files.parents[parent_id].keys()) == metadata_keys
+            assert set(my_files.parents[parent_id].keys()) == metadata_keys
[…]
