+++ mapreader/load/images.py
@@ -6,6 +6,7 @@ except ImportError:
     pass
 
 import os
+import pathlib
 import random
 import re
 import warnings
@@ -22,12 +23,16 @@ import rasterio
 from PIL import Image, ImageOps, ImageStat
 from pyproj import Transformer
 from rasterio.plot import reshape_as_raster
-from shapely import wkt
 from shapely.geometry import Polygon, box
 from tqdm.auto import tqdm
 
 from mapreader.download.data_structures import GridBoundingBox, GridIndex
 from mapreader.download.downloader_utils import get_polygon_from_grid_bb
+from mapreader.utils.load_frames import (
+    load_from_csv,
+    load_from_excel,
+    load_from_geojson,
+)
 
 os.environ[
     "USE_PYGEOS"
@@ -45,15 +50,15 @@ class MapImages:
     Parameters
     ----------
     path_images : str or None, optional
-        Path to the directory containing images (accepts wildcards). By
-        default, ``False``
-    file_ext : str or False, optional
+        Path to the directory containing images (accepts wildcards).
+        By default, ``None`` (no images will be loaded).
+    file_ext : str or None, optional
         The file extension of the image files to be loaded, ignored if file types are specified in ``path_images`` (e.g. with ``"./path/to/dir/*png"``).
-        By default ``False``.
+        By default ``None``.
     tree_level : str, optional
         Level of the image hierarchy to construct. The value can be
         ``"parent"`` (default) and ``"patch"``.
-    parent_path : str, optional
+    parent_path : str or None, optional
         Path to parent images (if applicable), by default ``None``.
     **kwargs : dict, optional
         Keyword arguments to pass to the
@@ -72,8 +77,8 @@ class MapImages:
     def __init__(
         self,
         path_images: str | None = None,
-        file_ext: str | bool | None = False,
-        tree_level: str | None = "parent",
+        file_ext: str | None = None,
+        tree_level: str = "parent",
         parent_path: str | None = None,
         **kwargs: dict,
     ):
@@ -90,6 +95,7 @@ class MapImages:
         self.images = {"parent": {}, "patch": {}}
         self.parents = self.images["parent"]
         self.patches = self.images["patch"]
+        self.georeferenced = False
 
         for image_path in tqdm(self.path_images):
             self._images_constructor(
@@ -99,28 +105,61 @@ class MapImages:
                 **kwargs,
             )
 
+        self.check_georeferencing()
+
+    def check_georeferencing(self):
+        if all(
+            "coordinates" in self.parents[parent_id].keys()
+            for parent_id in self.list_parents()
+        ):
+            self.georeferenced = True
+            self.add_patch_coords()
+
     @staticmethod
-    def _resolve_file_path(file_path, file_ext=None):
-        if file_ext:
-            if os.path.isdir(file_path):
-                files = glob(os.path.abspath(f"{file_path}/*.{file_ext}"))
-            else:  # if not dir
-                files = glob(os.path.abspath(file_path))
-                files = [file for file in files if file.split(".")[-1] == file_ext]
+    def _resolve_file_path(file_path: str, file_ext: str | None = None):
+        """Resolves file path to list of files.
+
+        Parameters
+        ----------
+        file_path : str
+            Path to the file(s) or directory containing files. Can contain wildcards.
+        file_ext : str or None, optional
+            The file extension of the images to be loaded, by default ``None``. Ignored if file types are specified in ``file_path`` (e.g. with ``"./path/to/dir/*png"``).
+
+        Returns
+        -------
+        list
+            List of file paths as strings.
+
+        Notes
+        -----
+        Valid inputs for file path are:
+
+        - Path to a directory containing images (e.g. ``"./path/to/dir"``).
+        - Path to a specific image file (e.g. ``"./path/to/image.png"``).
[…]
