### 2024-08-12 daa12fc
add labels_map argument when loading annots
