### 2024-08-12 6ecf0af
Set version to 0.6 and document changes
