### 2024-08-13 #6434 @emdupre
Thank you for confirming !

You should be able to mint a new version of the software. If you have the Zenodo-GitHub integration enabled, this should then automatically create a new Zenodo record. If not, you can create a new version (with a matching version number to your GitHub release) directly on Zenodo: https://help.zenodo.org/docs/deposit/manage-versions/

If you can then share the resulting version number and doi here, I can move this forward ! 🚀
