### 2024-08-16 fcde40c
add "update changelog" into PR template
