+++ docs/source/community-and-contributions/contribution-guide/getting-started-with-contributions/add-or-update-code.rst
@@ -37,8 +37,8 @@ Style guide
 
 When making your changes, please:
 
-- Try to align to the `PEP 8 style guide for Python code <https://peps.python.org/pep-0008/>`.
-- Try to use the numpy-style docstrings (as per `this link <https://numpydoc.readthedocs.io/en/latest/format.html#>_`).
+- Try to align to the `PEP 8 style guide for Python code <https://peps.python.org/pep-0008/>`__.
+- Try to use the numpy-style docstrings (as per `this link <https://numpydoc.readthedocs.io/en/latest/format.html#>`__).
 - Ensure all docstrings are kept up to date and reflect any changes to code functionality you have made.
 - Add and run tests for your code.
 - If you add new dependencies, add these to our ``setup.py``.
@@ -46,8 +46,8 @@ When making your changes, please:
 
 When you are done making changes, please:
 
-- Run `black <https://black.readthedocs.io/en/stable/>` to reformat your code
-- Run `flake8 <https://flake8.pycqa.org/en/latest/index.html#>` to lint your code.
+- Run `black <https://black.readthedocs.io/en/stable/>`__ to reformat your code
+- Run `flake8 <https://flake8.pycqa.org/en/latest/index.html#>`__ to lint your code.
 
 
 Running tests
