### 2024-08-21 #6434 @emdupre
Hi @rwood-97, please let me know if you're able to create this new DOI, or if you're hitting any issues I can help in troubleshooting !
