### 2024-08-21 95bca37
add docs on updating the changelog
