### 2024-08-21 #6434 @rwood-97
Hi, sorry for being slow on this - I'll do the updated zenodo record today.
We are still trying to move MapReader to its own repo but are trying to get in touch with a github user called MapReader (who has been inactive for 2 years) to ask if we can steal his username.
