### 2024-08-21 27f5fb2
Adding in links to upcoming community calls + workshops
