### 2024-08-21 #6434 @emdupre
Thank you, @rwood-97 ! I've updated the archive and confirmed that it the paper rendering looks correct.

I'm now happy to recommend _MapReader_ to the JOSS EiC team for publication.

Thank you again to @geekysquirrel, @PipGrylls, and @jordibc for all of your work throughout this review process ! 💐
