### 2024-08-22 #6434 @rwood-97
Hi @emdupre

Would it be possible to hold off publishing for ~1 week as we've decided to do the repo transfer and would ideally like to do this before publication of the JOSS paper incase we need to update any links.
I do this asap will update you when its done.
