### 2024-08-22 2a7993a
Update events.rst
