+++ tests/test_load/test_images.py
@@ -4,7 +4,7 @@ import os
 from pathlib import Path
 from random import randint
 
-import geopandas as geopd
+import geopandas as gpd
 import pandas as pd
 import pytest
 from PIL import Image
@@ -666,7 +666,7 @@ def test_save_to_geojson(init_maps, tmp_path, capfd):
     maps, _, _ = init_maps
     maps.save_patches_to_geojson(geojson_fname=f"{tmp_path}/patches.geojson")
     assert os.path.exists(f"{tmp_path}/patches.geojson")
-    geo_df = geopd.read_file(f"{tmp_path}/patches.geojson")
+    geo_df = gpd.read_file(f"{tmp_path}/patches.geojson")
     assert "geometry" in geo_df.columns
     assert str(geo_df.crs.to_string()) == "EPSG:4326"
     assert isinstance(geo_df["geometry"][0], Polygon)
@@ -684,7 +684,7 @@ def test_save_to_geojson_missing_data(sample_dir, image_id, tmp_path):
     )
     maps.save_patches_to_geojson(geojson_fname=f"{tmp_path}/patches.geojson")
     assert os.path.exists(f"{tmp_path}/patches.geojson")
-    geo_df = geopd.read_file(f"{tmp_path}/patches.geojson")
+    geo_df = gpd.read_file(f"{tmp_path}/patches.geojson")
     assert "geometry" in geo_df.columns
     assert str(geo_df.crs.to_string()) == "EPSG:4326"
     assert isinstance(geo_df["geometry"][0], Polygon)
@@ -702,7 +702,7 @@ def test_save_to_geojson_polygon_strings(
     assert isinstance(maps.patches[patch_id]["polygon"], str)
     maps.save_patches_to_geojson(geojson_fname=f"{tmp_path}/patches.geojson")
     assert os.path.exists(f"{tmp_path}/patches.geojson")
-    geo_df = geopd.read_file(f"{tmp_path}/patches.geojson")
+    geo_df = gpd.read_file(f"{tmp_path}/patches.geojson")
     assert "geometry" in geo_df.columns
     assert str(geo_df.crs.to_string()) == "EPSG:4326"
     assert isinstance(geo_df["geometry"][0], Polygon)
