+++ tests/test_download/test_sheet_downloader.py
@@ -47,24 +47,17 @@ def test_init_errors(sample_dir):
         SheetDownloader(test_json, 10)
 
 
-def test_get_polygons(sheet_downloader):
-    sd = sheet_downloader
-    sd.get_polygons()
-    assert sd.polygons is True
-    assert (isinstance(sd.features[i]["polygon"], Polygon) for i in sd.features)
-
-
 def test_get_grid_bb(sheet_downloader):
     sd = sheet_downloader
     sd.get_grid_bb()
-    assert sd.grid_bbs is True
-    assert (isinstance(sd.features[i]["grid_bb"], GridBoundingBox) for i in sd.features)
-    assert str(sd.features[0]["grid_bb"].lower_corner) == "(14, 8147, 5300)"
-    assert str(sd.features[0]["grid_bb"].upper_corner) == "(14, 8150, 5302)"
+    assert "grid_bb" in sd.metadata.columns
+    assert (isinstance(i, GridBoundingBox) for i in sd.metadata["grid_bb"])
+    assert str(sd.metadata.iloc[0]["grid_bb"].lower_corner) == "(14, 8147, 5300)"
+    assert str(sd.metadata.iloc[0]["grid_bb"].upper_corner) == "(14, 8150, 5302)"
 
     sd.get_grid_bb(10)
-    assert str(sd.features[0]["grid_bb"].lower_corner) == "(10, 509, 331)"
-    assert str(sd.features[0]["grid_bb"].upper_corner) == "(10, 509, 331)"
+    assert str(sd.metadata.iloc[0]["grid_bb"].lower_corner) == "(10, 509, 331)"
+    assert str(sd.metadata.iloc[0]["grid_bb"].upper_corner) == "(10, 509, 331)"
 
 
 def test_get_grid_bb_errors(sample_dir):
@@ -80,41 +73,35 @@ def test_get_grid_bb_errors(sample_dir):
 def test_extract_wfs_id_nos(sheet_downloader):
     sd = sheet_downloader
     sd.extract_wfs_id_nos()
-    assert sd.wfs_id_nos is True
-    assert sd.features[0]["wfs_id_no"] == 16320
+    assert "wfs_id_no" in sd.metadata.columns
+    assert sd.metadata.iloc[0]["wfs_id_no"] == 16320
 
 
 def test_extract_published_dates(sheet_downloader):
     sd = sheet_downloader
     sd.extract_published_dates()
-    assert sd.published_dates is True
-    assert sd.features[0]["properties"]["published_date"] == 1900
+    assert "published_date" in sd.metadata.columns
+    assert sd.metadata.iloc[0]["published_date"] == 1900
     assert (
-        sd.features[3]["properties"]["published_date"] == 1896
+        sd.metadata.iloc[3]["published_date"] == 1896
     )  # metadata has "1894 to 1896" - method takes end date only
 
-    sd.extract_published_dates(date_col=["properties", "YEAR"])  # list of keys
-    assert sd.features[0]["properties"]["published_date"] == 2023
-    assert sd.features[3]["properties"]["published_date"] == 1894
+    sd.extract_published_dates(date_col="YEAR")  # str
+    assert sd.metadata.iloc[0]["published_date"] == 2023
+    assert sd.metadata.iloc[3]["published_date"] == 1894
 
     sd.extract_published_dates(date_col="test_date")  # keys as str
-    assert sd.features[0]["properties"]["published_date"] == 2021
-    assert sd.features[3]["properties"]["published_date"] == 2021
+    assert sd.metadata.iloc[0]["published_date"] == 2021
+    assert sd.metadata.iloc[3]["published_date"] == 2021
 
 
 def test_published_dates_value_errors(sheet_downloader):
     sd = sheet_downloader
-    with pytest.raises(ValueError, match="as string or list of strings"):
[…]
