### 2024-08-27 6312f71
update geopandas for text spotting
