+++ mapreader/spot_text/dptext_detr_runner.py
@@ -9,6 +9,7 @@ except ImportError:
         "[ERROR] Please install DPText-DETR from the following link: https://github.com/rwood-97/DPText-DETR"
     )
 
+import geopandas as gpd
 import numpy as np
 import pandas as pd
 import torch
@@ -33,8 +34,8 @@ from .runner_base import Runner
 class DPTextDETRRunner(Runner):
     def __init__(
         self,
-        patch_df: pd.DataFrame | str,
-        parent_df: pd.DataFrame | str = None,
+        patch_df: pd.DataFrame | gpd.GeoDataFrame | str | pathlib.Path,
+        parent_df: pd.DataFrame | gpd.GeoDataFrame | str | pathlib.Path = None,
         cfg_file: str
         | pathlib.Path = "./DPText-DETR/configs/DPText_DETR/ArT/R_50_poly.yaml",
         weights_file: str | pathlib.Path = "./art_final.pth",
@@ -45,10 +46,10 @@ class DPTextDETRRunner(Runner):
 
         Parameters
         ----------
-        patch_df : pd.DataFrame | str
-            The dataframe containing the patch information. If a string, it should be a path to a CSV file.
-        parent_df : pd.DataFrame | str, optional
-            The dataframe containing the parent information. If a string, it should be a path to a CSV file, by default None.
+        patch_df : pd.DataFrame | gpd.GeoDataFrame | str | pathlib.Path,
+            The dataframe containing the patch information. If a string/path, it should be a path to a CSV/TSV/etc or geojson file.
+        parent_df : pd.DataFrame | gpd.GeoDataFrame | str | pathlib.Path,
+            The dataframe containing the parent information. If a string/path, it should be a path to a CSV/TSV/etc or geojson file, by default None.
         cfg_file : str | pathlib.Path, optional
             The path to the config file (yaml), by default "./DPText-DETR/configs/DPText_DETR/ArT/R_50_poly.yaml"
         weights_file : str | pathlib.Path, optional
@@ -153,9 +154,9 @@ class DPTextDETRRunner(Runner):
             A pandas DataFrame containing the predictions.
         """
         if geo:
-            columns = ["polygon", "crs", "score"]
+            columns = ["geometry", "crs", "score"]
         else:
-            columns = ["polygon", "score"]
+            columns = ["geometry", "score"]
 
         if parent:
             columns.append("patch_id")
@@ -168,6 +169,19 @@ class DPTextDETRRunner(Runner):
             )
             for k in preds.keys()
         )
+
+        if geo:
+            # get the crs (should be the same for all)
+            if not preds_df["crs"].nunique() == 1:
+                raise ValueError("[ERROR] Multiple crs found in the predictions.")
+            crs = preds_df["crs"].unique()[0]
+
+            preds_df = gpd.GeoDataFrame(
+                preds_df,
+                geometry="geometry",
+                crs=crs,
+            )
+
         preds_df.index.name = "image_id"
         preds_df.reset_index(inplace=True)
         return preds_df
