+++ mapreader/spot_text/runner_base.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import os
 import pathlib
-from ast import literal_eval
+import re
 from itertools import combinations
 
 import geopandas as gpd
@@ -13,6 +13,8 @@ import pandas as pd
 from PIL import Image
 from shapely import Polygon
 from tqdm.auto import tqdm
+
+from mapreader.utils.load_frames import load_from_csv, load_from_geojson
 
 
 class Runner:
@@ -37,52 +39,52 @@ class Runner:
         delimiter : str, optional
             The delimiter used in the csv file, by default ",".
         """
-        if isinstance(patch_df, pd.DataFrame):
+        if isinstance(patch_df, (pd.DataFrame, gpd.GeoDataFrame)):
             self.patch_df = patch_df
 
         elif isinstance(patch_df, str):
-            if os.path.isfile(patch_df):
-                print(f'[INFO] Reading "{patch_df}".')
-                patch_df = pd.read_csv(patch_df, sep=delimiter, index_col=0)
-                # ensure tuple/list columns are read as such
-                patch_df = self._eval_df(patch_df)
-                self.patch_df = patch_df
+            print(f'[INFO] Reading "{patch_df}"')
+            if re.search(r"\..?sv$", patch_df):
+                self.patch_df = load_from_csv(
+                    patch_df,
+                    delimiter=delimiter,
+                )
+            elif re.search(r"\..?json$", patch_df):
+                self.patch_df = load_from_geojson(patch_df)
             else:
-                raise ValueError(f'[ERROR] "{patch_df}" cannot be found.')
+                raise ValueError(
+                    "[ERROR] ``patch_df`` must be a path to a CSV/TSV/etc or geojson file, a pandas DataFrame or a geopandas GeoDataFrame."
+                )
 
         else:
             raise ValueError(
-                "[ERROR] Please pass ``patch_df`` as a string (path to csv file) or pd.DataFrame."
+                "[ERROR] ``patch_df`` must be a path to a CSV/TSV/etc or geojson file, a pandas DataFrame or a geopandas GeoDataFrame."
             )
 
-        if isinstance(parent_df, pd.DataFrame):
+        if isinstance(parent_df, (pd.DataFrame, gpd.GeoDataFrame)):
             self.parent_df = parent_df
 
         elif isinstance(parent_df, str):
-            if os.path.isfile(parent_df):
-                print(f'[INFO] Reading "{parent_df}".')
-                parent_df = pd.read_csv(parent_df, sep=delimiter, index_col=0)
-                parent_df = self._eval_df(parent_df)
-                self.parent_df = parent_df
+            print(f'[INFO] Reading "{parent_df}"')
+            if re.search(r"\..?sv$", parent_df):
+                self.parent_df = load_from_csv(
+                    parent_df,
+                    delimiter=delimiter,
+                )
+            elif re.search(r"\..?json$", parent_df):
+                self.parent_df = load_from_geojson(parent_df)
             else:
-                raise ValueError(f'[ERROR] "{parent_df}" cannot be found.')
+                raise ValueError(
+                    "[ERROR] ``parent_df`` must be a path to a CSV/TSV/etc or geojson file, a pandas DataFrame or a geopandas GeoDataFrame."
+                )
 
         else:
-            print("[WARNING] No parent dataframe provided.")
-            self.parent_df = parent_df
-
-    @staticmethod
-    def _eval_df(df):
-        for col in df.columns:
-            try:
-                df[col] = df[col].apply(literal_eval)
-            except (ValueError, TypeError, SyntaxError):
-                pass
-        return df
+            raise ValueError(
[…]
