### 2024-08-27 7ebf2bf
compatibility with old polygon vs geometry saving name
