### 2024-09-05 #6434 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
