### 2024-09-10 #6434 @kthyng
Ok I changed the repo link at the top of this review issue to https://github.com/maps-as-data/MapReader.

@openjournals/dev this submission is just about to be published and the authors changed the organization of the repo. Is there anywhere else the link needs to be updated?
