### 2024-09-10 #6434 @kthyng
@kmcdono2 I'm seeing multiple versions associated with this JOSS paper:
* github v1.3.3 (https://github.com/maps-as-data/MapReader/releases/tag/v1.3.3)
* zenodo v1.3.10 (https://doi.org/10.5281/zenodo.13643609)
* JOSS v1.3.9 (top of this issue)

Which is it meant to be?
