### 2024-09-10 #6434 @kthyng
Also in the JOSS-associated Zenodo version archive we request that the title and author list match the JOSS paper (not sure if you already discussed this above because the page is really long). Can you update the Zenodo metadata so these match? (it's a request not a requirement)
