### 2024-09-11 #6434 @rwood-97
Hi @kthyng ,

We'd like to link to v1.3.10 which is the version after we moved to maps-as-data org.
I'll update the order of authors on [this zenodo record](https://doi.org/10.5281/zenodo.13643609) now :)
