### 2024-09-11 #6434 @kthyng
@rwood-97 Ok great about the author order in Zenodo — how about the title? Did you want to update that for this version to match your JOSS paper or leave it as is?
