### 2024-09-12 #6434 @rwood-97
Hi, just done this. Title for the zenodo record is now "MapReader: Open software for the visual analysis of
maps" too
