### 2024-09-23 #6434 @xuanxu
@kthyng I've updated the repository URL for this submission in the JOSS site, nothing else needs updating before acceptance.
