### 2024-09-26 #6434 @kthyng
Congratulations on your new publication @kmcdono2! Many thanks to editor @emdupre and to reviewers @geekysquirrel,  @PipGrylls, and @jordibc for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.

@kmcdono2 I see you are in our reviewer database (thank you!) though you don't have any keywords associated with yourself, so you will never come up in a search when we're trying to find reviewers. It would be great if you'd be willing to update that. Just a heads up!  https://reviewers.joss.theoj.org/
